% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/App0.R
\docType{class}
\name{App0}
\alias{App0}
\title{App0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
App0
}
\description{
\code{App0} Abstract class  for the App family of classes.
}
\details{
This abstract class from which all *App classes are derived.
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
