% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Archive0.R
\docType{class}
\name{Archive0}
\alias{Archive0}
\title{Archive0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Archive0
}
\arguments{
\item{object}{\describe{
 \item{object$name}{Character string with name of the object}
 \item{object$desc}{Character string with description of the object e.g. "development-environment-archive"}
 \item{object$path}{Character string containing the directory containing the object}
}}
}
\description{
\code{Archive} Abstract class for archiving objects in the NLPStudio
}
\details{
This abstract class defines the methods used by the concrete classes for
archiving objects in the NLPStudio
}
\section{Method}{

\describe{
 \item{\code{new()}}{Creates an object of class Archive0}
 \item{\code{archive(object)}}{Archives the object in the NLPStudio archive subdirectory }
}
}

\section{Parameters}{

}

\examples{
\dontrun{
a <- archive$new()
a$archive0("development")
}

}
\author{
John James, \email{j2sdatalab@gmail.com}
}
\keyword{datasets}
