% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base.R
\docType{class}
\name{Base}
\alias{Base}
\title{Base}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Base
}
\description{
\code{Base} Base class from which all classes are descended.
}
\details{
This base class defines members and methods common across all classes, such
as validation and logging.
}
\section{Base methods}{

 \itemize{
  \item{\code{getName()}}{Returns name assigned to an object.}
  \item{\code{getParams()}}{Returns the parameters passed to the constructor.}
  \item{\code{logIt(level = c("Info", "Warn", "Error"))}}{Writes to the system log.}
 }
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
