% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildCorpus.R
\docType{class}
\name{BuildCorpus}
\alias{BuildCorpus}
\title{BuildCorpus}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BuildCorpus
}
\arguments{
\item{x}{FileBuildCorpus Object}

\item{fileStudio}{FileStudio object}
}
\description{
\code{BuildCorpus} Encapsulates the command to source a FileSet object.
}
\seealso{
Other FileStudio Command Classes: \code{\link{CleanCorpus}},
  \code{\link{CopyFileSet}}, \code{\link{MoveFileSet}},
  \code{\link{ReadCorpus}}, \code{\link{RepairFileSet}},
  \code{\link{SampleCorpus}}, \code{\link{Source}},
  \code{\link{SplitCorpus}}, \code{\link{TokenizeCorpus}},
  \code{\link{WriteCorpus}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
