% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildFileSetCmd.R
\docType{class}
\name{BuildFileSetCmd}
\alias{BuildFileSetCmd}
\title{BuildFileSetCmd}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BuildFileSetCmd
}
\arguments{
\item{x}{Receiver object, e.g. the FileStudio object.}

\item{path}{Character string containing the relative file or directory path
for the file(s).}

\item{name}{Character string containing the name of the FileSet object.}

\item{target}{Not used in this command.}
}
\description{
\code{BuildFileSetCmd} Encapsulates the command to execute to build a FileSet object.
}
\seealso{
Other FileStudio Command Classes: \code{\link{RepairFileSetCmd}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
