% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_readability.R
\name{calculate_readability}
\alias{calculate_readability}
\title{Fast Calculation of Readability Measures}
\usage{
calculate_readability(x, ncores, ...)
}
\arguments{
\item{x}{A \strong{quanteda} \link{corpus} or a character vector containing the documents to process.}

\item{ncores}{The number of \link{multisession} workers to be allocated for
the calculation of readability.}

\item{...}{Additional arguments passed to \link{textstat_readability}.}
}
\value{
A \link{data.table} with as many columns as passed to
\code{measure} and \code{doc_id} as the document identifier.
}
\description{
Compute readability measures in parallel with the \strong{\link{future}} paradigm.
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}
}
