% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanCorpus.R
\docType{class}
\name{CleanCorpus}
\alias{CleanCorpus}
\title{CleanCorpus}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CleanCorpus
}
\arguments{
\item{x}{Corpus or CVSet Object}

\item{corpusStudio}{CorpusStudio object}

\item{name}{Character string containing the name to assign to the Corpus
or CVSet object.}

\item{textConfig}{TextConfig object containing the instructions for cleaning
the object.}
}
\value{
Token object
}
\description{
\code{CleanCorpus} Encapsulates the command to clean a Corpus or CVSet object.
}
\seealso{
Other FileStudio Command Classes: \code{\link{BuildCorpus}},
  \code{\link{CopyFileSet}}, \code{\link{MoveFileSet}},
  \code{\link{ReadCorpus}}, \code{\link{RepairFileSet}},
  \code{\link{SampleCorpus}}, \code{\link{Source}},
  \code{\link{SplitCorpus}}, \code{\link{TokenizeCorpus}},
  \code{\link{WriteCorpus}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
