% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clone.R
\docType{class}
\name{Clone}
\alias{Clone}
\title{Clone}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Clone
}
\arguments{
\item{x}{Object to be cloned. Currently supports Corpus and Document objects.}

\item{name}{Character string containing the name to be assigned to the
replicant.}
}
\value{
A Replicant of the input object
}
\description{
\code{Clone} Abstract factory class that creates a clone of an object.
}
\details{
Given a Corpus or Document object, and an optional name, this class produces
a replica of the object via the appropriate concrete factory class, then
returns the replicant to the calling environment. The clones are actually
quasi clones as the replicants have their own identifiers, names,
and descriptions.
}
\section{Clone methods}{

\describe{
 \item{\code{new(x, name = NULL)}}{Instantiates the factory.}
 \item{\code{execute()}}{Returns the replicant object.}
}
}

\seealso{
Other Clone Classes: \code{\link{CloneCorpus}},
  \code{\link{CloneDocument}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
