% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CloneCorpus.R
\docType{class}
\name{CloneCorpus}
\alias{CloneCorpus}
\title{CloneCorpus}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CloneCorpus
}
\arguments{
\item{x}{Object to be cloned. Currently supports Corpus and Document objects.}

\item{name}{Character string containing the name to be assigned to the
replicant.}
}
\value{
Corpus object
}
\description{
\code{CloneCorpus} Concrete factory for cloning Corpus objects.
}
\details{
Class creates quasi clones of Corpus objects. Attachments are not cloned
within this Class.  The Documents attached to a corpus must be cloned
then attached.
}
\section{CloneCorpus methods}{

\describe{
 \item{\code{new(x, name = NULL)}}{Instantiates the factory.}
 \item{\code{execute()}}{Returns the replicant Corpus object.}
}
}

\seealso{
Other Clone Classes: \code{\link{CloneDocument}},
  \code{\link{Clone}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
