% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CloneDocument.R
\docType{class}
\name{CloneDocument}
\alias{CloneDocument}
\title{CloneDocument}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CloneDocument
}
\arguments{
\item{x}{Object to be cloned. Currently supports Corpus and Document objects.}

\item{name}{Character string containing the name to be assigned to the
replicant.}
}
\value{
Document object
}
\description{
\code{CloneDocument} Concrete factory for cloning Document objects.
}
\details{
Class creates quasi clones of Document objects including certain metadata
and text.
}
\section{CloneDocument methods}{

\describe{
 \item{\code{new(x, name = NULL)}}{Instantiates the factory.}
 \item{\code{execute()}}{Returns the replicant Document object.}
}
}

\seealso{
Other Clone Classes: \code{\link{CloneCorpus}},
  \code{\link{Clone}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
