% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CloneFiles.R
\docType{class}
\name{CloneFiles}
\alias{CloneFiles}
\title{CloneFiles}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CloneFiles
}
\arguments{
\item{x}{Object to be cloned. Currently supports Collection and File objects.}

\item{name}{Character string containing the name to be assigned to the
replicant.}
}
\value{
A Replicant of the input object
}
\description{
\code{CloneFiles} Class responsible for cloning Collection and File objects.
}
\details{
Class responsible for cloning Collection and File objects.
}
\section{Methods}{

\describe{
 \item{\code{new(x, name = NULL)}}{Instantiates the factory.}
 \item{\code{execute()}}{Returns the replicant object.}
}
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
