% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneMeta.R
\name{cloneMeta}
\alias{cloneMeta}
\title{CloneMeta}
\usage{
cloneMeta(from, to)
}
\arguments{
\item{from}{Object from which the metadata will be cloned}

\item{to}{Object to which the metadata will be cloned}
}
\value{
to Object to which the metadata was cloned
}
\description{
\code{CloneMeta} Function for copying functional and descriptive meta data between objects.
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
