% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cmd0.R
\docType{class}
\name{Cmd0}
\alias{Cmd0}
\title{Cmd0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Cmd0
}
\arguments{
\item{x}{Receiver class object}

\item{target}{Target class object}

\item{...}{Additional parameters}
}
\description{
\code{Cmd0} Abstract class  for the App family of classes.
}
\details{
This abstract class from which all *App classes are derived.
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
