% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection0.R
\docType{class}
\name{Collection0}
\alias{Collection0}
\title{Collection0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Collection0
}
\description{
\code{Collection0} Abstract class for all Collection classes
}
\details{
This base class defines members and methods common across all the Collection
classes, such as Corpus, Data, Analysis, Model and Evaluation classes.
}
\section{Collection0 methods}{

 \itemize{
  \item{\code{attach(x)}}{Method for attaching objects.}
  \item{\code{detach(key, value)}}{Method for detaching objects.}
  \item{\code{attachments(key = NULL, value = NULL)}}{Lists an object's attachments.}
 }

 @param x Object or list of objects to be attached
 @param key Character string or vector of strings indicating the metadata
 variable or variables used for matching
 @param value Character string or vector of strings indicating the
 metadata value associated with the key(s) parameter.
}

\seealso{
Other Collection Classes: \code{\link{DataCollection0}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
