% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Composite.R
\docType{class}
\name{Composite}
\alias{Composite}
\title{Composite}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Composite
}
\description{
\code{Composite} Base class for all Composite classes
}
\details{
This base class defines members and methods common across all the Composite
classes NLPStudio, Pipeline, Corpus and Document.
}
\section{Composite methods}{

 \itemize{
  \item{\code{attach(x)}}{Method for attaching objects.}
  \item{\code{detach(key, value)}}{Method for detaching objects.}
  \item{\code{attachments(key = NULL, value = NULL)}}{Lists an object's attachments.}
 }

 @param x Object or list of objects to be attached
 @param key Character string or vector of strings indicating the metadata
 variable or variables used for matching
 @param value Character string or vector of strings indicating the
 metadata value associated with the key(s) parameter.
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
