% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Composite0.R
\docType{class}
\name{Composite0}
\alias{Composite0}
\title{Composite0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Composite0
}
\description{
\code{Composite0} Abstract class from which all composite classes are derived.
}
\details{
This abstract class defines members and methods common to the composite classes.
}
\section{Composite0 methods}{

 \itemize{
  \item{\code{new()}}{Not implemented.}
  \item{\code{getChildren(key, value)}}{Returns the child or children that
  match the key/value pair(s). }
  \item{\code{addChild(x)}}{Adds a child to the Set object.}
  \item{\code{removeChild(x)}}{Removes a child from the Set
  object.}
  \item{\code{summary(x)}}{Summarizes a Set child.}
 }

 @param x Object or list of objects to be attached
 @param key Character string or vector of strings indicating the metadata
 variable or variables used for matching
 @param value Character string or vector of strings indicating the
 metadata value associated with the key(s) parameter.
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\concept{Child Classes}
\keyword{datasets}
