% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Converter.R
\docType{class}
\name{Converter}
\alias{Converter}
\title{Converter}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Converter
}
\arguments{
\item{x}{Object to be converted}
}
\value{
Converted object.
}
\description{
\code{Converter} Converts Corpus objects to and from formats of other NLP packages
}
\section{Methods}{

 \describe{
  \item{\code{new(...)}}{Instantiates the Converter object.}
  \item{\code{quanteda()}}{Converts to and from quanteda format.}
  \item{\code{tm()}}{Converts to and from tm format.}
}
}

\seealso{
Other Converter family of classes: \code{\link{Converter0}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
