% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Converter0.R
\docType{class}
\name{Converter0}
\alias{Converter0}
\title{Converter0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Converter0
}
\arguments{
\item{x}{Object to be converted}
}
\description{
\code{Converter0} Abstract class for Converter classes.
}
\details{
Defines interface for concrete Converter classes.
}
\section{Methods}{

 \describe{
  \item{\code{new(...)}}{Instantiates the class object.}
  \item{\code{convert()}}{Converts between NLPStudio and the target package format,
  based upon the class of the object parameter.}
}
}

\seealso{
Other Converter family of classes: \code{\link{Converter}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
