% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConverterQuanteda.R
\docType{class}
\name{ConverterQuanteda}
\alias{ConverterQuanteda}
\title{ConverterQuanteda}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ConverterQuanteda$new()$convert(x)
}
\arguments{
\item{x}{Object to be converted}
}
\description{
\code{ConverterQuanteda} Converts NLPStudio Corpus objects to and from Quanteda corpus objects.
}
\seealso{
Other Converter Classes: \code{\link{ConverterTM}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
