% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConverterTM.R
\docType{class}
\name{ConverterTM}
\alias{ConverterTM}
\title{ConverterTM}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ConverterTM$new()$convert(x)
}
\arguments{
\item{x}{Object to be converted}
}
\description{
\code{ConverterTM} Converts NLPStudio Corpus objects to and from tm corpus objects.
}
\details{
Given an NLPStudio Corpus object, this object returns a tm SimpleCorpus object.
Alternatively, a SimpleCorpus, VCorpus, or PCorpus produces a NLPStudio
Corpus object. Note: Document level metadata is not maintained when
converting from NLPStudio Document objects to tm Corpus objects. Document
level metadata must be added separately to the tm Corpus object.
}
\seealso{
Other Converter Classes: \code{\link{ConverterQuanteda}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
