% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Copy.R
\docType{class}
\name{Copy}
\alias{Copy}
\title{Copy}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Copy
}
\arguments{
\item{x}{Object to be Copyd.}

\item{name}{Character string containing the name to be assigned to the
replicant.}
}
\value{
A Replicant of the input object
}
\description{
\code{Copy0} Abstract class for the Copy family of classes.
}
\details{
Abstract class that defines the interface and common methods for the
Copy family of classes.
}
\section{Methods}{

\describe{
 \item{\code{new(x, name = NULL)}}{Instantiates the factory.}
 \item{\code{execute()}}{Returns the replicant object.}
}
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\concept{Copy Classes}
\keyword{datasets}
