% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyFileSet.R
\docType{class}
\name{CopyFileSet}
\alias{CopyFileSet}
\title{CopyFileSet}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CopyFileSet
}
\arguments{
\item{to}{Director into which the FileSet should be moved.}

\item{fileStudio}{FileStudio object}
}
\description{
\code{CopyFileSet} Encapsulates the command to copy a FileSet object.
}
\seealso{
Other FileStudio Command Classes: \code{\link{BuildCorpus}},
  \code{\link{CleanCorpus}}, \code{\link{MoveFileSet}},
  \code{\link{ReadCorpus}}, \code{\link{RepairFileSet}},
  \code{\link{SampleCorpus}}, \code{\link{Source}},
  \code{\link{SplitCorpus}}, \code{\link{TokenizeCorpus}},
  \code{\link{WriteCorpus}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
