% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorpusBuilder0.R
\docType{class}
\name{CorpusBuilder0}
\alias{CorpusBuilder0}
\title{CorpusBuilder0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CorpusBuilder0
}
\arguments{
\item{name}{String containing the name of the corpus to be built.}

\item{desc}{String containing the dscription of the corpus to be built.}

\item{path}{String indicating the directory path in which the corpus resides}

\item{created}{Datetime object indicating create date/time}

\item{modified}{Datetime object indicating modification date/time}
}
\description{
\code{CorpusBuilder} Abstract builder class of the CorpusBuilder set of classes
}
\details{
This abstract builder class defines the interface for the CorpusBuilder
subclasses.  The class interacts with director and the concrete builder
classes to build the document corpora required for the analysis, processing,
and modeling phases. The standard interface methods obtain and process the
corpus, create the composite document and return the corpus meta data and
content to the calling environment.
}
\section{Methods}{

The following methods are defined in this class.
\describe{
 \item{\code{new(name, desc)}}{Abstract method for instantiating object of CorpusBuilder0 class.}
 \item{\code{buildComposite()}}{Abstract method for building the composite document.}
 \item{\code{buildComponent()}}{Abstract method for building the component document.}
 \item{\code{obtainCorpus()}}{Abstract method for obtaining the corpus. }
 \item{\code{processCorpus()}}{Abstract method for processing the corpus. }
 \item{\code{getCorpus()}}{Abstract method for returning the constructed corpus to the director class. }
}
}

\seealso{
Other CorpusBuilder Classes: \code{\link{CorpusBuilderRawWeb}},
  \code{\link{CorpusBuilderWeb}},
  \code{\link{CorpusDirector}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
