% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorpusBuilderRaw.R
\docType{class}
\name{CorpusBuilderRawWeb}
\alias{CorpusBuilderRawWeb}
\title{CorpusBuilderRawWeb}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CorpusBuilderRawWeb
}
\description{
\code{CorpusBuilder} Builds corpus sourced from the web.
}
\details{
This concrete builder class obtains the corpus, creates the document
composite and returns the corpus content and meta data to the calling
environment.
}
\section{Methods}{

The following methods are included in this class.
\describe{
 \item{\code{obtainCorpus()}}{Downloads the corpus from its internet web source.}
 \item{\code{processCorpus()}}{Unzips the corpus.}
 \item{\code{createComposite()}}{Creates the document composite and leaf objects for the corpus.}
 \item{\code{getCorpus()}}{Returns an object of the Document class, containing the corpus content and meta data.}
}
}

\seealso{
Other CorpusBuilder Classes: \code{\link{CorpusBuilder0}},
  \code{\link{CorpusBuilderWeb}},
  \code{\link{CorpusDirector}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
