% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorpusBuilderWeb.R
\docType{class}
\name{CorpusBuilderWeb}
\alias{CorpusBuilderWeb}
\title{CorpusBuilderWeb}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CorpusBuilderWeb
}
\arguments{
\item{fileNames}{String indicating the names of the files in the compressed document to obtain}

\item{url}{String containing the URL for the web source}
}
\description{
\code{CorpusBuilder} Builds a corpus obtained from the web.
}
\details{
This concrete builder class obtains the corpus, creates the component
document and returns the object to the calling environment. This will
download a single compressed file from the web source.
}
\section{Methods}{

The following methods are defined in this class.
\describe{
 \item{\code{new(name, desc, url)}}{Method for instantiating the builder}
 \item{\code{buildComposite()}}{Method for building the composite document. }
 \item{\code{buildComponent()}}{Method for building the component document.}
 \item{\code{obtainCorpus()}}{Method for obtaining the corpus. The corpus is downloaded from its web source }
 \item{\code{processCorpus()}}{Method for processing the corpus. This returns true as there is no processing.}
 \item{\code{getCorpus()}}{Method for returning the constructed composite document to the director class. }
 }
}

\seealso{
Other CorpusBuilder Classes: \code{\link{CorpusBuilder0}},
  \code{\link{CorpusBuilderRawWeb}},
  \code{\link{CorpusDirector}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
