% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorpusStudio.R
\docType{class}
\name{CorpusStudio}
\alias{CorpusStudio}
\title{CorpusStudio}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CorpusStudio
}
\arguments{
\item{x}{a FileSet object containing .txt files or a character string
containing the directory containing .txt files.}

\item{name}{Character string containing the name to assign to the Corpus
object.}
}
\description{
\code{CorpusStudio}  Class responsible for building and manipulating Corpus objects.
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
