% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Counter.R
\docType{class}
\name{Counter}
\alias{Counter}
\title{Counter}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Counter$new(x, tolower = TRUE, stem = FALSE, dictionary = NULL)$execute()
}
\arguments{
\item{x}{A Corpus object}

\item{tolower}{if TRUE, converts all letters to lower case. Default is TRUE.}

\item{stem}{if TRUE, stem words. Default is FALSE.}

\item{dictionary}{a quanteda dictionary class object or a named list
of character vector dictionary entries to apply to the tokens when creating the dfm.
See \code{\link[quanteda]{dfm}} for details.}

\item{...}{Other parameters passed to  \code{\link[quanteda]{dfm}}}
}
\value{
\code{\link{Counter}} object.
}
\description{
\code{Counter}  Strategy for building term frequency matrix objects from quanteda dfm objects.
}
\details{
A wrapper for \code{\link[quanteda]{dfm}}, this classes creates a sparse
document frequency matrix for a Corpus object.
Source \url{https://cran.r-project.org/web/packages/quanteda/quanteda.pdf}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
