% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Court.R
\docType{class}
\name{Court}
\alias{Court}
\title{Court}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Court
}
\arguments{
\item{object}{The object in which the parent is to be changed.}

\item{parent}{The new parent object.}
}
\description{
\code{Court} This class controls the process of re-assigning an object to
a new ancestor, via the VCourt visitor object. This object receives a request
to change the parent of an object.  The request is dispatched to the
VCourt visitor, via an accept method called on the originator. The VCourt
visitor validates the request, and if valid, removes the object from the
current parent and returns control back to the originator which updates
itself with the new parent.
}
\details{
\strong{Court Class Family Participants:}
\itemize{
 \item Court: This class validates the request, dispatches the appropriate
 visitor via the accept method to the calling object, then returns control
 to the calling object.
 \item VCourt: This visitor class validates the request, removes the object
 from the current parent and returns control back to the originating
 object.
 }

\strong{Court Class Collaborators:}
\itemize{
 \item VCourt: Visitor class that validates the request and removes the object from the current parent.
 \item Lab: The Lab class, a container for document collections. Invokes Court class object.
 \item DocumentCollection: Composite class containing DocumentCollection and Document class objects. Invokes Court class object.
 \item DocumentText: Class of text documents. Invokes Court class object.
 \item DocumentRdata: Class of Rdata documents. Invokes Court class object.
 \item DocumentCsv: Class of Csv documents. Invokes Court class object.
 \item DocumentXlsx: Class of excel documents. Invokes Court class object.#'
 }

\strong{Court Class Methods:}
The following methods are defined for this class:
\describe{
\item{\code{adjourn(object, parent)}}{Dispatches the accept method in the calling object, which dispatches the VCourt visitor}
}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
