% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSource0.R
\docType{class}
\name{CSource0}
\alias{CSource0}
\title{CSource0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CSource0
}
\arguments{
\item{name}{Optional character vector indicating name for Corpus object.}

\item{x}{Character vector containing text, the name of a file, or directory
or a corpus object from a supported package.}
}
\description{
\code{CSource0} Abstract class that defines the interface for the concrete CSource classes
}
\section{Methods}{

 \itemize{
  \item{\code{new()}}{Initializes an object of the CSource0 class.}
  \item{\code{execute()}}{Executes the process of sourcing the Corpus object.}
 }
}

\seealso{
Other Corpus Source Classes: \code{\link{CSourceVector}},
  \code{\link{CSource}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
