% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSourceDir.R
\docType{class}
\name{CSourceDir}
\alias{CSourceDir}
\title{CSourceDir}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CSourceDir
}
\arguments{
\item{name}{Optional character vector indicating name for Corpus object.}

\item{x}{Character vector or a list of character vectors containing text.}
}
\description{
\code{CSourceDir} Sources a Corpus object from a directory source.
}
\details{
Sources a Corpus object from a directory source. Each file yields a single Document
object and a single associated Text object.
}
\section{Methods}{

 \itemize{
  \item{\code{new(X, name = NULL)}}{Initializes an object of the CSourceDir class.}
  \item{\code{execute()}}{Executes the process of sourcing the Corpus object.}
 }
}

\examples{
dir <- "./foo"
corpus <- CSource$new(x = dir, name = "Foo")$dir()

}
\seealso{
Other Corpus Source Classes: \code{\link{CSource0}},
  \code{\link{CSourceQuanteda}}, \code{\link{CSourceTM}},
  \code{\link{CSourceVector}}, \code{\link{CSource}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
