% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSourceFileCollection.R
\docType{class}
\name{CSourceFileCollection}
\alias{CSourceFileCollection}
\title{CSourceFileCollection}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CSourceFileCollection
}
\arguments{
\item{name}{Optional character vector indicating name for Corpus object.}

\item{x}{Character vector or a list of character vectors containing text.}
}
\description{
\code{CSourceFileCollection} Sources a Corpus object from a FileCollection object.
}
\section{Methods}{

 \itemize{
  \item{\code{new(x, name = NULL)}}{Initializes an object of the CSourceFileCollection class.}
  \item{\code{source()}}{Executes the process of sourcing the Corpus object.}
 }
}

\examples{
text <- c("The firm possesses unparalleled leverage in Washington,
thanks in part to its track record of funneling executives into
senior government posts. Even the Trump administration, which
rode a populist wave to electoral victory, is stocked with
Goldman alumni, including Treasury Secretary Steven Mnuchin and
the departing White House economic adviser Gary D. Cohn.",
"Goldman is also an adviser to many of America’s — and the
world’s — largest companies, ranging from stalwarts like Walt
Disney to upstarts like Uber.")

corpus <- CSource$new(x = txt, name = "Goldman")$vector()

}
\seealso{
Other Corpus Source Classes: \code{\link{CSource0}},
  \code{\link{CSourceDir}}, \code{\link{CSourceQuanteda}},
  \code{\link{CSourceVector}}, \code{\link{CSource}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
