% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSourceQuanteda.R
\docType{class}
\name{CSourceQuanteda}
\alias{CSourceQuanteda}
\title{CSourceQuanteda}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CSourceQuanteda
}
\arguments{
\item{name}{Optional character vector indicating name for Corpus object.}

\item{x}{Character vector or a list of character vectors containing text.}
}
\description{
\code{CSourceQuanteda} Sources a Corpus object from a Quanteda corpus object
}
\section{Methods}{

 \itemize{
  \item{\code{new()}}{Initializes an object of the CSourceQuanteda class.}
  \item{\code{execute()}}{Executes the process of sourcing the Corpus object.}
 }
}

\seealso{
Other Corpus Source Classes: \code{\link{CSource0}},
  \code{\link{CSourceDir}}, \code{\link{CSourceVector}},
  \code{\link{CSource}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
