% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSourceTM.R
\docType{class}
\name{CSourceTM}
\alias{CSourceTM}
\title{CSourceTM}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CSourceTM
}
\arguments{
\item{name}{Optional character vector indicating name for Corpus object.}

\item{x}{Character vector or a list of character vectors containing text.}
}
\description{
\code{CSourceTM} Sources a Corpus object from a TM corpus object
}
\section{Methods}{

 \itemize{
  \item{\code{new(x, name = NULL)}}{Initializes an object of the CSourceTM class.}
  \item{\code{execute()}}{Executes the process of sourcing the Corpus object.}
 }

 tmc <- tm::VCorpus(DirSource("./foo))
 corpus <- CSource$new(x = tmc, name = "Alexandra")$tm()
}

\seealso{
Other Corpus Source Classes: \code{\link{CSource0}},
  \code{\link{CSourceDir}}, \code{\link{CSourceQuanteda}},
  \code{\link{CSourceVector}}, \code{\link{CSource}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
