% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSourceV.R
\docType{class}
\name{CSourceV}
\alias{CSourceV}
\title{CSourceV}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CSourceV
}
\arguments{
\item{name}{Optional character vector indicating name for Corpus object.}

\item{x}{Character vector or a list of character vectors containing text.}

\item{collapse}{Logical. If true, collapse the character vectors into
a single document.}
}
\description{
\code{CSourceV} Sources a Corpus object from a character vector or vectors.
}
\details{
Sources a Corpus object from a character vector or vectors. Each vector element
is treated as a separate document unless the collapse paramter is set to TRUE.
If a list is encountered, each list element is treated as a separate document
and the vectors are collapsed into a single document within the list
structure.
}
\section{Methods}{

 \itemize{
  \item{\code{new(x, name = NULL)}}{Initializes an object of the CSourceV class.}
  \item{\code{execute()}}{Executes the process of sourcing the Corpus object.}
 }
}

\examples{
text <- c("The firm possesses unparalleled leverage in Washington,
thanks in part to its track record of funneling executives into
senior government posts. Even the Trump administration, which
rode a populist wave to electoral victory, is stocked with
Goldman alumni, including Treasury Secretary Steven Mnuchin and
the departing White House economic adviser Gary D. Cohn.",
"Goldman is also an adviser to many of America’s — and the
world’s — largest companies, ranging from stalwarts like Walt
Disney to upstarts like Uber.")

corpus <- CSource$new(x = txt, name = "Goldman")$vector()

}
\seealso{
Other Corpus Source Classes: \code{\link{CSource0}},
  \code{\link{CSourceDir}}, \code{\link{CSourceFileSet}},
  \code{\link{CSourceQ}}, \code{\link{CSourceTM}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
