% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSourceVector.R
\docType{class}
\name{CSourceVector}
\alias{CSourceVector}
\title{CSourceVector}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CSourceVector
}
\arguments{
\item{name}{Optional character vector indicating name for Corpus object.}

\item{x}{Character vector or a list of character vectors containing text.}
}
\description{
\code{CSourceVector} Sources a Corpus object from a character vector or a list of character vectors.
}
\details{
Sources a Corpus object from a character vector or a list thereof. Each vector will create
a single Text object and one Document object.
}
\section{Methods}{

 \itemize{
  \item{\code{new()}}{Initializes an object of the CSourceVector class.}
  \item{\code{execute()}}{Executes the process of sourcing the Corpus object.}
 }
}

\seealso{
Other Corpus Source Classes: \code{\link{CSource0}},
  \code{\link{CSource}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
