% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Curator.R
\docType{class}
\name{Curator}
\alias{Curator}
\title{Curator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Curator
}
\arguments{
\item{current}{Object of the Lab, DocumentCollection, or Document classes in its current state.}

\item{restored}{Object of the Lab, DocumentCollection, or Document classes in a prior state.}
}
\description{
\code{Curator} This class is responsible for (re)building objects and
restoring  object composite hierarchies to prior states. The statebuilder
object takes as its parameters, the current and prior state objects. A
comparison method determines the elements common to both objects or
object composites, as well as the objects only found in the current (which
will be removed from the composition), and those found only in the
prior state (to be reinstated and added to composition). The (re)build operation
happens in the following four stages.
}
\seealso{
Other State Classes: \code{\link{StateManager}},
  \code{\link{StateServer}}, \code{\link{State}},
  \code{\link{VCurator}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
