% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVFactory0.R
\docType{class}
\name{CVFactory0}
\alias{CVFactory0}
\title{CVFactory0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CVFactory0
}
\arguments{
\item{x}{Corpus object for which the CVSetHoldOut object is to be created.}

\item{unit}{Character string indicating whether the unit for random sampling is
the vector or sentence. If the latter, the documents will be tokenized into sentences
prior to sampling.}

\item{stratify}{Logical. If TRUE, sampling will be conducted at the individual
document level. Otherwise, the Document objects will be combined into a single
Document object prior to sampling.}

\item{proportions}{Numeric vector of length 2 or 3, indicating the proportions to allocate
to the training (validation) and test sets.  If the length of the vector is 2, a
training and test set will be created in accordance with the proportions indicated. If
the length is 3, a validation set will be created in addition to the training and
test sets.}

\item{k}{The number of K-Fold validation sets to create.}

\item{seed}{Integer used to set the seed for random sampling}
}
\description{
\code{CVFactory0} Abstract factory class for the family of  cross validation factory classes.
}
\section{Methods}{

 \itemize{
  \item{\code{new()}}{Not implemented for this abstract class.}
  \item{\code{build()}}{Executes the process of sourcing the Corpus object.}
 }
}

\seealso{
Other Cross Validation Classes: \code{\link{CVFactoryHoldOut}},
  \code{\link{CVFactoryKFold}}, \code{\link{CVSet}},
  \code{\link{CV}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
