% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVKFold.R
\docType{class}
\name{CVKFold}
\alias{CVKFold}
\title{CVKFold}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CVKFold
}
\arguments{
\item{x}{A CVSet object.}

\item{name}{Character string containing the name for the CVKFold object.}

\item{k}{Numeric index into the enclosed CVSets, used for retrieving
a CVSet by index.}
}
\description{
\code{CVKFold} Class representing a K-FOLD collection of cross-validation sets
}
\details{
This composite class contains a collection of K CVSet objects.
}
\section{Methods}{

 \itemize{
  \item{\code{new(name = NULL)}}{Initializes an object of the CVKFold class.}
  \item{\code{getCVSet(what = NULL)}}{Obtains one or several CVSet objects
  from the CVKFold.}
  \item{\code{addCVSet(corpus)}}{Adds a CVSet object to the CVKFold object.}
  \item{\code{removeCVSet(corpus)}}{Removes a CVSet object from the CVKFold object.}
 }
}

\seealso{
Other Cross Validation Classes: \code{\link{CVSet}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
