% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVSet.R
\docType{class}
\name{CVSet}
\alias{CVSet}
\title{CVSet}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CVSet
}
\arguments{
\item{x}{A Corpus object.}

\item{name}{Character string containing the name for the CVSet object.}

\item{what}{Character indicating which Corpus object to obtain. Valid values are
c("training', 'validation', 'test', NULL).}
}
\description{
\code{CVSet} Class representing a cross-validation set
}
\details{
Class representing cross-validation sets. A cross-validation contains
a training, test, and optional validation set.
}
\section{Methods}{

 \itemize{
  \item{\code{new(name = NULL)}}{Initializes an object of the CVSet class.}
  \item{\code{getCorpus(what = NULL)}}{Obtains one or several Corpus objects
  from the CVSet. What may contain c("training', 'validation', 'test'). If
  NULL all Corpus objects are returned.}
  \item{\code{addCorpus(corpus)}}{Adds a Corpus object to the CVSet object.}
  \item{\code{removeCorpus(corpus)}}{Removes a Corpus object from the CVSet object.}
 }
}

\seealso{
Other Cross Validation Classes: \code{\link{CVKFold}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
