% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVSetKFold.R
\docType{class}
\name{CVSetKFold}
\alias{CVSetKFold}
\title{CVSetKFold}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CVSetKFold
}
\arguments{
\item{name}{Character string containing the name for the CVSetKFold object.}

\item{purpose}{Character string used to indicate how the document will be used, e.g. 'traian', 'test'.}

\item{note}{Character string containing a comment associated with a call to the
text method. The texts of the note variable are written to the CVSetKFolds
log. This is used to track changes to the text, perhaps made during preprocessing.}
}
\value{
CVSetKFold object, containing the CVSetKFold text, the metadata and
the methods to manage both.
}
\description{
\code{CVSetKFold} Class containing k CVSets created by the KFold Class.
}
\section{Core Methods}{

 \itemize{
  \item{\code{new(name = NULL)}}{Initializes an object of the CVSetKFold class.}
  \item{\code{text(x, note = NULL)}}{Method for obtaining/adding/updating text. If no
  parameters are presented, the current text is returned.  Otherwise, the text
  is updated with the texts of the character vector 'x'. Sentence, word, token, type,
  sentence and word length statistics are also computed and the metadata is updated
  accordingly.}
  \item{\code{summary()}}{Summarizes the CVSetKFold object.}
 }
}

\seealso{
Other CVSet Classes: \code{\link{CVSet}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\concept{CVSet Classes}
\keyword{datasets}
