% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{data_dictionary_BozanicRoulstoneVanBuskirk_FLS}
\alias{data_dictionary_BozanicRoulstoneVanBuskirk_FLS}
\title{Bozanic Roulstone VanBuskirk Forward Looking Statement Dictionary}
\format{
An object of class \code{dictionary2} of length 1.
}
\source{
\url{https://www.sciencedirect.com/science/article/abs/pii/S0165410117300733}
}
\usage{
data_dictionary_BozanicRoulstoneVanBuskirk_FLS
}
\description{
A \code{\link[quanteda]{dictionary}} containing the
the 2018 version of the Bozanic, Roulstone, and VanBuskirk Forward Looking Statement Word List.
The list contains 90 words.
}
\references{
Bozanic, Z., Roulstone, D.T., Van Buskirk, A. (2018), Management earnings forecasts and other
forward-looking statements, \emph{Journal of Accounting and Economics}, Volume 65, Issue 1, 2018,
Pages 1-20.
}
\keyword{data}
