% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{data_dictionary_Cannon_Ling_Wang_Watanabe}
\alias{data_dictionary_Cannon_Ling_Wang_Watanabe}
\title{Cannon, Ling, Wang, and Watanabe CSR Dictionary}
\format{
An object of class \code{dictionary2} of length 4.
}
\source{
\url{https://www.tandfonline.com/doi/epdf/10.1080/09638180.2019.1670223?src=getftr}
}
\usage{
data_dictionary_Cannon_Ling_Wang_Watanabe
}
\description{
A \link[quanteda:dictionary]{quanteda::dictionary} containing the Cannon, Ling, Wang, and Watanabe corporate
social responsibility word list. The object has four top-level categories
and mixes unigrams with multi-word expressions.
}
\details{
The dictionary contains four categories:
\itemize{
\item \code{Philantropy}
\item \code{BusinessPractice}
\item \code{Product}
\item \code{General}
}
}
\references{
Cannon, J. N., Ling, Z., Wang, Q., & Watanabe, O. V. (2020).
10-K disclosure of corporate social responsibility and firms’ competitive advantages.
\emph{European Accounting Review}, 29(1), 85-113.
}
\keyword{data}
