% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{data_dictionary_Li_FLS}
\alias{data_dictionary_Li_FLS}
\title{Feng Li Forward Looking Statement Dictionary}
\format{
An object of class \code{dictionary2} of length 1.
}
\source{
\url{https://onlinelibrary.wiley.com/doi/abs/10.1111/j.1475-679X.2010.00382.x}
}
\usage{
data_dictionary_Li_FLS
}
\description{
A \code{\link[quanteda]{dictionary}} containing the
the 2010 version of the Feng Li's Forward Looking Statement Word List. The list contains 18 words.
}
\references{
Li, F. (2010). The Information Content of Forward-Looking Statements in
Corporate Filings—A Naïve Bayesian Machine Learning Approach,
\emph{Journal of Accounting Research}, 48 (5), 1049--1102.
}
\keyword{data}
