% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{data_dictionary_LoughranMcDonald_Complexity}
\alias{data_dictionary_LoughranMcDonald_Complexity}
\title{Loughran and McDonald Firm Complexity Dictionary}
\format{
An object of class \code{dictionary2} of length 1.
}
\source{
\url{https://sraf.nd.edu/complexity/}

\url{https://papers.ssrn.com/sol3/papers.cfm?abstract_id=3645372}
}
\usage{
data_dictionary_LoughranMcDonald_Complexity
}
\description{
A \code{\link[quanteda]{dictionary}} containing the
the 2024 version of the Loughran and McDonald Firm Complexity Word List. The list contains 53 words.
}
\references{
Loughran, T. & McDonald, B. (2024). Measuring Firm Complexity,
\emph{Journal of Financial and Quantitative Analysis},
\href{https://papers.ssrn.com/sol3/papers.cfm?abstract_id=3645372}{Forthcoming}.
}
\keyword{data}
