% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{data_dictionary_SDG}
\alias{data_dictionary_SDG}
\title{U.N. Sustainable Development Goals (SDG) Mapping Dictionary}
\format{
An object of class \code{dictionary2} of length 16.
}
\source{
\url{https://www.researchsquare.com/article/rs-2544385/v2}

\url{https://www.sdgmapping.auckland.ac.nz/}
}
\usage{
data_dictionary_SDG
}
\description{
A \link[quanteda:dictionary]{dictionary} containing key terms for mapping research text to
Sustainable Development Goal themes following Wang et al. (2023). The object
has 16 top-level entries named \code{SDG1} through \code{SDG16}.
}
\details{
The Auckland Approach, an enhanced method for mapping research publications to the Sustainable
Development Goals (SDGs), utilizes advanced text-mining techniques and n-gram analyses to
refine and expand keyword lists derived from publication metadata. This approach, built on
foundational work by Elsevier, Sustainable Development Solutions Network, and the
United Nations (UN), improves the precision and coverage of SDG
mapping by incorporating both globally standardized and locally relevant keywords.
The methodology ensures a more comprehensive, contextualized, and data-driven mapping process,
effectively tailoring global SDG themes to specific local research needs and narratives.
}
\references{
Wang, W., Kang, W., & Mu, J. (2023). Mapping research to the sustainable development goals
\emph{Working Paper} available at \url{https://www.researchsquare.com/article/rs-2544385/v2}.
}
\keyword{data}
