% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataStudio.R
\docType{class}
\name{DataStudio}
\alias{DataStudio}
\title{DataStudio}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataStudio
}
\arguments{
\item{object}{The object to be processed.}

\item{queue}{The job queue containing text processing commands.}
}
\description{
\code{DataStudio} Class converts text to data.
}
\details{
Class automates the process of converting to text to data via stemming,
POS tagging, and the creation of document frequency matrices, ngrams, tokens.
}
\section{DataStudio Family of Classes Overview}{

The DataStudio family of classes include the DataStudio class, which is
responsible for executing one or more data processing commands on a
Corpus or Document resource, the core class group that performs the
data processing functions on the objects, and the command class
group which instructs the DataStudio class to perform the
data processing function on the Corpus or Document object.
}

\section{DataStudio methods}{

\strong{Core Methods:}
 \itemize{
  \item{\code{new()}}{Method for instantiating a DataStudio.}
  \item{\code{addCommand()}}{Method which adds a data processing command to the queue. }
  \item{\code{removeCommand()}}{Method that removes a data processing command from the queue.}
  \item{\code{execute()}}{Method that executes the job queue. }
  \item{\code{getResult()}}{Method that returns the object following execution of the job queue. }
 }
}

\section{Parameters}{

}

\seealso{
Other DataStudio classes: \code{\link{DataStudio0}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
