% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_corpus.R
\name{define_corpus}
\alias{define_corpus}
\alias{define_corpus.character}
\alias{define_corpus.data.table}
\title{Generate a Quanteda Corpus from Text Data}
\usage{
define_corpus(x, ...)

\method{define_corpus}{character}(x, ncores = 1, ...)

\method{define_corpus}{data.table}(x, ...)
}
\arguments{
\item{x}{Input object, either a character indicating a path to stored data or a data.table as created by \link{from_json_to_df}.}

\item{...}{Additional arguments passed to \link[readtext:readtext]{readtext::readtext}.}

\item{ncores}{The number of \link[future:multisession]{future::multisession} workers to be allocated for
the lookup. Default to 1.}
}
\value{
A quanteda \link[quanteda:corpus]{quanteda::corpus} object.
}
\description{
Convert structured text data into a quanteda \link[quanteda:corpus]{quanteda::corpus}, supporting a variety of formats with parallel processing via \strong{\link[future:future]{future::future}}.
}
\details{
If \code{x} points directly to a specific file containing the textual data, this is treated as a single-item
list. If \code{x} is a folder, \code{define_corpus} looks for the following formats: .txt, .csv, .json, or .xml. This is in compliance with \link[readtext:readtext]{readtext::readtext}.

When passing either a single item or multiple small items, setting \code{ncores > 1} does not necessarily show a tangible improvement in speed. This is why by default \code{ncores = 1}.
}
