% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFM2.R
\docType{class}
\name{DFM}
\alias{DFM}
\title{DFM}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DFM$new(x tolower = TRUE, stem = FALSE, select = NULL, remove = NULL,
#' dictionary = NULL, thesaurus = NULL, valuetype = c("glob", "regex","fixed"),
groups = NULL, verbose = quanteda_options("verbose"), ...)$execute()
}
\arguments{
\item{x}{Object of to be processed. This can be an object of the Corpus, Document, character classes, or a list of character vectors.}

\item{tolower}{if TRUE, converts all letters to lower case.}

\item{stem}{if TRUE, stem words}

\item{select}{a pattern of user-supplied features to keep, while excluding all others.
See \code{\link[quanteda]{dfm}} for details.}

\item{remove}{a pattern of user-supplied features to ignore, such as "stop words".
See \code{\link[quanteda]{dfm}} for details.}

\item{dictionary}{a dictionary object to apply to the tokens when creating the dfm.
See \code{\link[quanteda]{dfm}} for details.}

\item{thesaurus}{a dictionary object that will be applied as if exclusive = FALSE.
See \code{\link[quanteda]{dfm}} for details.}

\item{valuetype}{A character string containing the type of pattern matching:
"glob" for "glob"-style wildcard expressions; "regex" for regular expressions; or
"fixed" for exact matching. Defaults to 'fixed'. See \code{\link[quanteda]{dfm}} for details.}

\item{groups}{Either: a character vector containing the names of document variables to
be used for grouping; or a factor or object that can be coerced into a factor
equal in length or rows to the number of documents. See \code{\link[quanteda]{dfm}}
for details.}

\item{verbose}{Display messages if TRUE.}
}
\value{
a \code{\link{DFM}} object.
}
\description{
\code{DFM}  Creates a document frequency matrix from a Corpus or Document object.
}
\details{
A wrapper for \code{\link[quanteda]{dfm}}, this classes creates a sparse
document frequency matrix for a Corpus or Document object.  The DFM is attached
to the object and returned to the calling environment.
Source \url{https://cran.r-project.org/web/packages/quanteda/quanteda.pdf}
}
\section{Methods}{

 \describe{
  \item{\code{new(...)}}{Instantiates the class object.}
  \item{\code{execute()}}{Executes the process on the object.}
}
}

\section{DataStudio Family of Classes Overview}{

The DataStudio family of classes include the DataStudio class, which is
responsible for executing one or more data processing commands on a
Corpus or Document resource, the core class group that performs the
data processing functions on the objects, and the command class
group which instructs the DataStudio class to perform the
data processing function on the Corpus or Document object.
}

\section{DataStudio Design}{

This DataStudio family of classes is an implementation of the Strategy
and Command design patterns, as described in the book "Design Patterns:
Elements of Reusable Object-Oriented Software" by Erich Gamma, Richard
Helm, Ralph Johnson and John Vlissides (hence Gang of Four). The
Strategy pattern defines and encapsulates the family of data processing
classes and allows text cleaning algorithms to be selected and
exhanged at runtime. THe Command pattern defines commands that
can be executed in batch mode or at a later time.
}

\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
