% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entity.R
\docType{class}
\name{Entity}
\alias{Entity}
\title{Entity}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Entity
}
\description{
\code{Entity} Base class for all entity related classes
}
\details{
This base class defines members and methods common across all entity related
classes, including Studio, Corpus, Set, and Document classes.
}
\section{Entity methods}{

 \itemize{
  \item{\code{getId()}}{Returns the identifier for the object.}
  \item{\code{getParams()}}{Returns the parameters passed to the constructor.}
 }
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
