% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCSource.R
\docType{class}
\name{FCSource}
\alias{FCSource}
\title{FCSource}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
myFileCollection <- FCSource$new(x)$vector()
}
\arguments{
\item{x}{Character vector containing text, the name of a file, or directory
or a FileCollection object from a supported package.}

\item{name}{Character vector containing the name to assign to the FileCollection object.}

\item{concatenate}{Logical for the vector method. If TRUE, character vectors are
concatenated into a single text for the Document object.}
}
\value{
FileCollection object.
}
\description{
\code{FCSource} Class for sourcing FileCollection objects.
}
\details{
FCSource is the client facing class used to source a FileCollection object from a variety
of formats. Methods support sourcing Corpora objects from character vectors,
JSON, and XML files or directories. Corpora may be imported from other
packages such as Quanteda, TM, KoRpus, and qdap.
}
\section{Methods}{

 \itemize{
  \item{\code{new()}}{Initializes an object of the FCSource class.}
  \item{\code{vector(x, name = NULL)}}{Sources the FileCollection object from a character vector.}
  \item{\code{directory(x, name = NULL)}}{Sources the FileCollection object from a directory.}
  \item{\code{json(x, name = NULL)}}{Sources the FileCollection object from a json source.}
  \item{\code{xml(x, name = NULL)}}{Sources the FileCollection object from an xml source.}
 }
}

\examples{
FileCollection <- FCSource$new()$vector(x)
FileCollection <- FCSource$new()$file(x)
FileCollection <- FCSource$new()$directory(x)
FileCollection <- FCSource$new()$json(x)
FileCollection <- FCSource$new()$xml(x)


}
\seealso{
Other File Collection Source Classes: \code{\link{FCSource0}},
  \code{\link{FCSourceDir}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
