% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCSource0.R
\docType{class}
\name{FCSource0}
\alias{FCSource0}
\title{FCSource0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FCSource0
}
\arguments{
\item{name}{Optional character vector indicating name for FileCollection object.}

\item{x}{Character vector containing text, the name of a file, or directory
from which the FileCollection object will be sourced.}
}
\description{
\code{FCSource0} Abstraction for the classes responsible for sourcing FileCollection objects.
}
\section{Methods}{

 \itemize{
  \item{\code{new()}}{Not implemented for this abstract.}
  \item{\code{source(x, name = NULL)}}{Executes the process of sourcing the FileCollection object.}
 }
}

\seealso{
Other File Collection Source Classes: \code{\link{FCSourceDir}},
  \code{\link{FCSource}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
