% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCSourceDir.R
\docType{class}
\name{FCSourceDir}
\alias{FCSourceDir}
\title{FCSourceDir}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FCSourceDir
}
\arguments{
\item{name}{Optional character vector indicating name for FileCollection object.}

\item{x}{Character string containing the directory from which the FileCollection
object will be source.}
}
\description{
\code{FCSourceDir} Sources a FileCollection object from a directory source.
}
\details{
Sources a FileCollection object from a directory source. Each file yields a single Document
object and a single associated Text object.
}
\section{Methods}{

 \itemize{
  \item{\code{new(X, name = NULL)}}{Initializes an object of the FCSourceDir class.}
  \item{\code{execute()}}{Executes the process of sourcing the FileCollection object.}
 }
}

\seealso{
Other File Collection Source Classes: \code{\link{FCSource0}},
  \code{\link{FCSource}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
