% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File.R
\docType{class}
\name{File}
\alias{File}
\title{File}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
File
}
\arguments{
\item{path}{Character string containing a relative path to a file.}
}
\description{
\code{File} Class respresenting an individual file.
}
\seealso{
Other File Classes: \code{\link{FileSet}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
