% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileSet.R
\docType{class}
\name{FileSet}
\alias{FileSet}
\title{FileSet}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileSet
}
\arguments{
\item{file}{File object.}
}
\description{
\code{FileSet} Class respresenting a set or collection of File objects.
}
\seealso{
Other File Classes: \code{\link{File}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
