% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileSourceCSV.R
\docType{class}
\name{FileSourceCSV}
\alias{FileSourceCSV}
\title{FileSourceCSV}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileSourceCSV
}
\arguments{
\item{x}{Character string containing the file path or URL for the source content}

\item{destination}{Character string indicating the file path for the resulting FileSet object.}
}
\description{
\code{FileSourceCSV} Class creates a File object from .txt file.
}
\seealso{
Other FileSource Family of Classes: \code{\link{FileSource0}},
  \code{\link{FileSourceDOCX}},
  \code{\link{FileSourceJSON}},
  \code{\link{FileSourcePDF}}, \code{\link{FileSourceTXT}},
  \code{\link{FileSourceURL}}, \code{\link{FileSourceXML}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\concept{FileSource Family of Classes}
\keyword{datasets}
