% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileSourceDocx.R
\docType{class}
\name{FileSourceDocx}
\alias{FileSourceDocx}
\title{FileSourceDocx}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileSourceDocx
}
\arguments{
\item{path}{Character string containing a relative path to a file.}
}
\description{
\code{FileSourceTxt} Class creates a File object from .docx file.
}
\seealso{
Other FileSource Family of Classes: \code{\link{FileSource0}},
  \code{\link{FileSourceCsv}},
  \code{\link{FileSourceHtml}},
  \code{\link{FileSourceJSON}},
  \code{\link{FileSourceTxt}}, \code{\link{FileSourceXml}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
