% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileSourceHtml.R
\docType{class}
\name{FileSourceHTML}
\alias{FileSourceHTML}
\title{FileSourceHTML}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileSourceHTML
}
\arguments{
\item{path}{Character string containing a relative path to a file.}
}
\description{
\code{FileSourceHTML} Class creates a File object from .HTML file.
}
\seealso{
Other FileSource Family of Classes: \code{\link{FileSource0}},
  \code{\link{FileSourceCsv}},
  \code{\link{FileSourceJSON}},
  \code{\link{FileSourceTxt}}, \code{\link{FileSourceXml}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
