% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileSourceJSON.R
\docType{class}
\name{FileSourceJSON}
\alias{FileSourceJSON}
\title{FileSourceJSON}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileSourceJSON
}
\arguments{
\item{path}{Character string containing a relative path to a file.}
}
\description{
\code{FileSourceTxt} Class creates a File object from .JSON file.
}
\seealso{
Other FileSource Family of Classes: \code{\link{FileSource0}},
  \code{\link{FileSourceCsv}},
  \code{\link{FileSourceDocx}},
  \code{\link{FileSourceHtml}},
  \code{\link{FileSourceTxt}}, \code{\link{FileSourceXml}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
