% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileSourcePDF.R
\docType{class}
\name{FileSourcePDF}
\alias{FileSourcePDF}
\title{FileSourcePDF}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileSourcePDF
}
\arguments{
\item{x}{Character string containing the file path or URL for the source content}

\item{destination}{Character string indicating the file path for the resulting FileSet object.}
}
\description{
\code{FileSourcePDF} Class creates a File object from .txt file.
}
\seealso{
Other FileSource Family of Classes: \code{\link{FileSource0}},
  \code{\link{FileSourceCSV}},
  \code{\link{FileSourceDOCX}},
  \code{\link{FileSourceJSON}},
  \code{\link{FileSourceTXT}}, \code{\link{FileSourceURL}},
  \code{\link{FileSourceXML}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
