% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileSourceTxt.R
\docType{class}
\name{FileSourceTxt}
\alias{FileSourceTxt}
\title{FileSourceTxt}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileSourceTxt
}
\arguments{
\item{path}{Character string containing a relative path to a file.}
}
\description{
\code{FileSourceTxt} Class creates a File object from .txt file.
}
\seealso{
Other FileSource Family of Classes: \code{\link{FileSource0}},
  \code{\link{FileSourceCsv}},
  \code{\link{FileSourceDocx}},
  \code{\link{FileSourceHtml}},
  \code{\link{FileSourceJSON}}, \code{\link{FileSourceXml}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
