% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileSourceURL.R
\docType{class}
\name{FileSourceURL}
\alias{FileSourceURL}
\title{FileSourceURL}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileSourceURL
}
\arguments{
\item{origin}{Character string containing the file path or URL for the origin content}

\item{destination}{Character string specifying the directory in which the text files will be stored.}
}
\description{
\code{FileSourceURL} Class creates a File object from .txt file.
}
\seealso{
Other FileSource Family of Classes: \code{\link{FileSource0}},
  \code{\link{FileSourceCSV}},
  \code{\link{FileSourceDOCX}},
  \code{\link{FileSourceJSON}},
  \code{\link{FileSourcePDF}}, \code{\link{FileSourceTXT}},
  \code{\link{FileSourceXML}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
