% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileSourceXML.R
\docType{class}
\name{FileSourceXML}
\alias{FileSourceXML}
\title{FileSourceXML}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileSourceXML
}
\arguments{
\item{x}{Character string containing the file path or URL for the source content}

\item{destination}{Character string indicating the file path for the resulting FileSet object.}
}
\description{
\code{FileSourceXML} Class creates a File object from .txt file.
}
\seealso{
Other FileSource Family of Classes: \code{\link{FileSource0}},
  \code{\link{FileSourceCSV}},
  \code{\link{FileSourceDOCX}},
  \code{\link{FileSourceJSON}},
  \code{\link{FileSourcePDF}}, \code{\link{FileSourceTXT}},
  \code{\link{FileSourceURL}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
