% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveCtrlApp.R
\docType{class}
\name{FileStudio}
\alias{FileStudio}
\title{FileStudio}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileStudio
}
\arguments{
\item{path}{Character string indicating the path to a file to be repaired.}

\item{code}{Numeric vector containing up to 32 numbers between 0 and 31,
in which each number relates to the decimal ASCII code for the control
characters to remove from the file.}
}
\description{
\code{FileStudio} Class for treating and repair files.
}
\details{
Class responsible for treating and performing repairs on text files
at binary level.
}
\seealso{
Other File Classes: \code{\link{FileCollection}},
  \code{\link{File}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
