% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileStudio0.R
\docType{class}
\name{FileStudio0}
\alias{FileStudio0}
\title{FileStudio0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileStudio0
}
\description{
\code{FileStudio0}  Abstract class for the FileStudio family of classes.
}
\details{
Abstract strategy class which defines the methods common to the FileStudio
family of classes.
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
