% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fold.R
\docType{class}
\name{Fold}
\alias{Fold}
\title{Fold}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Fold
}
\arguments{
\item{name}{Character string containing the name for the Fold object.}
}
\description{
\code{Fold} Class containing 1 of 'K' folds.  Eeach fold containing a training and test set.
}
\section{Core Methods}{

 \itemize{
  \item{\code{new(name = NULL)}}{Initializes an object of the Fold class.}
 }
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
