% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSSource.R
\docType{class}
\name{FSSource}
\alias{FSSource}
\title{FSSource}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
myFileSet <- FSSource$new(x)$vector()
}
\arguments{
\item{x}{Character vector containing text, the name of a file, or directory
or a FileSet object from a supported package.}

\item{name}{Character vector containing the name to assign to the FileSet object.}

\item{concatenate}{Logical for the vector method. If TRUE, character vectors are
concatenated into a single text for the Document object.}
}
\value{
FileSet object.
}
\description{
\code{FSSource} Class for sourcing FileSet objects.
}
\details{
FSSource is the client facing class used to source a FileSet object from a variety
of formats. Methods support sourcing Corpora objects from character vectors,
JSON, and XML files or directories. Corpora may be imported from other
packages such as Quanteda, TM, KoRpus, and qdap.
}
\section{Methods}{

 \itemize{
  \item{\code{new()}}{Initializes an object of the FSSource class.}
  \item{\code{vector(x, name = NULL)}}{Sources the FileSet object from a character vector.}
  \item{\code{directory(x, name = NULL)}}{Sources the FileSet object from a directory.}
  \item{\code{json(x, name = NULL)}}{Sources the FileSet object from a json source.}
  \item{\code{xml(x, name = NULL)}}{Sources the FileSet object from an xml source.}
 }
}

\examples{
FileSet <- FSSource$new()$vector(x)
FileSet <- FSSource$new()$file(x)
FileSet <- FSSource$new()$directory(x)
FileSet <- FSSource$new()$json(x)
FileSet <- FSSource$new()$xml(x)


}
\seealso{
Other File Set Source Classes: \code{\link{FSSource0}},
  \code{\link{FSSourceDir}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
