% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSSource0.R
\docType{class}
\name{FSSource0}
\alias{FSSource0}
\title{FSSource0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FSSource0
}
\arguments{
\item{name}{Optional character vector indicating name for FileSet object.}

\item{x}{Character vector containing text, the name of a file, or directory
from which the FileSet object will be sourced.}
}
\description{
\code{FSSource0} Abstraction for the classes responsible for sourcing FileSet objects.
}
\section{Methods}{

 \itemize{
  \item{\code{new()}}{Not implemented for this abstract.}
  \item{\code{source(x, name = NULL)}}{Executes the process of sourcing the FileSet object.}
 }
}

\seealso{
Other File Set Source Classes: \code{\link{FSSourceDir}},
  \code{\link{FSSource}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
