% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSSourceDir.R
\docType{class}
\name{FSSourceDir}
\alias{FSSourceDir}
\title{FSSourceDir}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FSSourceDir
}
\arguments{
\item{name}{Optional character vector indicating name for FileSet object.}

\item{x}{Character string containing the directory from which the FileSet
object will be source.}
}
\description{
\code{FSSourceDir} Sources a FileSet object from a directory source.
}
\details{
Sources a FileSet object from a directory source. Each file yields a single Document
object and a single associated Text object.
}
\section{Methods}{

 \itemize{
  \item{\code{new(X, name = NULL)}}{Initializes an object of the FSSourceDir class.}
  \item{\code{execute()}}{Executes the process of sourcing the FileSet object.}
 }
}

\seealso{
Other File Set Source Classes: \code{\link{FSSource0}},
  \code{\link{FSSource}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
