% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_json_files.R
\name{get_json_files}
\alias{get_json_files}
\title{JSON Files Container}
\usage{
get_json_files(root_path, ...)
}
\arguments{
\item{root_path}{Path to root folder containing JSON files.}

\item{...}{Additional parameters passed to \code{list.files()}.}
}
\value{
A character vector of file paths. (Deprecated.)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated and will be removed in a future release.
Please use \code{\link[base:list.files]{base::list.files()}} directly instead:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{files <- list.files(root_path, pattern = "\\\\.json$", recursive = TRUE, full.names = TRUE)
}\if{html}{\out{</div>}}

Then pass the resulting character vector of file paths to \code{\link[=from_json_to_df]{from_json_to_df()}}.
}
