% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sec_master_files.R
\name{get_sec_master_files}
\alias{get_sec_master_files}
\title{SEC Master Files Container}
\usage{
get_sec_master_files(
  root_path,
  pattern = NULL,
  fyear = NULL,
  drop_late = FALSE
)
}
\arguments{
\item{root_path}{Character giving the path to the root folder location (see 'Details').}

\item{pattern}{An optional \link[base:regex]{base::regex}. Only file names which match the regular expression
will be returned as in \link{list.files}. Default to \code{NULL}.}

\item{fyear}{Optional numeric vector specifying the time window in fiscal years. When \code{NULL},
all the fiscal years detected in \code{root_path} are considered. Default to \code{NULL}.}

\item{drop_late}{When \code{TRUE}, remove filings referring to distant fiscal period. Default to
\code{FALSE}.}
}
\value{
A \code{data.table} containing all the SEC master files as detected by the function
with the following columns:

\item{\code{cik}}{The Central Index Key as given by the SEC EDGAR database. Integer.}
\item{\code{cname}}{The company name. Character.}
\item{\code{type}}{The filing type. Character.}
\item{\code{date_filed}}{The filing date on the SEC. IDat.}
\item{\code{fyear_end}}{The fiscal year end date. IDat.}
\item{\code{sic}}{The four-digit SIC industry code. Integer.}
\item{\code{state_of_inc}}{U.S. state of incorporation (abbreviation). Character.}
\item{\code{state_location}}{U.S. state location (abbreviation). Character.}
\item{\code{filing_detail}}{Pointer to filing summary page in HTML. Character.}
\item{\code{filing_html}}{Pointer to filing in HTML. Character.}
\item{\code{filing_txt}}{Pointer to complete submission file in txt. Character.}
\item{\code{filename}}{Filename. Character.}
}
\description{
Create a container list with the pointers to all the JSON files detected in the input folder.
}
\details{
The function works in conjunction with the EDGAR crawler structure. In general, this creates three
main folders as containers: (i) the original HTML files, (ii) the metadata of each filing, (iii)
the extracted items in JSON format. Each of these structures is usually, and conveniently, organized
by fiscal year. This is not an imposition but rather a strong suggestion. Ideally, one could put
all the filings into the root_path and work from that.

This function works on the metadata files, also called SEC master files. Typically, such data are
stored in a .csv file and are located in \code{root_path} with no sub-directory. It is recommended
to specify at least a \code{pattern} to detect those files.
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}
}
