% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalFunctions.R
\name{identity}
\alias{identity}
\title{identity}
\usage{
identity(cls, name = NULL)
}
\arguments{
\item{cls}{Character string containing the class of the object}

\item{name}{Character string containing the name of the object. Optional. If
NULL, a name is generated from the class and id.}
}
\description{
\code{identity} Creates identity card which are assigned to all business domain objects.
}
\seealso{
Other Internal Functions: \code{\link{proper}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
