% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\docType{class}
\name{IO}
\alias{IO}
\title{IO}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
myText <- IO$new(name = "skiReport", x = avalanche)
}
\arguments{
\item{path}{Character string containing the file path}

\item{content}{Character vectors containing the content to be written}
}
\value{
Read method returns character vectors containing text. The write method
returns an instance of the class.
}
\description{
\code{IO} Class responsible for Text IO within the NLPStudio package.
}
\details{
Class responsible for reading and writing texts into and from the NLPStudio
package. Read methods support a range of formats and has a repair option
that converts select ASCII control characters to spaces by default.
}
\section{Methods}{

 \itemize{
  \item{\code{read(path)}}{Reads text. If the file is a ".txt" file, the repair
  process replaces select ASCII control characters with spaces.}
  \item{\code{write(path, content)}}{Write text to designated file location.}
 }
}

\seealso{
Other IO Classes: \code{\link{IOFactory}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
