% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOFactory.R
\docType{class}
\name{IOFactory}
\alias{IOFactory}
\title{IOFactory}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
IOFactory
}
\arguments{
\item{path}{Character string containing the path for file}
}
\value{
Concrete IO strategy.
}
\description{
\code{IOFactory} Factory class that creates the appropriate IO class.
}
\details{
Class creates IO objects based upon the file type
}
\section{IOFactory methods}{

\describe{
 \item{\code{new()}}{Instantiates the factory.}
 \item{\code{getFileStrategy(fileName)}}{Instantiates and returns the appropriate File Strategy object.}
}
}

\seealso{
Other IO Classes: \code{\link{IO}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
