% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IORdata.R
\docType{class}
\name{IORdata}
\alias{IORdata}
\title{IORdata}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
IORdata$new()$read(path)
}
\arguments{
\item{path}{Character string designated the relative path to a file.}

\item{content}{Content to be written to the designated file path.}
}
\value{
Character vector if read method is called. For the write method, a logical value
of TRUE is returned if the write was successful, FALSE otherwise.
}
\description{
\code{IORdata} Class responsible for reading and writing Rdata files.
}
\section{IO Methods}{

 \itemize{
  \item{\code{read(path)}}{Reads content from a specified file path.}
  \item{\code{write(path, content)}}{Writes the content to the designated
  file path.}
 }
}

\seealso{
Other Input / Output Classes: \code{\link{IOBin}},
  \code{\link{IOCSV}}, \code{\link{IOText}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
