% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOText.R
\docType{class}
\name{IOText}
\alias{IOText}
\title{IOText}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
IOText$new()$read(path = path)
}
\arguments{
\item{path}{Character string containing the relative file path}

\item{content}{Character vector to be written to file}
}
\value{
Character string if read method is called. The write method
returns TRUE if the write was sucessful, FALSE otherwise.
}
\description{
\code{IOText} Class responsible for reading and writing text files.
}
\section{Methods}{

 \describe{
  \item{\code{read(path)}}{Read method.}
  \item{\code{write(path, content)}}{Write method.}
}
}

\seealso{
Other Input / Output Classes: \code{\link{IOBin}},
  \code{\link{IOCSV}}, \code{\link{IORdata}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
