% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KFold.R
\docType{class}
\name{KFold}
\alias{KFold}
\title{KFold}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
KFold
}
\arguments{
\item{x}{Corpus object.}

\item{k}{Numeric of cvSets}

\item{stratify}{Logical. If TRUE (default), KFolds will be taken from
each document in accordance with the proportions or counts indicated
in the 'n' parameter.}

\item{seed}{Numeric used to initialize a pseudorandom number generator.}
}
\description{
\code{KFold} Creates a K-Fold cross-validation set.
}
\section{Methods}{

 \itemize{
  \item{\code{new()}}{Not implemented.}
 }
}

\seealso{
Other Corpus Sample Family of Classes: \code{\link{Sample0}},
  \code{\link{Sample}}, \code{\link{Segment}}

Other CorpusStudio Family of Classes: \code{\link{CorpusStudio}},
  \code{\link{Sample0}}, \code{\link{Sample}},
  \code{\link{Segment}}, \code{\link{Split}},
  \code{\link{TokenizerNLP}}, \code{\link{TokenizerQ}},
  \code{\link{Tokenizer}}, \code{\link{Token}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\concept{Corpus Sample Family of Classes}
\concept{CorpusStudio Family of Classes}
\keyword{datasets}
