% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNBuild.R
\docType{class}
\name{KNBuild}
\alias{KNBuild}
\title{KNBuild}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
KNBuild
}
\arguments{
\item{x}{Language model object}
}
\description{
\code{KNBuild} Computes counts used to estimate Kneser Ney probabilities
}
\details{
Computes counts used in the probability calculations for the Kneser Ney
probabilities.
}
\seealso{
Other KNStudio Classes: \code{\link{KNEstimate}},
  \code{\link{MKNBuild}}, \code{\link{SLMEvaluate}}

Other SLMStudio Classes: \code{\link{KNBuild0}},
  \code{\link{KNEstimate}}, \code{\link{MKNBuild}},
  \code{\link{SLMCorpora}}, \code{\link{SLMEvaluate}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
