% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNBuild0.R
\docType{class}
\name{KNBuild0}
\alias{KNBuild0}
\title{KNBuild0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
KNBuild0
}
\description{
\code{KNBuild0} Base class for Kneser-Ney Build Family of classes.
}
\details{
Defines the default methods for computing the counts used in the
probability calculations for the Kneser Ney and Modified Kneser-Ney
probabilities.
}
\seealso{
Other SLMStudio Classes: \code{\link{KNBuild}},
  \code{\link{KNEstimate}}, \code{\link{MKNBuild}},
  \code{\link{SLMCorpora}}, \code{\link{SLMEvaluate}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
