% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNCounts.R
\docType{class}
\name{KNCounts}
\alias{KNCounts}
\title{KNCounts}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
KNCounts
}
\arguments{
\item{x}{Language model object}
}
\description{
\code{KNCounts} Computes counts used to estimate Kneser Ney probabilities
}
\details{
Computes counts used in the probability calculations for the Kneser Ney
probabilities.
}
\seealso{
Other KNStudio Classes: \code{\link{KNEstimate}},
  \code{\link{KNEvaluate}}, \code{\link{MKNEvaluate}}

Other LMStudio Classes: \code{\link{KNEstimate}},
  \code{\link{KNEvaluate}}, \code{\link{LMCorpora}},
  \code{\link{LMStudio0}}, \code{\link{MKNCounts}},
  \code{\link{MKNEstimate}}, \code{\link{MKNEvaluate}},
  \code{\link{MKNStudio0}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
