% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNEvaluate.R, R/KNEvaluate2.R
\docType{class}
\name{KNEvaluate}
\alias{KNEvaluate}
\alias{KNEvaluate}
\title{KNEvaluate}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
KNEvaluate

KNEvaluate
}
\arguments{
\item{x}{Language model object}

\item{x}{Language model object}
}
\description{
\code{KNEvaluate} Evaluates a Kneser-Ney Language Model on Test Set

\code{KNEvaluate} Evaluates a Kneser-Ney Language Model on Test Set
}
\details{
Evaluates a Kneser-Ney Language Model on Test Set by computing
test set perplexity

Evaluates a Kneser-Ney Language Model on Test Set by computing
test set perplexity
}
\seealso{
Other KNStudio Classes: \code{\link{KNCounts}},
  \code{\link{KNEstimate}}, \code{\link{KNStudio0}}

Other LMStudio Classes: \code{\link{KNCounts}},
  \code{\link{KNEstimate}}, \code{\link{KNStudio0}},
  \code{\link{LMCorpora}}, \code{\link{LMStudio0}},
  \code{\link{MKNCounts}}, \code{\link{MKNEstimate}},
  \code{\link{MKNPrepDocuments0}}, \code{\link{MKNStudio0}}

Other KNStudio Classes: \code{\link{KNCounts}},
  \code{\link{KNEstimate}}, \code{\link{KNStudio0}}

Other LMStudio Classes: \code{\link{KNCounts}},
  \code{\link{KNEstimate}}, \code{\link{KNStudio0}},
  \code{\link{LMCorpora}}, \code{\link{LMStudio0}},
  \code{\link{MKNCounts}}, \code{\link{MKNEstimate}},
  \code{\link{MKNPrepDocuments0}}, \code{\link{MKNStudio0}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}

John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
