% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNModel.R
\docType{class}
\name{KNModel}
\alias{KNModel}
\title{KNModel}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
KNModel
}
\arguments{
\item{corpora}{training and test Corpora object upon which the model
will be trained and evaluated}

\item{size}{Numeric indicating the model size in terms of nGrams. Defaults to trigram model.}

\item{open}{Logical indicating whether the vocabulary is open
or closed. Open indicates that it is possible to encounter out of vocabulary
words in the test set. If TRUE, OOV processing will be performed on
the training set. Default is TRUE.}
}
\description{
\code{KNModel} Class containing the Kneser Ney Language Model object.
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
