% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNStudio.R
\docType{class}
\name{KNStudio}
\alias{KNStudio}
\title{KNStudio}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
KNStudio
}
\arguments{
\item{config}{SLMConfig object containing the model parameters}

\item{corpora}{SLMCorpora object containing the test set.}
}
\description{
\code{KNStudio} Builds Kneser Ney language model object.
}
\seealso{
Other Statistical Language Model Classes: \code{\link{SLM0}},
  \code{\link{SLM}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
