% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File.R, R/Lab.R
\docType{class}
\name{Lab}
\alias{Lab}
\alias{Lab}
\title{File}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Lab

Lab
}
\arguments{
\item{desc}{A chararacter string containing the description of the Lab}

\item{document}{An object of the DocumentCollection class to be added to the Lab object's list of document collections.}

\item{name}{A character string containing the name of the Lab object. This variable is used in the instantiation and remove methods.}

\item{purge}{A boolean variable.  Used in the removeDocument method. Indicates whether to purge a document from memory, disk, and state.}

\item{type}{A character string indicating the return format for the "get" methods.  Valid values are c("object", "list", "df")}

\item{name}{A character string containing the name of the Lab object. This variable is used in the instantiation and remove methods.}

\item{desc}{A chararacter string containing the description of the Lab}

\item{document}{An object of the DocumentCollection class to be added to the Lab object's list of document collections.}
}
\description{
\code{File} Class responsible for copy, moving and deleting files on disk

\code{Lab} Class that contains document collections and the environment in which NLP happens.
}
\details{
This class is a wrapper for procedural copy, move, delete functions in R

The environment in which NLP happens. There are two groups of methods. The
first group allows clients to instantiate, retrieve, print, enter, leave,
and archive a Lab object.  The second set of methods allow clients to retrieve
the contained documents, add a document, and remove a document.
}
\section{Fields}{

\describe{
\item{\code{parent}}{An object of class NLPStudio that contains the Lab object. This is always equal to the singleton object "nlpStudio".}

\item{\code{parentName}}{Character string indicating the name of the parent, which is always "nlpStudio".}

\item{\code{path}}{Character string indicating the directory path of the Lab object.}

\item{\code{documents}}{List of documents of class DocumentCollection that are members of the Lab object.}

\item{\code{directories}}{List of directories included in each lab object.}

\item{\code{modified}}{Datetime variable indicating the date/time the object was created.}

\item{\code{created}}{Datetime variable indicating the date/time the object was last modified.}
}}

\section{Methods}{

\describe{
 \item{\code{copy(from, to}}{Copies file(s) from one object's file location to another object's file location.}
 \item{\code{move(from, to}}{Move file(s) from one object's file location to another object's file location.}
 \item{\code{delete(from}}{Delete file(s) for an object.}
}
}

\section{Document Methods}{

\describe{
 \item{\code{getDocuments(type = "list")}}{Returns a list of documents in a range of formats.  Valid types (formats) are c("object", "list", "df").  The default is "list".}
 \item{\code{addDocument(document)}}{Adds a document to the Lab object's list of document collections.}
 \item{\code{removeDocument(document, purge = FALSE)}}{Removes a document from the Lab object's list of document collections.  If the purge variable is set to TRUE, the Lab object is archived, the document is removed from the global environment, its directory is deleted, and it is removed from state.}
}
}

\section{Lab Methods}{

\describe{
 \item{\code{new(name, desc = NULL)}}{Creates an object of Lab Class}
 \item{\code{desc}}{A getter/setter method allowing clients to retrieve and set the Lab description variable.}
 \item{\code{getObject()}}{Retrieves the meta data for the Lab object.}
 \item{\code{getAncestor()}}{Retrieves the parent object for the Lab object.}
 \item{\code{setAncestor()}}{Sets the parent object for the Lab object.}
 \item{\code{accept(visitor)}}{Accepts a visitor and responds by dispatching the appropriate method, sending itself as a parameter.}
}
}

\section{Document collection (child) Methods}{

\describe{
 \item{\code{getChildren()}}{Retrieves a list containing meta data for child objects of the DocumentCollection class.}
 \item{\code{addChild(document)}}{Adds a child document, an object of the DocumentCollection class, to the Lab object.}
 \item{\code{removeChild(document)}}{Removes a child document, an object of the DocumentCollection class, from the Lab object.}
}
}

\author{
John James, \email{jjames@datasciencesalon.org}

John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
