% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{listFiles}
\alias{listFiles}
\title{listFiles}
\usage{
listFiles(x)
}
\description{
\code{listFiles} Returns the list of files associated with a directory or a glob
}
\seealso{
Other Internal Functions: \code{\link{printHeading}},
  \code{\link{proper}}, \code{\link{slice}},
  \code{\link{tokenize}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
