% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMCorpora.R
\docType{class}
\name{LMCorpora}
\alias{LMCorpora}
\title{LMCorpora}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LMCorpora
}
\arguments{
\item{x}{Language model object}
}
\description{
\code{LMCorpora} Class responsible for preparing training Corpus objects for language modeling.
}
\details{
Class responsible for preparing training Corpus objects for language modeling.
}
\seealso{
Other LMStudio Classes: \code{\link{KNBuild}},
  \code{\link{KNEstimate}}, \code{\link{KNEvaluate}},
  \code{\link{LMStudio0}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
