% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMStudio0.R
\docType{class}
\name{LMStudio0}
\alias{LMStudio0}
\title{LMStudio0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LMStudio0
}
\arguments{
\item{x}{Language model object}
}
\description{
\code{LMStudio0} Abstract class for the language model family of classes.
}
\details{
This abstract class defines a common interface and methods of the LMStudio
family of classes.
}
\seealso{
Other LMStudio Classes: \code{\link{KNCounts}},
  \code{\link{KNEstimate}}, \code{\link{KNEvaluate}},
  \code{\link{LMCorpora}}, \code{\link{MKNCounts}},
  \code{\link{MKNEstimate}}, \code{\link{MKNEvaluate}},
  \code{\link{MKNStudio0}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
