% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_tokens.R
\name{lookup_tokens}
\alias{lookup_tokens}
\title{Fast Tokens Lookup}
\usage{
lookup_tokens(x, ncores, ...)
}
\arguments{
\item{x}{A \strong{quanteda} \link[quanteda:tokens]{quanteda::tokens} object.}

\item{ncores}{The number of \link[future:multisession]{future::multisession} workers to be allocated for
the lookup.}

\item{...}{Additional arguments passed to \link[quanteda:tokens_lookup]{quanteda::tokens_lookup}.}
}
\value{
A \link[quanteda:tokens]{quanteda::tokens} object.
}
\description{
Lookup a \link[quanteda:tokens]{quanteda::tokens} object in parallel via \strong{\link[future:future]{future::future}}.
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}
}
