% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meta.R
\docType{class}
\name{Meta}
\alias{Meta}
\title{Meta}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Meta
}
\description{
\code{Meta} Contains metadata and methods for maintaining and reporting them.
}
\details{
Contains object metadata and the methods for adding, removing, and reporting
metadata.  Also includes convenience methods for updating state metadata
relating to datetimes of creating, update, and access.
}
\section{Meta methods}{

 \itemize{
  \item{\code{new()}}{Intantiates an instance of the Meta class.}
  \item{\code{set(key, value)}}{Adds metadata via key value pairs.}
  \item{\code{get(key)}}{Returns metadata associated with the key parameter.}
  \item{\code{created()}}{Sets state datetime parameters when an object is created.}
  \item{\code{modified()}}{Sets state datetime parameters when an object is modified.}
  \item{\code{accessed()}}{Sets state datetime parameters when an object is accessed.}
 }
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
