% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MKN.R
\docType{class}
\name{MKN}
\alias{MKN}
\title{MKN}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MKN
}
\arguments{
\item{x}{Corpus object upon which the model will be trained.}

\item{size}{Numeric indicating the model size in terms of nGrams. Defaults to trigram model.}

\item{open}{Logical indicating whether the vocabulary is open
or closed. Open indicates that it is possible to encounter out of vocabulary
words in the test set. If TRUE, OOV processing will be performed on
the training set. Default is TRUE.}
}
\description{
\code{MKN} Modified Kneser Ney language model object.
}
\details{
Modified Kneser Ney language model object.
}
\seealso{
Other LMStudio Classes: \code{\link{KNCounts}},
  \code{\link{KNDocument}}, \code{\link{KNStudio0}},
  \code{\link{LMStudio0}}, \code{\link{MKNCounts}},
  \code{\link{MKNDocument}}, \code{\link{MKNEstimate}},
  \code{\link{MKNStudio0}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
