% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MKNBuild.R
\docType{class}
\name{MKNBuild}
\alias{MKNBuild}
\title{MKNBuild}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MKNBuild
}
\arguments{
\item{x}{Language model object}
}
\description{
\code{MKNBuild} Computes counts used to estimate Modified Kneser Ney probabilities
}
\details{
Computes counts used in the probability calculations for the
Modified Kneser Ney probabilities.
}
\seealso{
Other KNStudio Classes: \code{\link{KNBuild}},
  \code{\link{KNEstimate}}, \code{\link{SLMEvaluate}}

Other SLMStudio Classes: \code{\link{KNBuild0}},
  \code{\link{KNBuild}}, \code{\link{KNEstimate}},
  \code{\link{SLMCorpora}}, \code{\link{SLMEvaluate}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
