% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MKNCounts.R
\docType{class}
\name{MKNCounts}
\alias{MKNCounts}
\title{MKNCounts}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MKNCounts
}
\arguments{
\item{x}{Language model object}
}
\description{
\code{MKNCounts} Prepares text data for downstream processing.
}
\details{
Gathers text data, creates sentence tokens and, if open is TRUE, converts
hapax legomenon to unknown tokens.
}
\seealso{
Other MKNStudio Classes: \code{\link{MKNDocument}},
  \code{\link{MKNEstimate}}, \code{\link{MKNStudio0}}

Other LMStudio Classes: \code{\link{KNCounts}},
  \code{\link{LMStudio0}}, \code{\link{MKNDocument}},
  \code{\link{MKNEstimate}}, \code{\link{MKNStudio0}},
  \code{\link{MKN}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
