% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MKNCounts.R, R/MKNData.R
\docType{class}
\name{MKNData}
\alias{MKNData}
\alias{MKNData}
\title{MKNData}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MKNData

MKNData
}
\arguments{
\item{lm}{Language model object}

\item{x}{Language model object}
}
\description{
\code{MKNData} Prepares text data for downstream processing.

\code{MKNData} Prepares text data for downstream processing.
}
\details{
Gathers text data, creates sentence tokens and, if open is TRUE, converts
hapax legomenon to unknown tokens.

Gathers text data, creates sentence tokens and, if open is TRUE, converts
hapax legomenon to unknown tokens.
}
\seealso{
Other MKNStudio Classes: \code{\link{MKNStudio0}}

Other LMStudio Classes: \code{\link{LMStudio0}},
  \code{\link{MKNStudio0}}, \code{\link{MKN}}

Other MKNStudio Classes: \code{\link{MKNStudio0}}

Other LMStudio Classes: \code{\link{LMStudio0}},
  \code{\link{MKNStudio0}}, \code{\link{MKN}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}

John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
