% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MKNDocument.R
\docType{class}
\name{MKNDocument}
\alias{MKNDocument}
\title{MKNDocument}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MKNDocument
}
\arguments{
\item{x}{Language model object}
}
\description{
\code{MKNDocument} Prepares text data for downstream processing.
}
\details{
Gathers text data, creates sentence tokens and, if open is TRUE, converts
hapax legomenon to unknown tokens.
}
\seealso{
Other MKNStudio Classes: \code{\link{MKNCounts}},
  \code{\link{MKNEstimate}}, \code{\link{MKNStudio0}}

Other LMStudio Classes: \code{\link{KNCounts}},
  \code{\link{KNDocument}}, \code{\link{KNEstimate}},
  \code{\link{KNStudio0}}, \code{\link{LMStudio0}},
  \code{\link{MKNCounts}}, \code{\link{MKNEstimate}},
  \code{\link{MKNStudio0}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
