% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MKNEstimate.R
\docType{class}
\name{MKNEstimate}
\alias{MKNEstimate}
\title{MKNEstimate}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MKNEstimate
}
\arguments{
\item{x}{Language model object}
}
\description{
\code{MKNEstimate} Computes Modified Kneser Ney estimates
}
\details{
Computes Modified Kneser Ney estimates for all nGrams in the Model
}
\seealso{
Other MKNStudio Classes: \code{\link{MKNCounts}},
  \code{\link{MKNStudio0}}

Other LMStudio Classes: \code{\link{KNCounts}},
  \code{\link{KNEstimate}}, \code{\link{KNEvaluate}},
  \code{\link{LMCorpora}}, \code{\link{LMStudio0}},
  \code{\link{MKNCounts}}, \code{\link{MKNEvaluate}},
  \code{\link{MKNStudio0}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
