% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MKNEvaluate.R
\docType{class}
\name{MKNEvaluate}
\alias{MKNEvaluate}
\title{MKNEvaluate}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MKNEvaluate
}
\arguments{
\item{x}{Language model object}
}
\description{
\code{MKNEvaluate} Evaluates a Kneser-Ney Language Model on Test Set
}
\details{
Evaluates a Kneser-Ney Language Model on Test Set by computing
test set perplexity
}
\seealso{
Other KNStudio Classes: \code{\link{KNCounts}},
  \code{\link{KNEstimate}}, \code{\link{KNEvaluate}}

Other LMStudio Classes: \code{\link{KNCounts}},
  \code{\link{KNEstimate}}, \code{\link{KNEvaluate}},
  \code{\link{LMCorpora}}, \code{\link{LMStudio0}},
  \code{\link{MKNCounts}}, \code{\link{MKNEstimate}},
  \code{\link{MKNStudio0}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
