% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MKNDocument.R
\docType{class}
\name{MKNPrepDocuments0}
\alias{MKNPrepDocuments0}
\title{MKNPrepDocuments0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MKNPrepDocuments0
}
\arguments{
\item{x}{Language model object}
}
\description{
\code{MKNPrepDocuments0} Prepares text data for downstream processing.
}
\details{
Gathers text data, creates sentence tokens and, if open is TRUE, converts
hapax legomenon to unknown tokens.
}
\seealso{
Other MKNStudio Classes: \code{\link{MKNCounts}},
  \code{\link{MKNEstimate}}, \code{\link{MKNStudio0}}

Other LMStudio Classes: \code{\link{KNCounts}},
  \code{\link{KNEstimate}}, \code{\link{KNEvaluate}},
  \code{\link{KNStudio0}}, \code{\link{LMCorpora}},
  \code{\link{LMStudio0}}, \code{\link{MKNCounts}},
  \code{\link{MKNEstimate}}, \code{\link{MKNStudio0}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
