% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MKNStudio.R
\docType{class}
\name{MKNStudio}
\alias{MKNStudio}
\title{MKNStudio}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MKNStudio
}
\arguments{
\item{config}{SLMConfig object containing the model parameters}

\item{corpora}{SLMCorpora object containing the test set.}
}
\description{
\code{MKNStudio} Builds Kneser Ney language model object.
}
\seealso{
Other Statistical Language Model Classes: \code{\link{KNStudio}},
  \code{\link{KN}}, \code{\link{MKN}}, \code{\link{SLM}}

Other Kneser Ney Model Classes: \code{\link{KNStudio}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
