% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MKNStudio0.R
\docType{class}
\name{MKNStudio0}
\alias{MKNStudio0}
\title{MKNStudio0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MKNStudio0
}
\arguments{
\item{x}{Language model object}
}
\description{
\code{MKNStudio0} Abstract class for the Kneser Ney language model state classes.
}
\details{
This abstract class defines a common interface and methods of the
Kneser Ney language model state classes.
}
\seealso{
Other MKNStudio Classes: \code{\link{MKNCounts}},
  \code{\link{MKNEstimate}},
  \code{\link{MKNPrepDocuments0}}

Other LMStudio Classes: \code{\link{KNCounts}},
  \code{\link{KNEstimate}}, \code{\link{KNEvaluate}},
  \code{\link{KNStudio0}}, \code{\link{LMCorpora}},
  \code{\link{LMStudio0}}, \code{\link{MKNCounts}},
  \code{\link{MKNEstimate}},
  \code{\link{MKNPrepDocuments0}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
