% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoveFileSet.R
\docType{class}
\name{MoveFileSet}
\alias{MoveFileSet}
\title{MoveFileSet}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MoveFileSet
}
\arguments{
\item{to}{The directory into which the FileSet shall be moved.}

\item{fileStudio}{FileStudio object}
}
\description{
\code{MoveFileSet} Encapsulates the command to move a FileSet object.
}
\seealso{
Other FileStudio Command Classes: \code{\link{BuildCorpus}},
  \code{\link{CleanCorpus}}, \code{\link{CopyFileSet}},
  \code{\link{ReadCorpus}}, \code{\link{RepairFileSet}},
  \code{\link{SampleCorpus}}, \code{\link{Source}},
  \code{\link{SplitCorpus}}, \code{\link{TokenizeCorpus}},
  \code{\link{WriteCorpus}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
