% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGrammer.R
\docType{class}
\name{NGrammer}
\alias{NGrammer}
\title{NGrammer}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NGrammer <- NGrammer$new()
}
\arguments{
\item{x}{The source Document object.}
}
\value{
NGrammer object, containing the NGrammer for a single Document object.
}
\description{
\code{NGrammer} Class representing nGram Document objects.

@param x The principal object
 @param key Character string or a vector of strings containing a key or
 keys for metadata key/value pair(s).
 @param value Character or numeric, or a vector thereof, containing a value
 or values for metadata key/value pair(s).
 @param format Character string used in the metadata method. Indicates
 whether to return the metadata as  list of lists, or a data.frame.This method
 returns metadata in a list if no key/value pair is provided. The list
 may contain other lists or data frames associated with the type of
 metadata. This parameter specifies whether to return the metadata
 as a list of lists, or a list of data frames. Valid values are:
 c('list', 'data.frame')
}
\details{
Class containing the nGramd representation of a Document object.
}
\section{Core Methods}{

 \itemize{
  \item{\code{new(x)}}{Initializes an object of the NGrammer class.}
  \item{\code{content}}{Active binding used to set and retrieve NGrammer content. NGrammer
  content may be changed via assignment. Referencing this method retrieves the current
  NGrammer content.}
 }
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
