% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGramSet.R
\docType{class}
\name{NGramSet}
\alias{NGramSet}
\title{NGramSet}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
skiReport <- NGramSet$new(name = "skiReport", purpose = 'Train')
}
\arguments{
\item{name}{Character string containing the name for the NGramSet object.}

\item{purpose}{Character string used to indicate how the document will be used, e.g. 'train', 'test'.}

\item{note}{Character string containing a comment associated with a call to the
text method. The texts of the note variable are written to the NGramSets
log. This is used to track changes to the text, perhaps made during preprocessing.}
}
\value{
NGramSet object, containing the NGramSet text, the metadata and
the methods to manage both.
}
\description{
\code{NGramSet} Class containing a collection of NGram objects

@param x The principal object
 @param key Character string or a vector of strings containing a key or
 keys for metadata key/value pair(s).
 @param value Character or numeric, or a vector thereof, containing a value
 or values for metadata key/value pair(s).
 @param format Character string used in the metadata method. Indicates
 whether to return the metadata as  list of lists, or a data.frame.This method
 returns metadata in a list if no key/value pair is provided. The list
 may contain other lists or data frames associated with the type of
 metadata. This parameter specifies whether to return the metadata
 as a list of lists, or a list of data frames. Valid values are:
 c('list', 'data.frame')
}
\details{
Class contains NGram objects
}
\section{Core Methods}{

 \itemize{
  \item{\code{new(name = NULL)}}{Initializes an object of the NGramSet class.}
  \item{\code{word()}}{Creates word tokens.}
  \item{\code{sentence()}}{Creates sentence tokens.}
  \item{\code{char()}}{Creates character tokens.}
  \item{\code{get()}}{Returns the tokens.}
  \item{\code{getText()}}{Returns the original text.}
 }
}

\seealso{
Other NGram Classes: \code{\link{NGram}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
