% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelize.R
\name{parallelize}
\alias{parallelize}
\title{parallelize}
\usage{
parallelize(task, ...)
}
\arguments{
\item{task}{- the function to parallelize}

\item{...}{- other parameters required by task}
}
\value{
r - return from task
}
\description{
This generic function used to parallelize a task (when possible)
}
\author{
John James
}
