% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_top_terms.R
\name{plot_top_terms}
\alias{plot_top_terms}
\title{Visualize Topic-Word Probabilities}
\usage{
plot_top_terms(top_terms, facet_args = list(scales = "free_y"), ...)
}
\arguments{
\item{top_terms}{A \code{data.table} returned by \code{\link[=get_top_terms]{get_top_terms()}}, containing the columns
\code{rank}, \code{topic}, \code{term}, and \code{probability}.}

\item{facet_args}{A named list of additional arguments passed to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}.
By default, \code{scales = "free_y"} is used.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:geom_bar]{ggplot2::geom_col()}}.}
}
\value{
A \code{ggplot} object showing one horizontal bar chart per topic. Each bar reflects
a term’s estimated contribution to that topic.
}
\description{
This function plots the top-ranked terms for each topic based on the long-format output from
\code{\link[=get_top_terms]{get_top_terms()}}. It produces a faceted horizontal bar chart where each facet represents
one topic, and the height of each bar reflects the term's estimated topic-word probability
( \eqn{\phi} ).
}
\details{
The function uses \strong{\link[ggplot2:ggplot2-package]{ggplot2::ggplot2}} to create a faceted bar chart showing the most relevant terms
for each topic. Terms are ranked and sorted within each facet by their associated probability
in the topic-word distribution matrix (\eqn{\phi}). Internally, the term ordering is adjusted using
\strong{\link[tidytext:tidytext-package]{tidytext::tidytext}} \code{reorder_within()} to ensure per-topic ranking. This function is typically
used in conjunction with \code{\link[=get_top_terms]{get_top_terms()}} with \code{format = "long"}, which prepares the data for
suitable visualization.
}
\seealso{
\code{\link[=get_top_terms]{get_top_terms()}}, \code{\link[=plot_dtw]{plot_dtw()}}
}
