% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POS.R
\docType{class}
\name{POS}
\alias{POS}
\title{POS}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
pos <- POS$new()
}
\arguments{
\item{x}{The source Document object.}
}
\value{
POS object, containing the tokens for a single Document object.
}
\description{
\code{POS} Creates part-of-speech tagged documents.

@param x The principal object
 @param key Character string or a vector of strings containing a key or
 keys for metadata key/value pair(s).
 @param value Character or numeric, or a vector thereof, containing a value
 or values for metadata key/value pair(s).
 @param format Character string used in the metadata method. Indicates
 whether to return the metadata as  list of lists, or a data.frame.This method
 returns metadata in a list if no key/value pair is provided. The list
 may contain other lists or data frames associated with the type of
 metadata. This parameter specifies whether to return the metadata
 as a list of lists, or a list of data frames. Valid values are:
 c('list', 'data.frame')
}
\details{
Class containing the methods for creating and reporting part-of-speech tags
for a Document object.
}
\section{Core Methods}{

 \itemize{
  \item{\code{new(x)}}{Initializes an object of the POS class.}
  \item{\code{content}}{Active binding used to set and retrieve POS content. POS
  content may be changed via assignment. Referencing this method retrieves the current
  POS content.}
 }
}

\seealso{
Other POS Classes: \code{\link{POSSet}},
  \code{\link{Tagger}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
