% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POSFactory.R
\docType{class}
\name{POSFactory}
\alias{POSFactory}
\title{POSFactory}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
POSFactory$new(x)$execute()
}
\arguments{
\item{x}{Object of to be processed. This can be an object of the Corpus, Document, character classes, or a list of character vectors.}
}
\value{
\code{POSCollection} object.
}
\description{
\code{POSFactory} Class responsible for creating a POSCollection object from Corpus objects.
}
\details{
This class is a wrapper for the  \code{\link[openNLP]{Maxent_POS_Tag_Annotator}}
function in the openNLP package.

Sources:
\url{https://cran.r-project.org/web/packages/openNLP/openNLP.pdf}
}
\section{Methods}{

 \describe{
  \item{\code{new(...)}}{Instantiates the class object.}
  \item{\code{execute()}}{Executes the process on the object.}
}
}

\section{DataStudio Family of Classes Overview}{

The DataStudio family of classes include the DataStudio class, which is
responsible for executing one or more data processing commands on a
Corpus or Document resource, the core class group that performs the
data processing functions on the objects, and the command class
group which instructs the DataStudio class to perform the
data processing function on the Corpus or Document object.
}

\section{DataStudio Design}{

This DataStudio family of classes is an implementation of the Strategy
and Command design patterns, as described in the book "Design Patterns:
Elements of Reusable Object-Oriented Software" by Erich Gamma, Richard
Helm, Ralph Johnson and John Vlissides (hence Gang of Four). The
Strategy pattern defines and encapsulates the family of data processing
classes and allows text cleaning algorithms to be selected and
exhanged at runtime. THe Command pattern defines commands that
can be executed in batch mode or at a later time.
}

\seealso{
Other DataStudio Classes: \code{\link{POSFactory0}},
  \code{\link{TermFreqFactory0}},
  \code{\link{TermFreqFactoryDfm}},
  \code{\link{TermFreqFactoryDtm}},
  \code{\link{TermFreqFactoryTdm}}, \code{\link{Tokenize}}

Other POS Classes: \code{\link{POSFactory0}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
