% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POSFactory0.R
\docType{class}
\name{POSFactory0}
\alias{POSFactory0}
\title{POSFactory0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
POSFactory0
}
\description{
\code{POSFactory0}  Abstract class. Defines interface for POSFactory class.
}
\seealso{
Other DataStudio Classes: \code{\link{POSFactory}},
  \code{\link{TermFreqFactory0}},
  \code{\link{TermFreqFactoryDfm}},
  \code{\link{TermFreqFactoryDtm}},
  \code{\link{TermFreqFactoryTdm}}, \code{\link{Tokenize}}

Other POS Classes: \code{\link{POSFactory}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
