% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Primative0.R
\docType{class}
\name{Primative0}
\alias{Primative0}
\title{Primative0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Primative0
}
\description{
\code{Primative0} Abstract class for all primative classes.
}
\details{
Abstract class which defines the methods common to all primatives
(and by inheritence, composite) objects.
}
\section{Primative0 methods}{

 \itemize{
  \item{\code{new()}}{Method not implemented for this component class.}
  \item{\code{summary()}}{Prints a summary of a Document family class object.}
 }
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
