% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Primitive0.R
\docType{class}
\name{Primitive0}
\alias{Primitive0}
\title{Primitive0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Primitive0
}
\description{
\code{Primitive0} Abstract class for all primative classes.
}
\details{
Abstract class which defines the methods common to all primatives
(and by inheritence, composite) objects.
}
\section{Primitive0 methods}{

 \itemize{
  \item{\code{new()}}{Method not implemented for this component class.}
  \item{\code{summary()}}{Prints a summary of a Document family class object.}
 }
}

\seealso{
Other Document Classes: \code{\link{Corpus}},
  \code{\link{Document}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
