% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_selection_summary.R
\name{print.nlp_k_selection_summary}
\alias{print.nlp_k_selection_summary}
\title{Print K-selection summary}
\usage{
\method{print}{nlp_k_selection_summary}(x, ...)
}
\arguments{
\item{x}{An \code{nlp_k_selection_summary} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print K-selection summary
}
