% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_diagnostics.R
\name{print.nlp_topic_stability}
\alias{print.nlp_topic_stability}
\title{Print Topic Stability Results}
\usage{
\method{print}{nlp_topic_stability}(x, ...)
}
\arguments{
\item{x}{An \code{nlp_topic_stability} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Topic Stability Results
}
