% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.R, R/Read0.R
\docType{class}
\name{Read0}
\alias{Read0}
\alias{Read0}
\title{Read0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Read0

Read0
}
\arguments{
\item{document}{An object of class Document}
}
\description{
\code{Read0} Class as in interface class to the ReadText, ReadCsv, ReadRdata
             subclasses

\code{Read0} Class as in abstract class to the ReadBin, ReadText, ReadCsv, ReadRdata read strategies
}
\details{
This class is an interface to the Read* sub classes.  It defines the core
variables and method(s) inherited by the Read* subclasses.

This class is an abstract class defining the methods to be implemented by
the concrete strategy classes: ReadText, ReadBin, ReadCsv, and ReadRdata.
}
\section{Fields}{

\describe{
\item{\code{directory}}{A character string containing the file directory}

\item{\code{fileName}}{A character string containing the name of the file to read}
}}

\section{Fields}{

}

\section{Methods}{

\describe{
 \item{\code{readData(directory, fileName)}}{Overwritten by subclasses}
}


\describe{
 \item{\code{new()}}{Returns an object of the selected concrete read strategy class. This method is overwritten by the concrete strategy classes}
 \item{\code{readData(document)}}{Returns a list containing the meta data and content}
}
}

\section{Parameters}{

}

\author{
John James, \email{j2sdatalab@gmail.com}

John James, \email{j2sdatalab@gmail.com}
}
\keyword{datasets}
