% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadBin.R
\docType{class}
\name{ReadBin}
\alias{ReadBin}
\title{ReadBin}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ReadBin
}
\description{
\code{ReadBin} Class contains the data and methods for reading files in binary format.
}
\details{
This class inherits from the Read0 inteface class and provides the data
and methods for reading files in binary format.
}
\section{Methods}{

\describe{
 \item{\code{readData(directory, fileName)}}{Reads data from the designated  directory and file name}
}
}

\seealso{
Other Document i/o classes: \code{\link{ReadCsv}},
  \code{\link{ReadRdata}}, \code{\link{ReadText}}
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
\keyword{datasets}
