% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadCorpus.R
\docType{class}
\name{ReadCorpus}
\alias{ReadCorpus}
\title{ReadCorpus}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ReadCorpus
}
\arguments{
\item{x}{Corpus or CVSet Object}

\item{corpusStudio}{CorpusStudio object}

\item{path}{Character string the directory or file path
the object.}
}
\value{
Token object
}
\description{
\code{ReadCorpus} Encapsulates the command to read a Corpus or its texts.
}
\seealso{
Other FileStudio Command Classes: \code{\link{CopyFileSet}},
  \code{\link{MoveFileSet}}, \code{\link{RepairFileSet}},
  \code{\link{SampleCorpus}}, \code{\link{Source}},
  \code{\link{TokenizeCorpus}}, \code{\link{WriteCorpus}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
