% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_tokens.R
\name{remove_tokens}
\alias{remove_tokens}
\title{Fast Corpus Tokenization}
\usage{
remove_tokens(x, ncores, ...)
}
\arguments{
\item{x}{A \code{\link[quanteda]{corpus}} as built by \code{create_corpus}.}

\item{ncores}{The number of \code{\link[future]{multisession}} workers to be allocated for the tokenization.}

\item{...}{Additional arguments passed to \code{\link[quanteda]{tokens}}.}
}
\value{
A \code{\link[quanteda]{tokens}} object.
}
\description{
Tokenize a \code{\link[quanteda]{corpus}} in parallel.
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}
}
