% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeNonPrintable.R
\name{removeNonPrintable}
\alias{removeNonPrintable}
\title{Remove NonPrintable}
\usage{
removeNonPrintable(path, codes = nonPrintables)
}
\arguments{
\item{path}{Character string including path to file}

\item{codes}{Numeric indicating the decimal codes for the characters
to be removed.}
}
\value{
content Character vector containing content with non-printable
characters removed.
}
\description{
\code{removeNonPrintable} Removes select non-printable characters from file
using binary facilities. Defaults to removing all but line and form feed
characters.
}
