% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Repair.R
\docType{class}
\name{Repair}
\alias{Repair}
\title{Repair}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Repair
}
\arguments{
\item{input}{Character string containing the relative file or directory path
for the file(s) to be repaired.}

\item{output}{Character string containing the relative file or directory path
to which the repaired file is stored.}

\item{codes}{The ASCII codes for non printable characters to be replaced.}
}
\description{
\code{Repair}  Class responsible for performing repairs on files.
}
\details{
Class replaces select control characters with spaces at the binary level.
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
