% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepairApp.R
\docType{class}
\name{RepairApp}
\alias{RepairApp}
\title{RepairApp}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RepairApp
}
\arguments{
\item{x}{Receiver object, e.g. the FileStudio object.}

\item{input}{Character string containing the relative file or directory path
for the file(s) to be repaired.}

\item{output}{Character string containing the relative file or directory path
to which the repaired file is stored.}

\item{codes}{= ASCII codes for the control characters to be replaced.}
}
\description{
\code{RepairApp} Encapsulates the command to execute an object of the Repair class.
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
