% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepairFile.R
\docType{class}
\name{RepairFile}
\alias{RepairFile}
\title{RepairFile}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RepairFile
}
\arguments{
\item{path}{Character string indicating the path to a file to be repaired.}
}
\value{
content Character vector with repaired content.
}
\description{
\code{RepairFile} Repairs ASCII encoding in text files
}
\details{
Class responsible for performing repairs on text files at binary level. Control
characters are converted to ASCII spaces either individually, by calling the
method associated with a specific control character, or all control
characters can be converted to an ASCII space at once by calling the
'ctrl' method.
}
\examples{
txtFile <- "./foo.txt"
txtContent <- RepairFile$new(txtFile)$null()  # Converts NULL characters to spaces
txtContent <- RepairFile$new(txtFile)$sub()  # Converts SUB characters to spaces
txtContent <- RepairFile$new(txtFile)$ctrl()  # Converts all control characters to spaces

}
\seealso{
Other TextStudio Classes: \code{\link{LowerCaseCmd}},
  \code{\link{LowerCase}}, \code{\link{RemoveEmailCmd}},
  \code{\link{RemoveEmail}},
  \code{\link{RemoveHyphensCmd}},
  \code{\link{RemoveHyphens}},
  \code{\link{RemoveNumbersCmd}},
  \code{\link{RemoveNumbers}},
  \code{\link{RemovePunctCmd}}, \code{\link{RemovePunct}},
  \code{\link{RemoveSymbolsCmd}},
  \code{\link{RemoveSymbols}},
  \code{\link{RemoveTwitterCmd}},
  \code{\link{RemoveTwitter}}, \code{\link{RemoveURLCmd}},
  \code{\link{RemoveURL}},
  \code{\link{RemoveWhiteSpaceCmd}},
  \code{\link{RemoveWhiteSpace}},
  \code{\link{ReplaceAbbreviationsCmd}},
  \code{\link{ReplaceAbbreviations}},
  \code{\link{ReplaceBacktickCmd}},
  \code{\link{ReplaceBacktick}},
  \code{\link{ReplaceContractionsCmd}},
  \code{\link{ReplaceContractions}},
  \code{\link{ReplaceCurlyQuotesCmd}},
  \code{\link{ReplaceCurlyQuotes}},
  \code{\link{ReplaceEmojiCmd}},
  \code{\link{ReplaceEmoji}},
  \code{\link{ReplaceEmoticonCmd}},
  \code{\link{ReplaceEmoticon}},
  \code{\link{ReplaceHTMLCmd}}, \code{\link{ReplaceHTML}},
  \code{\link{ReplaceInternetSlangCmd}},
  \code{\link{ReplaceInternetSlang}},
  \code{\link{ReplaceKernCmd}}, \code{\link{ReplaceKern}},
  \code{\link{ReplaceNamesCmd}},
  \code{\link{ReplaceNames}},
  \code{\link{ReplaceNonAsciiCmd}},
  \code{\link{ReplaceNonAscii}},
  \code{\link{ReplaceNumbersCmd}},
  \code{\link{ReplaceNumbers}},
  \code{\link{ReplaceOrdinalCmd}},
  \code{\link{ReplaceOrdinal}},
  \code{\link{ReplaceSymbolCmd}},
  \code{\link{ReplaceSymbol}},
  \code{\link{ReplaceTokensCmd}},
  \code{\link{ReplaceTokens}},
  \code{\link{ReplaceWordElongationCmd}},
  \code{\link{ReplaceWordElongation}},
  \code{\link{SplitCorpusCmd}}, \code{\link{SplitCorpus}},
  \code{\link{SplitDocument}}, \code{\link{StripTextCmd}},
  \code{\link{Tokenize}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
