% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepairFileSet.R
\docType{class}
\name{RepairFileSet}
\alias{RepairFileSet}
\title{RepairFileSet}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RepairFileSet
}
\arguments{
\item{x}{Receiver object, e.g. the FileStudio object.}

\item{path}{Character string containing the relative file or directory path
to which the repaired file will be stored.}

\item{target}{FileSet object}
}
\description{
\code{RepairFileSet} Encapsulates the command to execute to repair a FileSet object.
}
\seealso{
Other FileStudio Command Classes: \code{\link{CopyFileSet}},
  \code{\link{MoveFileSet}}, \code{\link{ReadCorpus}},
  \code{\link{SampleCorpus}}, \code{\link{Source}},
  \code{\link{WriteCorpus}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\concept{FileStudio Command Classes}
\keyword{datasets}
