% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepairFileSetCmd.R
\docType{class}
\name{RepairFileSetCmd}
\alias{RepairFileSetCmd}
\title{RepairFileSetCmd}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RepairFileSetCmd
}
\arguments{
\item{x}{Receiver object, e.g. the FileStudio object.}

\item{path}{Character string containing the relative file or directory path
to which the repaired file will be stored.}

\item{target}{FileSet object}
}
\description{
\code{RepairFileSetCmd} Encapsulates the command to execute to repair a FileSet object.
}
\seealso{
Other FileStudio Command Classes: \code{\link{BuildFileSetCmd}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
