% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplaceSymbolApp.R
\docType{class}
\name{ReplaceSymbolApp}
\alias{ReplaceSymbolApp}
\title{ReplaceSymbolApp}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ReplaceSymbolApp$new(x, dollar = TRUE)$execute()
}
\arguments{
\item{x}{Object of to be processed. This can be an object of the Corpus, Document, character classes, or a list of character vectors.}

\item{dollar}{logical. If TRUE replaces dollar sign (\$) with "dollar".}

\item{percent}{logical. If TRUE replaces percent sign (\%) with "percent".}

\item{pound}{logical. If TRUE replaces pound sign (\#) with "number".}

\item{at}{logical. If TRUE replaces at sign (\@) with "at".}

\item{and}{logical. If TRUE replaces and sign (\&) with "and".}

\item{with}{logical. If TRUE replaces with sign (w/) with "with"}
}
\value{
\code{ReplaceSymbolApp} Returns a vector with symbols replaced.
}
\description{
\code{ReplaceSymbolApp}  Replace Symbols With Word Equivalents
}
\details{
A wrapper for \code{\link[textclean]{replace_symbol}} This function replaces symbols
with word equivalents (e.g., @ becomes "at".
Source \url{https://cran.r-project.org/web/packages/textclean/textclean.pdf}
}
\section{Methods}{

 \describe{
  \item{\code{new(...)}}{Instantiates the class object.}
  \item{\code{execute()}}{Executes the process on the object.}
}
}

\section{TextStudio Family of Classes Overview}{

The TextStudio family of classes include four types of classes. The manager classes which
are responsible for executing one or more core class objects through the implementation
of a command class object, supported by support class objects.
}

\section{DataStudio Design}{

This DataStudio family of classes is an implementation of the Strategy
design pattern, as described in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). The Strategy pattern defines
the family of text cleaning classes, encapsulates class and allows
text cleaning algorithms to be selected and exchagned at runtime.
}

\examples{

}
\seealso{
Other TextStudio Classes: \code{\link{AddCommaSpaceApp}},
  \code{\link{AddCommaSpace}},
  \code{\link{RemoveEmailApp}}, \code{\link{RemoveEmail}},
  \code{\link{RemoveHyphensApp}},
  \code{\link{RemoveHyphens}},
  \code{\link{RemoveNumbersApp}},
  \code{\link{RemoveNumbers}},
  \code{\link{RemovePunctApp}}, \code{\link{RemovePunct}},
  \code{\link{RemoveSymbolsApp}},
  \code{\link{RemoveSymbols}},
  \code{\link{RemoveTwitterApp}},
  \code{\link{RemoveTwitter}}, \code{\link{RemoveURLApp}},
  \code{\link{RemoveURL}},
  \code{\link{RemoveWhiteSpaceApp}},
  \code{\link{RemoveWhiteSpace}},
  \code{\link{ReplaceAbbreviationsApp}},
  \code{\link{ReplaceAbbreviations}},
  \code{\link{ReplaceBacktickApp}},
  \code{\link{ReplaceBacktick}},
  \code{\link{ReplaceContractionsApp}},
  \code{\link{ReplaceContractions}},
  \code{\link{ReplaceCurlyQuotesApp}},
  \code{\link{ReplaceCurlyQuotes}},
  \code{\link{ReplaceEmojiApp}},
  \code{\link{ReplaceEmoji}},
  \code{\link{ReplaceEmoticonApp}},
  \code{\link{ReplaceEmoticon}},
  \code{\link{ReplaceHTMLApp}}, \code{\link{ReplaceHTML}},
  \code{\link{ReplaceInternetSlangApp}},
  \code{\link{ReplaceInternetSlang}},
  \code{\link{ReplaceKernApp}}, \code{\link{ReplaceKern}},
  \code{\link{ReplaceNonAsciiApp}},
  \code{\link{ReplaceNonAscii}},
  \code{\link{ReplaceNumbersApp}},
  \code{\link{ReplaceNumbers}},
  \code{\link{ReplaceOrdinalApp}},
  \code{\link{ReplaceOrdinal}},
  \code{\link{ReplaceSymbol}},
  \code{\link{ReplaceTokensApp}},
  \code{\link{ReplaceTokens}},
  \code{\link{ReplaceWordElongationApp}},
  \code{\link{ReplaceWordElongation}},
  \code{\link{StripTextApp}}, \code{\link{StripText}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
