% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplaceTokens.R, R/SplitText.R
\docType{class}
\name{ReplaceTokens}
\alias{ReplaceTokens}
\alias{ReplaceTokens}
\title{ReplaceTokens}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ReplaceTokens$new(x, pattern, replacement, leadspace = FALSE)$execute()

ReplaceTokens$new(x, pattern, replacement, leadspace = FALSE)$execute()
}
\arguments{
\item{x}{Object of to be processed. This can be an object of the Corpus, Document, character classes, or a list of character vectors.}

\item{tokens}{Character string(s) to be matched in the given character vector.}

\item{replacement}{Character string equal in length to pattern or of length
one which are  a replacement for matched pattern.}

\item{leadspace}{logical.  If \code{TRUE} inserts a leading space in the
replacements.}

\item{trailspace}{logical.  If \code{TRUE} inserts a trailing space in the
replacements.}

\item{fixed}{logical. If \code{TRUE}, pattern is a string to be matched as is.
Overrides all conflicting arguments.}

\item{trim}{logical.  If \code{TRUE} leading and trailing white spaces are
removed and multiple white spaces are reduced to a single white space.}

\item{order.pattern}{logical.  If \code{TRUE} and \code{fixed = TRUE}, the
\code{pattern} string is sorted by number of characters to prevent substrings
replacing meta strings (e.g., \code{pattern = c("the", "then")} resorts to
search for "then" first).}

\item{x}{Object of to be processed. This can be an object of the Corpus, Document, character classes, or a list of character vectors.}

\item{tokens}{Character string(s) to be matched in the given character vector.}

\item{replacement}{Character string equal in length to pattern or of length
one which are  a replacement for matched pattern.}

\item{leadspace}{logical.  If \code{TRUE} inserts a leading space in the
replacements.}

\item{trailspace}{logical.  If \code{TRUE} inserts a trailing space in the
replacements.}

\item{fixed}{logical. If \code{TRUE}, pattern is a string to be matched as is.
Overrides all conflicting arguments.}

\item{trim}{logical.  If \code{TRUE} leading and trailing white spaces are
removed and multiple white spaces are reduced to a single white space.}

\item{order.pattern}{logical.  If \code{TRUE} and \code{fixed = TRUE}, the
\code{pattern} string is sorted by number of characters to prevent substrings
replacing meta strings (e.g., \code{pattern = c("the", "then")} resorts to
search for "then" first).}
}
\value{
\code{ReplaceTokens} - Returns a vector with the pattern replaced.

\code{ReplaceTokens} - Returns a vector with the pattern replaced.
}
\description{
\code{ReplaceTokens}  Replace a vector of search tokens

\code{ReplaceTokens}  Replace a vector of search tokens
}
\details{
A wrapper for \code{\link[textclean]{mgsub}} that takes a vector
of search tokens and a vector or single value of replacements.
Source \url{https://cran.r-project.org/web/packages/textclean/textclean.pdf}

A wrapper for \code{\link[textclean]{mgsub}} that takes a vector
of search tokens and a vector or single value of replacements.
Source \url{https://cran.r-project.org/web/packages/textclean/textclean.pdf}
}
\section{Methods}{

 \describe{
  \item{\code{new(...)}}{Instantiates the class object.}
  \item{\code{execute()}}{Executes the process on the object.}
}


 \describe{
  \item{\code{new(...)}}{Instantiates the class object.}
  \item{\code{execute()}}{Executes the process on the object.}
}
}

\section{TextStudio Family of Classes Overview}{

The TextStudio family of classes include the TextStudio class, which is
responsible for executing one or more text processing commands on a
text resource, the core class group that performs the processing functions
on the text, and the command class group which instructs the TextStudio
class to perform an a text processing function on a text.


The TextStudio family of classes include the TextStudio class, which is
responsible for executing one or more text processing commands on a
text resource, the core class group that performs the processing functions
on the text, and the command class group which instructs the TextStudio
class to perform an a text processing function on a text.
}

\section{TextStudio Design}{

This TextStudio family of classes is an implementation of the Strategy
design pattern, as described in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). The Strategy pattern defines
the family of text cleaning classes, encapsulates class and allows
text cleaning algorithms to be selected and exchagned at runtime.


This TextStudio family of classes is an implementation of the Strategy
design pattern, as described in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). The Strategy pattern defines
the family of text cleaning classes, encapsulates class and allows
text cleaning algorithms to be selected and exchagned at runtime.
}

\seealso{
Other TextStudio Classes: \code{\link{LowerCaseCmd}},
  \code{\link{LowerCase}}, \code{\link{RemoveEmailCmd}},
  \code{\link{RemoveEmail}},
  \code{\link{RemoveHyphensCmd}},
  \code{\link{RemoveHyphens}},
  \code{\link{RemoveNumbersCmd}},
  \code{\link{RemoveNumbers}},
  \code{\link{RemovePunctCmd}}, \code{\link{RemovePunct}},
  \code{\link{RemoveSymbolsCmd}},
  \code{\link{RemoveSymbols}},
  \code{\link{RemoveTwitterCmd}},
  \code{\link{RemoveTwitter}}, \code{\link{RemoveURLCmd}},
  \code{\link{RemoveURL}},
  \code{\link{RemoveWhiteSpaceCmd}},
  \code{\link{RemoveWhiteSpace}}, \code{\link{RepairFile}},
  \code{\link{ReplaceAbbreviationsCmd}},
  \code{\link{ReplaceAbbreviations}},
  \code{\link{ReplaceBacktickCmd}},
  \code{\link{ReplaceBacktick}},
  \code{\link{ReplaceContractionsCmd}},
  \code{\link{ReplaceContractions}},
  \code{\link{ReplaceCurlyQuotesCmd}},
  \code{\link{ReplaceCurlyQuotes}},
  \code{\link{ReplaceEmojiCmd}},
  \code{\link{ReplaceEmoji}},
  \code{\link{ReplaceEmoticonCmd}},
  \code{\link{ReplaceEmoticon}},
  \code{\link{ReplaceHTMLCmd}}, \code{\link{ReplaceHTML}},
  \code{\link{ReplaceInternetSlangCmd}},
  \code{\link{ReplaceInternetSlang}},
  \code{\link{ReplaceKernCmd}}, \code{\link{ReplaceKern}},
  \code{\link{ReplaceNamesCmd}},
  \code{\link{ReplaceNames}},
  \code{\link{ReplaceNonAsciiCmd}},
  \code{\link{ReplaceNonAscii}},
  \code{\link{ReplaceNumbersCmd}},
  \code{\link{ReplaceNumbers}},
  \code{\link{ReplaceOrdinalCmd}},
  \code{\link{ReplaceOrdinal}},
  \code{\link{ReplaceSymbolCmd}},
  \code{\link{ReplaceSymbol}},
  \code{\link{ReplaceTokensCmd}},
  \code{\link{ReplaceWordElongationCmd}},
  \code{\link{ReplaceWordElongation}},
  \code{\link{SplitCorpusCmd}}, \code{\link{SplitCorpus}},
  \code{\link{SplitDocumentCmd}},
  \code{\link{SplitDocument}}, \code{\link{StripTextCmd}},
  \code{\link{Tokenize}}

Other TextStudio Classes: \code{\link{LowerCaseCmd}},
  \code{\link{LowerCase}}, \code{\link{RemoveEmailCmd}},
  \code{\link{RemoveEmail}},
  \code{\link{RemoveHyphensCmd}},
  \code{\link{RemoveHyphens}},
  \code{\link{RemoveNumbersCmd}},
  \code{\link{RemoveNumbers}},
  \code{\link{RemovePunctCmd}}, \code{\link{RemovePunct}},
  \code{\link{RemoveSymbolsCmd}},
  \code{\link{RemoveSymbols}},
  \code{\link{RemoveTwitterCmd}},
  \code{\link{RemoveTwitter}}, \code{\link{RemoveURLCmd}},
  \code{\link{RemoveURL}},
  \code{\link{RemoveWhiteSpaceCmd}},
  \code{\link{RemoveWhiteSpace}}, \code{\link{RepairFile}},
  \code{\link{ReplaceAbbreviationsCmd}},
  \code{\link{ReplaceAbbreviations}},
  \code{\link{ReplaceBacktickCmd}},
  \code{\link{ReplaceBacktick}},
  \code{\link{ReplaceContractionsCmd}},
  \code{\link{ReplaceContractions}},
  \code{\link{ReplaceCurlyQuotesCmd}},
  \code{\link{ReplaceCurlyQuotes}},
  \code{\link{ReplaceEmojiCmd}},
  \code{\link{ReplaceEmoji}},
  \code{\link{ReplaceEmoticonCmd}},
  \code{\link{ReplaceEmoticon}},
  \code{\link{ReplaceHTMLCmd}}, \code{\link{ReplaceHTML}},
  \code{\link{ReplaceInternetSlangCmd}},
  \code{\link{ReplaceInternetSlang}},
  \code{\link{ReplaceKernCmd}}, \code{\link{ReplaceKern}},
  \code{\link{ReplaceNamesCmd}},
  \code{\link{ReplaceNames}},
  \code{\link{ReplaceNonAsciiCmd}},
  \code{\link{ReplaceNonAscii}},
  \code{\link{ReplaceNumbersCmd}},
  \code{\link{ReplaceNumbers}},
  \code{\link{ReplaceOrdinalCmd}},
  \code{\link{ReplaceOrdinal}},
  \code{\link{ReplaceSymbolCmd}},
  \code{\link{ReplaceSymbol}},
  \code{\link{ReplaceTokensCmd}},
  \code{\link{ReplaceWordElongationCmd}},
  \code{\link{ReplaceWordElongation}},
  \code{\link{SplitCorpusCmd}}, \code{\link{SplitCorpus}},
  \code{\link{SplitDocumentCmd}},
  \code{\link{SplitDocument}}, \code{\link{StripTextCmd}},
  \code{\link{Tokenize}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}

John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
