% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplaceTokensApp.R
\docType{class}
\name{ReplaceTokensApp}
\alias{ReplaceTokensApp}
\title{ReplaceTokensApp}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ReplaceTokensApp$new(x, tokens, replacement, leadspace, trailspace, fixed, trim, orderPattern)$execute()
}
\arguments{
\item{x}{Object of to be processed. This can be an object of the Corpus, Document, character classes, or a list of character vectors.}

\item{tokens}{Character string of tokens to be matched in the
given character vector.}

\item{replacement}{Character string equal in length to tokens containing
the long forms of the tokens.}

\item{leadspace}{logical.  If \code{TRUE} inserts a leading space in the
replacements.}

\item{trailspace}{logical.  If \code{TRUE} inserts a trailing space in the
replacements.}

\item{fixed}{logical. If \code{TRUE}, tokens is a string to be matched as is.
Overrides all conflicting arguments.}

\item{trim}{logical.  If \code{TRUE} leading and trailing white spaces are
removed and multiple white spaces are reduced to a single white space.}

\item{orderPattern}{logical.  If \code{TRUE} and \code{fixed = TRUE}, the
\code{tokens} string is sorted by number of characters to prevent substrings
replacing meta strings (e.g., \code{tokens = c("the", "then")} resorts to
search for "then" first).}

\item{\dots}{ignored.}
}
\value{
\code{ReplaceTokensApp} Returns a vector with tokens replaced.
}
\description{
\code{ReplaceTokensApp}  Replace tokens.
}
\details{
A wrapper for \code{\link[textclean]{mgsub}}, A wrapper for gsub that takes
a vector of search terms and a vector or single value of replacements.
Source \url{https://cran.r-project.org/web/packages/textclean/textclean.pdf}
}
\section{Methods}{

 \describe{
  \item{\code{new(...)}}{Instantiates the class object.}
  \item{\code{execute()}}{Executes the process on the object.}
}
}

\section{TextStudio Family of Classes Overview}{

The TextStudio family of classes include four types of classes. The manager classes which
are responsible for executing one or more core class objects through the implementation
of a command class object, supported by support class objects.
}

\section{DataStudio Design}{

This DataStudio family of classes is an implementation of the Strategy
design pattern, as described in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). The Strategy pattern defines
the family of text cleaning classes, encapsulates class and allows
text cleaning algorithms to be selected and exchagned at runtime.
}

\examples{

}
\seealso{
Other TextStudio Classes: \code{\link{AddCommaSpaceApp}},
  \code{\link{AddCommaSpace}}, \code{\link{GSubApp}},
  \code{\link{GSub}}, \code{\link{NormalizeEncodingApp}},
  \code{\link{NormalizeEncoding}},
  \code{\link{RemoveEmailApp}}, \code{\link{RemoveEmail}},
  \code{\link{RemoveNumbersApp}},
  \code{\link{RemoveNumbers}},
  \code{\link{RemovePunctApp}}, \code{\link{RemovePunct}},
  \code{\link{RemoveTwitterApp}},
  \code{\link{RemoveTwitter}}, \code{\link{RemoveURLApp}},
  \code{\link{RemoveURL}},
  \code{\link{RemoveWhiteSpaceApp}},
  \code{\link{RemoveWhiteSpace}},
  \code{\link{ReplaceAbbreviationsApp}},
  \code{\link{ReplaceAbbreviations}},
  \code{\link{ReplaceBacktickApp}},
  \code{\link{ReplaceBacktick}},
  \code{\link{ReplaceContractionsApp}},
  \code{\link{ReplaceContractions}},
  \code{\link{ReplaceCurlyQuotesApp}},
  \code{\link{ReplaceCurlyQuotes}},
  \code{\link{ReplaceEmojiApp}},
  \code{\link{ReplaceEmoji}},
  \code{\link{ReplaceEmoticonApp}},
  \code{\link{ReplaceEmoticon}},
  \code{\link{ReplaceHTMLApp}}, \code{\link{ReplaceHTML}},
  \code{\link{ReplaceInternetSlangApp}},
  \code{\link{ReplaceInternetSlang}},
  \code{\link{ReplaceKernApp}}, \code{\link{ReplaceKern}},
  \code{\link{ReplaceNonAsciiApp}},
  \code{\link{ReplaceNonAscii}},
  \code{\link{ReplaceNumbersApp}},
  \code{\link{ReplaceNumbers}},
  \code{\link{ReplaceOrdinalApp}},
  \code{\link{ReplaceOrdinal}},
  \code{\link{ReplaceSymbolApp}},
  \code{\link{ReplaceSymbol}}, \code{\link{ReplaceTokens}},
  \code{\link{ReplaceWhiteApp}},
  \code{\link{ReplaceWhite}},
  \code{\link{ReplaceWordElongationApp}},
  \code{\link{ReplaceWordElongation}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
