% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplaceWordElongationApp.R
\docType{class}
\name{ReplaceWordElongationApp}
\alias{ReplaceWordElongationApp}
\title{ReplaceWordElongationApp}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ReplaceWordElongationApp$new(x, impartMeaning = FALSE)$execute()
}
\arguments{
\item{x}{Object of to be processed. This can be an object of the Corpus, Document, character classes, or a list of character vectors.}

\item{impartMeaning}{logical. If TRUE, known elongation semantics are used as replacements
(see textclean:::meaning_elongations for known elongation semantics and replacements).}
}
\value{
\code{ReplaceWordElongationApp} Returns a vector with word elongations replaced.
}
\description{
\code{ReplaceWordElongationApp}  Replace Word Elongations.
}
\details{
A wrapper for \code{\link[textclean]{replace_word_elongation}} In informal writing
people may use a form of text embellishment to emphasize or alter word meanings
called elongation (a.k.a. "word lengthening"). For example, the use of "Whyyyyy" conveys
frustration. Other times the usage may be to be more sexy (e.g., "Heyyyy there"). Other times it
may be used for emphasis (e.g., "This is so gooood"). This function uses an augmented form of
Armstrong & Fogarty’s (2007) algorithm. The algorithm first attempts to replace the elongation
with known semantic replacements (optional; default is FALSE). After this the algorithm locates all
places were the same letter (case insensitive) appears 3 times consecutively. These elements are then
further processed. The matches are replaces via fgsub by first taking the elongation to it’s canonical
form (drop all > 1 consecutive letters to a single letter) and then replacing with the most common
word used in 2008 in Google’s ngram data set that takes the canonical form. If the canonical form
is not found in the Google data set then the canonical form is used as the replacement.
Source \url{https://cran.r-project.org/web/packages/textclean/textclean.pdf}
}
\section{Methods}{

 \describe{
  \item{\code{new(...)}}{Instantiates the class object.}
  \item{\code{execute()}}{Executes the process on the object.}
}
}

\section{TextStudio Family of Classes Overview}{

The TextStudio family of classes include four types of classes. The manager classes which
are responsible for executing one or more core class objects through the implementation
of a command class object, supported by support class objects.
}

\section{DataStudio Design}{

This DataStudio family of classes is an implementation of the Strategy
design pattern, as described in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). The Strategy pattern defines
the family of text cleaning classes, encapsulates class and allows
text cleaning algorithms to be selected and exchagned at runtime.
}

\examples{

}
\seealso{
Other TextStudio Classes: \code{\link{AddCommaSpaceApp}},
  \code{\link{AddCommaSpace}}, \code{\link{GSubApp}},
  \code{\link{GSub}}, \code{\link{NormalizeEncodingApp}},
  \code{\link{NormalizeEncoding}},
  \code{\link{RemoveEmailApp}}, \code{\link{RemoveEmail}},
  \code{\link{RemoveTwitterApp}},
  \code{\link{RemoveTwitter}}, \code{\link{RemoveURLApp}},
  \code{\link{RemoveURL}},
  \code{\link{RemoveWhiteSpaceApp}},
  \code{\link{RemoveWhiteSpace}},
  \code{\link{ReplaceAbbreviationsApp}},
  \code{\link{ReplaceAbbreviations}},
  \code{\link{ReplaceBacktickApp}},
  \code{\link{ReplaceBacktick}},
  \code{\link{ReplaceContractionsApp}},
  \code{\link{ReplaceContractions}},
  \code{\link{ReplaceCurlyQuotesApp}},
  \code{\link{ReplaceCurlyQuotes}},
  \code{\link{ReplaceEmojiApp}},
  \code{\link{ReplaceEmoji}},
  \code{\link{ReplaceEmoticonApp}},
  \code{\link{ReplaceEmoticon}},
  \code{\link{ReplaceHTMLApp}}, \code{\link{ReplaceHTML}},
  \code{\link{ReplaceInternetSlangApp}},
  \code{\link{ReplaceInternetSlang}},
  \code{\link{ReplaceKernApp}}, \code{\link{ReplaceKern}},
  \code{\link{ReplaceNonAsciiApp}},
  \code{\link{ReplaceNonAscii}},
  \code{\link{ReplaceNumbersApp}},
  \code{\link{ReplaceNumbers}},
  \code{\link{ReplaceOrdinalApp}},
  \code{\link{ReplaceOrdinal}},
  \code{\link{ReplaceSymbolApp}},
  \code{\link{ReplaceSymbol}},
  \code{\link{ReplaceTokensApp}},
  \code{\link{ReplaceTokens}},
  \code{\link{ReplaceWordElongation}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
