% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reshape.R
\docType{class}
\name{Reshape}
\alias{Reshape}
\title{Reshape}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Reshape$new(x, to = c("document", "sentence"))$execute()
}
\arguments{
\item{x}{Object of to be processed. This can be an object of the Corpus, Document, character classes, or a list of character vectors.}

\item{to}{Character vector indicating the level to which the Corpus or
Document object should be reshaped. Valid values are "document" and
"sentence".  The starting letters, "d", and "s" may also be used.}
}
\value{
\code{Reshape} Returns a Corpus or Document object, with text
reshaped accordingly.
}
\description{
\code{Reshape} Reshapes objects into document or sentence level of aggregation.
}
\details{
Class responsible for reshaping Corpus and Document object text into a single
document level text vector, or sentence vectors.
}
\section{Methods}{

 \describe{
  \item{\code{new(...)}}{Instantiates the class object.}
  \item{\code{execute()}}{Executes the process on the object.}
}
}

\section{TextStudio Family of Classes Overview}{

The TextStudio family of classes include the TextStudio class, which is
responsible for executing one or more text processing commands on a
text resource, the core class group that performs the processing functions
on the text, and the command class group which instructs the TextStudio
class to perform an a text processing function on a text.
}

\section{DataStudio Design}{

This DataStudio family of classes is an implementation of the Strategy
design pattern, as described in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). The Strategy pattern defines
the family of text cleaning classes, encapsulates class and allows
text cleaning algorithms to be selected and exchagned at runtime.
}

\examples{

}
\seealso{
Other TextStudio Classes: \code{\link{LowerCaseCmd}},
  \code{\link{LowerCase}}, \code{\link{RemoveEmailCmd}},
  \code{\link{RemoveEmail}},
  \code{\link{RemoveHyphensCmd}},
  \code{\link{RemoveHyphens}},
  \code{\link{RemoveNumbersCmd}},
  \code{\link{RemoveNumbers}},
  \code{\link{RemovePunctCmd}}, \code{\link{RemovePunct}},
  \code{\link{RemoveSymbolsCmd}},
  \code{\link{RemoveSymbols}},
  \code{\link{RemoveTwitterCmd}},
  \code{\link{RemoveTwitter}}, \code{\link{RemoveURLCmd}},
  \code{\link{RemoveURL}},
  \code{\link{RemoveWhiteSpaceCmd}},
  \code{\link{RemoveWhiteSpace}}, \code{\link{RepairFile}},
  \code{\link{ReplaceAbbreviationsCmd}},
  \code{\link{ReplaceAbbreviations}},
  \code{\link{ReplaceBacktickCmd}},
  \code{\link{ReplaceBacktick}},
  \code{\link{ReplaceContractionsCmd}},
  \code{\link{ReplaceContractions}},
  \code{\link{ReplaceCurlyQuotesCmd}},
  \code{\link{ReplaceCurlyQuotes}},
  \code{\link{ReplaceEmojiCmd}},
  \code{\link{ReplaceEmoji}},
  \code{\link{ReplaceEmoticonCmd}},
  \code{\link{ReplaceEmoticon}},
  \code{\link{ReplaceHTMLCmd}}, \code{\link{ReplaceHTML}},
  \code{\link{ReplaceInternetSlangCmd}},
  \code{\link{ReplaceInternetSlang}},
  \code{\link{ReplaceKernCmd}}, \code{\link{ReplaceKern}},
  \code{\link{ReplaceNonAsciiCmd}},
  \code{\link{ReplaceNonAscii}},
  \code{\link{ReplaceNumbersCmd}},
  \code{\link{ReplaceNumbers}},
  \code{\link{ReplaceOrdinalCmd}},
  \code{\link{ReplaceOrdinal}},
  \code{\link{ReplaceSymbolCmd}},
  \code{\link{ReplaceSymbol}},
  \code{\link{ReplaceTokensCmd}},
  \code{\link{ReplaceTokens}},
  \code{\link{ReplaceWordElongationCmd}},
  \code{\link{ReplaceWordElongation}},
  \code{\link{SplitCorpus}}, \code{\link{SplitDocument}},
  \code{\link{StripTextCmd}}, \code{\link{StripText}},
  \code{\link{Tokenize}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
