% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_corpus.R
\name{reshape_corpus}
\alias{reshape_corpus}
\title{Fast Corpus Reshape}
\usage{
reshape_corpus(x, ncores, ...)
}
\arguments{
\item{x}{A \code{\link[quanteda]{corpus}} as built by \code{create_corpus}.}

\item{ncores}{The number of \code{\link[future]{multisession}} workers to be allocated for the tokenization.}

\item{...}{Additional arguments passed to \code{\link[quanteda]{corpus_reshape}} (see 'Details').}
}
\value{
A reshaped \code{\link[quanteda]{corpus}} object.
}
\description{
Reshape a \code{\link[quanteda]{corpus}} in parallel under the \strong{future} paradigm.
}
\details{
This function is a wrapper of \code{\link[quanteda]{corpus_reshape}} but it is much faster as it
uses the \href{https://www.futureverse.org/packages-overview.html}{\strong{future}}
paradigm to parallelize the task.

Use \code{...} to control the reshaping process. For instance, whether to reshape the corpus to
sentences (\code{to = "sentences"}) which is likely the most common usage.
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}
}
