% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample.R
\docType{class}
\name{Sample}
\alias{Sample}
\title{Sample}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Sample <- Sample$new()
}
\arguments{
\item{x}{The source Document or Corpus object.}

\item{n}{Numeric. If n <= 1, the proportion of 'unit' to sample. If n>1, the number
of 'unit' to sample.}

\item{unit}{Character indicating the units to sample. Values
include c('word', 'sentence', 'vector'). The default is 'sentence'}

\item{stratify}{Logical. Applies to Corpus objects with with two or more documents.
If TRUE, stratified sampling will occur within each document; otherwise, samples
will be taken from a single combined document. Default is TRUE}

\item{replace}{Logical. If TRUE, samples will be taken with replacement. Default is FALSE.}

\item{seed}{Numeric. Seed for sampling reproducibility. Default is NULL}
}
\value{
Sample object, containing the Sample for a single Document object.
}
\description{
\code{Sample} Class representing sampled Document objects.
}
\details{
Class containing the sampled representation of a Document object.
}
\section{Core Methods}{

 \itemize{
  \item{\code{new(x)}}{Initializes an object of the Sample class.}
  \item{\code{content}}{Active binding used to set and retrieve Sample content. Sample
  content may be changed via assignment. Referencing this method retrieves the current
  Sample content.}
 }
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
