% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample0.R
\docType{class}
\name{Sample0}
\alias{Sample0}
\title{Sample0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Sample0
}
\arguments{
\item{corpus}{Corpus object.}

\item{name}{Character string indicating the name for the cross-validation set.}

\item{stratify}{Logical. If TRUE (default), splits and sampling will
be stratefied.}

\item{seed}{Numeric used to initialize a pseudorandom number generator.}
}
\description{
\code{Sample0} Base class for Sample and Split Classes
}
\details{
Base class that defines the methods common to the Split and Sample Classes.
}
\section{Methods}{

 \itemize{
  \item{\code{new()}}{Not implemented.}
 }
}

\seealso{
Other Corpus Sample Family of Classes: \code{\link{KFold}},
  \code{\link{Sample}}, \code{\link{Segment}}

Other CorpusStudio Family of Classes: \code{\link{CorpusStudio}},
  \code{\link{KFold}}, \code{\link{Sample}},
  \code{\link{Segment}}, \code{\link{Split}},
  \code{\link{TokenizerNLP}}, \code{\link{TokenizerQ}},
  \code{\link{Tokenizer}}, \code{\link{Token}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
