% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleCorpus.R
\docType{class}
\name{SampleCorpus}
\alias{SampleCorpus}
\title{SampleCorpus}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SampleCorpus
}
\arguments{
\item{x}{Corpus or Token Object}

\item{corpusStudio}{CorpusStudio object}

\item{name}{Character string containing the name to assign to the Token object.}

\item{n}{Numeric. If n <= 1, the proportion of 'unit' to sample. If n>1, the number
of 'unit' to sample.}

\item{stratify}{Logical. Applies to Corpus objects with with two or more documents.
If TRUE, stratified sampling will occur within each document; otherwise, samples
will be taken from a single combined document. Default is TRUE}

\item{replace}{Logical. If TRUE, samples will be taken with replacement. Default is FALSE.}

\item{seed}{Numeric. Seed for sampling reproducibility. Default is NULL}
}
\value{
Token object
}
\description{
\code{SampleCorpus} Encapsulates the command to sample a Corpus object.
}
\seealso{
Other FileStudio Command Classes: \code{\link{CopyFileSet}},
  \code{\link{MoveFileSet}}, \code{\link{ReadCorpus}},
  \code{\link{RepairFileSet}}, \code{\link{Source}},
  \code{\link{WriteCorpus}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
