% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment}
\alias{segment}
\title{segment}
\usage{
segment(corpus, n = 3, weights = NULL, stratify = TRUE, seed = NULL)
}
\arguments{
\item{corpus}{Corpus object.}

\item{n}{Numeric indicator of the number of segments.}

\item{weights}{Numeric vector indicating the weights for each segment. If
NULL, all segments will have equal weight.}

\item{stratify}{Logical. If TRUE (default), splits and sampling will
be stratefied.}

\item{seed}{Numeric used to initialize a pseudorandom number generator.}
}
\value{
List of Corpus objects
}
\description{
\code{segment} Splits a Corpus object into n segments of sizes indicated by a
weighting parameter.
}
\section{Methods}{

 \itemize{
  \item{\code{new()}}{Initializes an object of the segment class.}
  \item{\code{execute(corpus, n=2, weights = c(0.75, 0.25) stratify = TRUE,
  seed = NULL)}}{Executes the corpus segmentation}
 }
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
