% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{slice}
\alias{slice}
\title{slice}
\usage{
slice(v, size)
}
\arguments{
\item{v}{Input vector to be split}

\item{size}{Numeric indicating the number of elements per slice.}
}
\description{
\code{tokenize} Slices a vector into chunks of the given size
}
