% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLM.R
\docType{class}
\name{SLM}
\alias{SLM}
\title{SLM}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SLM
}
\arguments{
\item{x}{a CVSet containing training and test Corpus objects}

\item{train}{Train Corpus object. Ignored if x is a CVSet, required otherwise.}

\item{test}{Test Corpus object. Ignored if x is a CVSet, required otherwise.}

\item{smooth}{Smoothing algorithm. Valid options include:
'katz': Katz Moothing
'kn': Kneser-Ney
'mkn': Modified Kneser-Ney
'sbo': Stupid Backoff}

\item{modelSize}{Numeric between 1, for unigram to 5 for quintgram.}

\item{openVocabulary}{Logical. If TRUE, preprocessing will replace all words
in the test corpus that are not in the training corpus with the pseudo-word
UNK. If FALSE, all words in test corpus are assumed to be in the training
corpus. The default is TRUE.}
}
\description{
\code{SLM} Statistical language model class.
}
\details{
Encapsulates a Statistical Language Model implementing one of several
smoothing algorithsm, such as Katz, Kneser-Ney, Modified Kneser-Ney,
and Stupid-Backoff.
}
\seealso{
Other Statistical Language Model Classes: \code{\link{KNStudio}},
  \code{\link{SLM0}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
