% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLM0.R
\docType{class}
\name{SLM0}
\alias{SLM0}
\title{SLM0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SLM0
}
\description{
\code{SLM0} Base class for the Statistical Language Model Studio (SLM) family of classes.
}
\details{
This base class defines common members for the SLM family of classes.
}
\seealso{
Other Language Model Classes: \code{\link{KN}},
  \code{\link{SLMConfig}}, \code{\link{SLMStudio0}}

Other Statistical Language Model Classes: \code{\link{KNStudio}},
  \code{\link{SLM}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
