% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLMConfig.R
\docType{class}
\name{SLMConfig}
\alias{SLMConfig}
\title{SLMConfig}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SLMConfig
}
\description{
\code{SLMConfig} Class representing a Statistical Language Model (SLM) configuration.
}
\seealso{
Other Language Model Classes: \code{\link{SLMStudio0}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
