% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLMCorpora.R
\docType{class}
\name{SLMCorpora}
\alias{SLMCorpora}
\title{SLMCorpora}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SLMCorpora
}
\arguments{
\item{config}{SLMConfig object containing the model parameters and the
training and test Corpus objects.}
}
\description{
\code{SLMCorpora} Class responsible for preparing training and test Corpora  objects for language modeling.
}
\details{
Class responsible for preparing training Corpus objects for language modeling.
}
\seealso{
Other SLMStudio Classes: \code{\link{KNBuild}},
  \code{\link{KNEstimate}}, \code{\link{KNEvaluate}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
