% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLMEvaluate.R
\docType{class}
\name{SLMEvaluate}
\alias{SLMEvaluate}
\title{SLMEvaluate}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SLMEvaluate
}
\arguments{
\item{config}{SLMConfig object}

\item{model}{A Language Model Object}

\item{test}{A test Corpus object.}
}
\description{
\code{SLMEvaluate} Evaluates a designated language model on a test set.
}
\seealso{
Other KNStudio Classes: \code{\link{KNBuild}},
  \code{\link{KNEstimate}}, \code{\link{MKNBuild}}

Other SLMStudio Classes: \code{\link{KNBuild0}},
  \code{\link{KNBuild}}, \code{\link{KNEstimate}},
  \code{\link{MKNBuild}}, \code{\link{SLMCorpora}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
