% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLMStudio0.R
\docType{class}
\name{SLMStudio0}
\alias{SLMStudio0}
\title{SLMStudio0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SLMStudio0
}
\description{
\code{SLMStudio0} Abstract class for the Statistical Language Model Studio (SLMStudio) family of classes.
}
\details{
This abstract class defines common members for the SLMStudio family of classes.
}
\seealso{
Other Language Model Classes: \code{\link{KN}},
  \code{\link{SLM0}}, \code{\link{SLMConfig}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
