% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snap0.R
\docType{class}
\name{Snap0}
\alias{Snap0}
\title{Snap0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Snap0
}
\arguments{
\item{object}{An NLPStudio object to be saved}

\item{snapName}{A character string containing the unique name / identifier for a particular snapshot of an object.}
}
\description{
\code{Snap0} Manages the process of saving and restoring snapshots of
NLPStudio objects.
}
\details{
\strong{Snap0 Class Overview:}

The Snap0 singleton class defines the interface for saving and restoring snapshots
of NLPStudio objects. Its single object delegates the actual save and restore
operations to the SnapSave and SnapRestore classes, respectively. Once the
request has been fulfilled by the subclasses, the request is logged.

\strong{Snap0 Class Collaborators:}
The collaborators of the Snap0 class are:
\itemize{
 \item Lab: Class in which NLP takes place. The class contains objects of the DocumentCollection class.
 \item DocumentCollection: Composite class containing objects of the Document class.
 \item Document: Leaf class of the Document objects.
 }

\strong{Snap Family of Classes Participants:}
\itemize{
 \item Snap0: This class which manages and reports the saving and restoration of snapshots.
 \item SnapSave: Class which saves the snapshot to disk.
 \item SnapRestore: Class which restores a snapshot from disk.
 }
}
\section{Methods}{

\describe{
 \item{\code{new()}}{Instantiates the Snap0 object. This singleton object is instantiated by the NLPStudio object at package load time.}
 \item{\code{saveSnap(object)}}{Instantiates a SnapSave object to carry out the client's save request.}
 \item{\code{restoreSnap(snapName)}}{Instantiates a SnapRestore object to carry out the client's restore request. The snapshotName variable identifies uniquely the specific snapshot to restore.}
 \item{\code{getSnaps(object)}}{Returns the list of snapshots for an object, list, or data frame formats. The objectName variable identifies the object for which snapshots should be reported.}
 \item{\code{printSnaps(object)}}{Prints a data frame of snapshots for a named object to the console. The objectName variable identifies the object for which snapshots should be reported.}
}
}

\seealso{
Other Snapshot classes: \code{\link{SnapSave}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
