% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source.R
\docType{class}
\name{Source}
\alias{Source}
\title{Source}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Source
}
\arguments{
\item{x}{FileSource Object}

\item{fileStudio}{FileStudio object}
}
\description{
\code{Source} Encapsulates the command to source a FileSet object.
}
\seealso{
Other FileStudio Command Classes: \code{\link{CopyFileSet}},
  \code{\link{MoveFileSet}}, \code{\link{ReadCorpus}},
  \code{\link{RepairFileSet}}, \code{\link{SampleCorpus}},
  \code{\link{TokenizeCorpus}}, \code{\link{WriteCorpus}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
