% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Split.R
\docType{class}
\name{Split}
\alias{Split}
\title{Split}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Split
}
\arguments{
\item{corpus}{Corpus object.}

\item{name}{Character string indicating the name for the cross-validation set.}

\item{train}{Numeric indicating the proportion of the Corpus to allocate
to the training set. Acceptable values are between 0 and 1. The total of the
values for the train, validation and test parameters must equal 1.}

\item{validation}{Numeric indicating the proportion of the Corpus to allocate
to the validation set. Acceptable values are between 0 and 1. The total of the
values for the train, validation and test parameters must equal 1.}

\item{test}{Numeric indicating the proportion of the Corpus to allocate
to the test set. Acceptable values are between 0 and 1. The total of the
values for the train, validation and test parameters must equal 1.}

\item{stratify}{Logical. If TRUE (default), splits and sampling will
be stratefied.}

\item{seed}{Numeric used to initialize a pseudorandom number generator.}
}
\value{
CVSet object
}
\description{
\code{Split} Splits a corpus into a training, test and optional validation set.
}
\details{
Splits a corpus into a training, test and optional validation set. These corpora
are combined into a single cross-validation set or CVSet object.
}
\section{Methods}{

 \itemize{
  \item{\code{new()}}{Initializes an object of the Split class.}
  \item{\code{execute(x, train = 0.75, validation = 0, test = 0.25,
  stratify = TRUE, seed = NULL)}}{Executes the corpus splits.}
 }
}

\seealso{
Other CorpusStudio Family of Classes: \code{\link{CorpusStudio}},
  \code{\link{Sample0}}, \code{\link{Sample}},
  \code{\link{Segment}}, \code{\link{SplitKFold}},
  \code{\link{Tokenizer}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
