% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Split0.R
\docType{class}
\name{Split0}
\alias{Split0}
\title{Split0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Split0
}
\description{
\code{Split0} Abstract class that defines the methods for the Corpus and Dcocument Split classes
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
