% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitCorpus.R
\docType{class}
\name{SplitCorpus}
\alias{SplitCorpus}
\title{SplitCorpus}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SplitCorpus$new(x, splits = c(.7, .15, .15),
                       setNames =  c("Train", "Validation", "Test"),
                       seed = 232)$execute()
}
\arguments{
\item{x}{Object of to be processed. This can be an object of the Corpus, Document, character classes, or a list of character vectors.}

\item{splits}{Numeric vector of splits.}

\item{setNames}{Character string vector, of equal length to the splits vector,
containing the names to which the cross-validation sets will be assigned.}

\item{seed}{Numeric seed used by the sample function.}
}
\description{
\code{SplitCorpus} Splits corpus into training, test and optionally, validation sets
}
\details{
Class reshapes Corpus object text into sentence vectors.
}
\section{Methods}{

 \describe{
  \item{\code{new(...)}}{Instantiates the class object.}
  \item{\code{execute()}}{Executes the process on the object.}
}
}

\section{TextStudio Family of Classes Overview}{

The TextStudio family of classes include four types of classes. The manager classes which
are responsible for executing one or more core class objects through the implementation
of a command class object, supported by support class objects.
}

\section{DataStudio Design}{

This DataStudio family of classes is an implementation of the Strategy
design pattern, as described in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). The Strategy pattern defines
the family of text cleaning classes, encapsulates class and allows
text cleaning algorithms to be selected and exchagned at runtime.
}

\seealso{
Other TextStudio Classes: \code{\link{AddCommaSpaceCmd}},
  \code{\link{AddCommaSpace}}, \code{\link{LowerCaseCmd}},
  \code{\link{LowerCase}}, \code{\link{RemoveEmailCmd}},
  \code{\link{RemoveEmail}},
  \code{\link{RemoveHyphensCmd}},
  \code{\link{RemoveHyphens}},
  \code{\link{RemoveNumbersCmd}},
  \code{\link{RemoveNumbers}},
  \code{\link{RemovePunctCmd}}, \code{\link{RemovePunct}},
  \code{\link{RemoveSymbolsCmd}},
  \code{\link{RemoveSymbols}},
  \code{\link{RemoveTwitterCmd}},
  \code{\link{RemoveTwitter}}, \code{\link{RemoveURLCmd}},
  \code{\link{RemoveURL}},
  \code{\link{RemoveWhiteSpaceCmd}},
  \code{\link{RemoveWhiteSpace}}, \code{\link{RepairFile}},
  \code{\link{ReplaceAbbreviationsCmd}},
  \code{\link{ReplaceAbbreviations}},
  \code{\link{ReplaceBacktickCmd}},
  \code{\link{ReplaceBacktick}},
  \code{\link{ReplaceContractionsCmd}},
  \code{\link{ReplaceContractions}},
  \code{\link{ReplaceCurlyQuotesCmd}},
  \code{\link{ReplaceCurlyQuotes}},
  \code{\link{ReplaceEmojiCmd}},
  \code{\link{ReplaceEmoji}},
  \code{\link{ReplaceEmoticonCmd}},
  \code{\link{ReplaceEmoticon}},
  \code{\link{ReplaceHTMLCmd}}, \code{\link{ReplaceHTML}},
  \code{\link{ReplaceInternetSlangCmd}},
  \code{\link{ReplaceInternetSlang}},
  \code{\link{ReplaceKernCmd}}, \code{\link{ReplaceKern}},
  \code{\link{ReplaceNonAsciiCmd}},
  \code{\link{ReplaceNonAscii}},
  \code{\link{ReplaceNumbersCmd}},
  \code{\link{ReplaceNumbers}},
  \code{\link{ReplaceOrdinalCmd}},
  \code{\link{ReplaceOrdinal}},
  \code{\link{ReplaceSymbolCmd}},
  \code{\link{ReplaceSymbol}},
  \code{\link{ReplaceTokensCmd}},
  \code{\link{ReplaceTokens}},
  \code{\link{ReplaceWordElongationCmd}},
  \code{\link{ReplaceWordElongation}},
  \code{\link{ReshapeCmd}}, \code{\link{Reshape}},
  \code{\link{SplitDocument}}, \code{\link{StripTextCmd}},
  \code{\link{StripText}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
