% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/State.R
\docType{class}
\name{State}
\alias{State}
\title{State}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
State
}
\arguments{
\item{object}{A serializable object to be serialized and saved.}

\item{stateId}{Character string which uniquely identifies an object and its state at a specific point in time.}
}
\description{
\code{State} This class is responsible for serializing and deserializing
object states. The class is comprised of two methods, saveState and
restoreState. The saveState method takes as its parameters, a state
identifier and the object to be saved, then serializes and saves the object
to the States directory. The restoreState method takes a state identifier
as its parameter and reads the object, designated by the state identifier,
from file.
}
\details{
The State class is a participant of the State family of classes.

\strong{State Family of Classes Overview:}
The State family of classes is responsible for object persistence and
data recovery in the NLPStudio package and is comprised of the following
participants.

\strong(StateSave Class Family Participants:)
\itemize{
 \item StateManager: This class receives save and retore requests from
 serializable objects, maintains the inventory of states, and dispatches
 the save/restore method to a State object.
 \item State: This class is responsible for serializing and deserializing
object states.
 }
}
\section{Methods}{

The following methods are defined for this class:
\describe{
 \item{\code{saveState(stateId, object)}}{Method for serializing and storing the object to disk.}
 \item{\code{restoreState(stateId)}}{Method for deserializing and returning the object from disk at state designated by the stateId parameter.}
 }
}

\seealso{
Other State Classes: \code{\link{StateManager}},
  \code{\link{StateRestore0}},
  \code{\link{StateRestoreFile}},
  \code{\link{StateRestoreObject}},
  \code{\link{StateSave0}}, \code{\link{StateSaveFile}},
  \code{\link{StateSaveObject}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
