% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateRestoreObject.R
\docType{class}
\name{StateRestoreObject}
\alias{StateRestoreObject}
\title{StateRestoreObject}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
StateRestoreObject
}
\arguments{
\item{requestId}{Character string indicating the unique request identifier.}

\item{stateId}{Character string which uniquely identifies an object and its state at a specific point in time.}
}
\value{
state The object in the state requested.
}
\description{
\code{StateRestoreObject} This class restores and returns serialized R6
objects from disk. It is a participant of the StateRestore0 family of
classes.
}
\details{
The StateRestore family of classes is an implementation of the strategy
pattern documented in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). This pattern allows for the
definition of a common interface to varying algorithms used to deserialize
and restore objects back into the NLPStudio package.

\strong{Family of StateRestore0 Classes Overview:}
The StateRestoreObject family of classes is responsible for deserializing and
restoring, objects from disk and is comprised of the following participants.

\strong(StateRestoreO Class Family Participants:)
\itemize{
 \item StateRestore0: The abstract class responsible for defining the common
 interface to be used by the participants of the StateRestore0 family of classes.
 \item StateRestoreObject: This class is responsible for deserializing and
 returning objects from disk.
 \item StateRestoreFile: This class is responsible for deserializing and
 returning objects of the File class from disk.
 }
}
\section{Methods}{

The following methods are defined for this class:
\describe{
 \item{\code{validateObject(requestId, stateId)}}{Base method not implemented for this abstract class.}
 \item{\code{restoreState(requestId, stateId)}}{Base method not implemented for this abstract class.}
 }
}

\seealso{
Other State Classes: \code{\link{StateManager}},
  \code{\link{StateRestore0}},
  \code{\link{StateRestoreFile}}, \code{\link{StateSave0}},
  \code{\link{StateSaveFile}},
  \code{\link{StateSaveObject}}, \code{\link{State}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
