% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StripTextApp.R
\docType{class}
\name{StripTextApp}
\alias{StripTextApp}
\title{StripTextApp}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
StripTextApp$new(x, keepChar = "~~", removeDigit = TRUE,
                     removeApostrophe = TRUE, lowerCase = TRUE)$execute()
}
\arguments{
\item{x}{Object of to be processed. This can be an object of the Corpus, Document, character classes, or a list of character vectors.}

\item{keepChar}{A character vector of symbols (i.e., punctuation) that strip
should keep. The default is to strip every symbol except apostrophes and a double
tilde "~~". The double tilde "~~" is included for a convenient means of keeping
word groups together in functions that split text apart based on spaces.
To remove double tildes "~~" set char.keep to NULL.}

\item{removeDigit}{Logical.  If TRUE, digits are removed from the text.}

\item{removeApostrophe}{Logical.  If TRUE, apostrophes are removed from the text.}

\item{lowerCase}{Logical.  If TRUE, text is converted to lower case.}
}
\value{
\code{StripTextApp} Returns text with unwanted characters removed.
}
\description{
\code{StripTextApp}  Strips text of unwanted characters.
}
\details{
A wrapper for \code{\link[qdap]{strip}} Strips text of unwanted characters.
Source \url{https://cran.r-project.org/web/packages/qdap/qdap.pdf}
}
\section{Methods}{

 \describe{
  \item{\code{new(...)}}{Instantiates the class object.}
  \item{\code{execute()}}{Executes the process on the object.}
}
}

\section{TextStudio Family of Classes Overview}{

The TextStudio family of classes include four types of classes. The manager classes which
are responsible for executing one or more core class objects through the implementation
of a command class object, supported by support class objects.
}

\section{DataStudio Design}{

This DataStudio family of classes is an implementation of the Strategy
design pattern, as described in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). The Strategy pattern defines
the family of text cleaning classes, encapsulates class and allows
text cleaning algorithms to be selected and exchagned at runtime.
}

\examples{

}
\seealso{
Other TextStudio Classes: \code{\link{AddCommaSpaceApp}},
  \code{\link{AddCommaSpace}},
  \code{\link{RemoveEmailApp}}, \code{\link{RemoveEmail}},
  \code{\link{RemoveHyphensApp}},
  \code{\link{RemoveHyphens}},
  \code{\link{RemoveNumbersApp}},
  \code{\link{RemoveNumbers}},
  \code{\link{RemovePunctApp}}, \code{\link{RemovePunct}},
  \code{\link{RemoveSymbolsApp}},
  \code{\link{RemoveSymbols}},
  \code{\link{RemoveTwitterApp}},
  \code{\link{RemoveTwitter}}, \code{\link{RemoveURLApp}},
  \code{\link{RemoveURL}},
  \code{\link{RemoveWhiteSpaceApp}},
  \code{\link{RemoveWhiteSpace}},
  \code{\link{ReplaceAbbreviationsApp}},
  \code{\link{ReplaceAbbreviations}},
  \code{\link{ReplaceBacktickApp}},
  \code{\link{ReplaceBacktick}},
  \code{\link{ReplaceContractionsApp}},
  \code{\link{ReplaceContractions}},
  \code{\link{ReplaceCurlyQuotesApp}},
  \code{\link{ReplaceCurlyQuotes}},
  \code{\link{ReplaceEmojiApp}},
  \code{\link{ReplaceEmoji}},
  \code{\link{ReplaceEmoticonApp}},
  \code{\link{ReplaceEmoticon}},
  \code{\link{ReplaceHTMLApp}}, \code{\link{ReplaceHTML}},
  \code{\link{ReplaceInternetSlangApp}},
  \code{\link{ReplaceInternetSlang}},
  \code{\link{ReplaceKernApp}}, \code{\link{ReplaceKern}},
  \code{\link{ReplaceNonAsciiApp}},
  \code{\link{ReplaceNonAscii}},
  \code{\link{ReplaceNumbersApp}},
  \code{\link{ReplaceNumbers}},
  \code{\link{ReplaceOrdinalApp}},
  \code{\link{ReplaceOrdinal}},
  \code{\link{ReplaceSymbolApp}},
  \code{\link{ReplaceSymbol}},
  \code{\link{ReplaceTokensApp}},
  \code{\link{ReplaceTokens}},
  \code{\link{ReplaceWordElongationApp}},
  \code{\link{ReplaceWordElongation}},
  \code{\link{StripText}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
