% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudioCache.R
\docType{class}
\name{StudioCache}
\alias{StudioCache}
\title{Cache}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
StudioCache
}
\description{
\code{Cache} Class for managing the cache for all NLP objects.
}
\details{
Class and methods to create a cache environment, retrieve objects from
the cache, and write objects to the cache.  There are also methods for
obtaininig cached objects from a file and generating a list of the contents
of the cache environment.
}
\section{Methods}{

\describe{
 \item{\code{new()}}{Creates the cache environment.}
 \item{\code{get(key)}}{Retrieves object from cache.}
 \item{\code{set(key, value)}}{Save a key value pair to the cache environment.}
 \item{\code{load(filePath)}}{Loads data from file into the cache environment.}
 \item{\code{ls}}{Lists the items in the cache environment.}
}
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
