% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Super.R
\docType{class}
\name{Super}
\alias{Super}
\title{Super}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Super
}
\description{
\code{Super} Base class for all classes.
}
\details{
Class loads dependencies such as the validator and logging objects,
and provides access to a private member used to hold method parameter data
for validation.
}
\section{Methods}{

 \itemize{
  \item{\code{new()}}{Not implemented for this abstract class.}
  \item{\code{getParams()}}{Returns the object's parameters.}
 }
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
