% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermFreq.R
\docType{class}
\name{TermFreq}
\alias{TermFreq}
\title{TermFreq}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TermFreq
}
\arguments{
\item{x}{A Corpus object}

\item{type}{Character string indicating the type of term frequency matrix
to produce. Valid values are 'dfm' for the quanteda dfm object, 'dtm' for
the tm package DocumentTermMatrix, and 'tdm' for the tm package
TermDocumentMatrix.}

\item{tolower}{Logical, if TRUE, converts text to lower case prior to creating
the term frequency matrix.  The Default is TRUE.}

\item{stem}{Logical, if TRUE, conducts stemming prior to creating
the term frequency matrix.}

\item{dictionary}{For the dfm objects, this is a list of named character
vectors of words to be included in the term frequency matrix. For
the tdm and dtm objects, the dictionary is a vector of terms. The default
is NULL.}

\item{name}{Optional character string containing the name of the object.}

\item{event}{Character string indicating a class associated with an event to be posted to the object's log.}

\item{key}{Character string or character vector containing the key or
keys associated a metadata field or fields.}

\item{value}{Character string or character vector, equal in length to
the key vector, containing the metadata value or values.}
}
\value{
A dfm, TermDocumentMatrix, or a DocumentTermMatrix as requested.
}
\description{
\code{TermFreq} Class responsible for producing term frequency matrices.
}
\details{
Class receives a request for a quanteda dfm, tm TermDocumentMatrix, or a
tm DocumenTermMatrix, then dispatches the request to the appropriate
concrete factory for production.
}
\section{Core Methods}{

 \itemize{
  \item{\code{new(x, type = c("dfm", "dtm", "tdm"), tolower = TRUE,
  stem = FALSE, dictionary = NULL)}}{Initializes an object of the TermFreq class.}
  \item{\code{execute()}}{Executes the build via the appropriate factory.}
  }
}

\section{Metadata Method}{

 \itemize{
  \item{\code{meta(key = NULL, value = NULL)}}{Provides facility for
  managing an object's metadata represented as key/value pairs in
  a list format. Metadata may be indicated via single key and
  value character strings or by pairs of character vectors. If
  no parameters are passed to the method, then it returns the
  current metadata, as well as system generated application
  and system metadata, in data frame format. If a character
  string or vector is passed via the key parameter, the value
  or values associated with the keys are returned.}
  \item{\code{log(cls = class(self)[1], event = NULL)}}{Class for posting to and
  retrieving an object's log. If the event parameter is provided,
  the event is posted to the object's log and the log is returned
  invisibly.  If the event parameter is NULL, the existing log
  is returned.}
}
}

\seealso{
Other Term Frequency Classes: \code{\link{TermFreqDfm}},
  \code{\link{TermFreqDtm}},
  \code{\link{TermFreqFactory0}}, \code{\link{TermFreqTdm}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
