% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermFreqDfm.R
\docType{class}
\name{TermFreqDfm}
\alias{TermFreqDfm}
\title{TermFreqDfm}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TermFreqDfm
}
\arguments{
\item{x}{A quanteda dfm object.}
}
\value{
The dfm method returns a quanteda dfm object.
}
\description{
\code{TermFreqDfm} Domain class containing a term frequency matrix produced by the \code{\link[NLPStudio]{TermFreqFactoryDfm}} class.

@param x The principal object
 @param key Character string or a vector of strings containing a key or
 keys for metadata key/value pair(s).
 @param value Character or numeric, or a vector thereof, containing a value
 or values for metadata key/value pair(s).
 @param format Character string used in the metadata method. Indicates
 whether to return the metadata as  list of lists, or a data.frame.This method
 returns metadata in a list if no key/value pair is provided. The list
 may contain other lists or data frames associated with the type of
 metadata. This parameter specifies whether to return the metadata
 as a list of lists, or a list of data frames. Valid values are:
 c('list', 'data.frame')
}
\details{
TermFreqDfm objects are wrappers for the \code{\link[quanteda]{dfm}} class.
}
\section{Core Methods}{

 \itemize{
  \item{\code{new(x)}}{Initializes an object of the TermFreqDfm class.}
  \item{\code{dfm(value)}}{Active binding that sets the value of the dfm object via
  assignment.  If no value is assigned, the method returns the current quanteda dfm object.}
  }
}

\seealso{
Other Term Frequency Classes: \code{\link{TermFreqDtm}},
  \code{\link{TermFreqFactory0}},
  \code{\link{TermFreqTdm}}, \code{\link{TermFreq}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
