% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermFreqFactory0.R
\docType{class}
\name{TermFreqFactory0}
\alias{TermFreqFactory0}
\title{TermFreqFactory0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TermFreqFactory0
}
\description{
\code{TermFreqFactory0}  Abstract "strategy" class that defines interface for the TermFreq classes.
}
\seealso{
Other DataStudio Classes: \code{\link{POSFactory0}},
  \code{\link{POSFactory}},
  \code{\link{TermFreqFactoryDfm}},
  \code{\link{TermFreqFactoryDtm}},
  \code{\link{TermFreqFactoryTdm}}, \code{\link{Tokenize}}

Other Term Frequency Classes: \code{\link{TermFreqDfm}},
  \code{\link{TermFreqDtm}}, \code{\link{TermFreqTdm}},
  \code{\link{TermFreq}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
