% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermFreqFactoryDfm.R
\docType{class}
\name{TermFreqFactoryDfm}
\alias{TermFreqFactoryDfm}
\title{TermFreqFactoryDfm}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TermFreqFactoryDfm$new(x, tolower = TRUE, stem = FALSE, dictionary = NULL)$execute()
}
\arguments{
\item{x}{A Corpus object}

\item{tolower}{if TRUE, converts all letters to lower case. Default is TRUE.}

\item{stem}{if TRUE, stem words. Default is FALSE.}

\item{dictionary}{a quanteda dictionary class object or a named list
of character vector dictionary entries to apply to the tokens when creating the dfm.
See \code{\link[quanteda]{dfm}} for details.}

\item{...}{Other parameters passed to  \code{\link[quanteda]{dfm}}}
}
\value{
\code{\link{TermFreqFactoryDfm}} object.
}
\description{
\code{TermFreqFactoryDfm}  Strategy for building term frequency matrix objects from quanteda dfm objects.
}
\details{
A wrapper for \code{\link[quanteda]{dfm}}, this classes creates a sparse
document frequency matrix for a Corpus object.
Source \url{https://cran.r-project.org/web/packages/quanteda/quanteda.pdf}
}
\section{Methods}{

 \describe{
  \item{\code{new(...)}}{Instantiates the class object.}
  \item{\code{execute()}}{Executes the process on the object.}
}
}

\section{DataStudio Family of Classes Overview}{

The DataStudio family of classes include the DataStudio class, which is
responsible for executing one or more data processing commands on a
Corpus or Document resource, the core class group that performs the
data processing functions on the objects, and the command class
group which instructs the DataStudio class to perform the
data processing function on the Corpus or Document object.
}

\seealso{
Other DataStudio Classes: \code{\link{TermFreqFactory0}},
  \code{\link{TermFreqFactoryDtm}},
  \code{\link{TermFreqFactoryTdm}}, \code{\link{Tokenize}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
