% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermFreqFactoryDtm.R
\docType{class}
\name{TermFreqFactoryDtm}
\alias{TermFreqFactoryDtm}
\title{TermFreqFactoryDtm}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TermFreqFactoryDtm$new(x, tolower = TRUE, stem = FALSE, dictionary = NULL)$execute()
}
\arguments{
\item{x}{A Corpus object}

\item{tolower}{if TRUE, converts all letters to lower case. Default is TRUE.}

\item{stem}{if TRUE, stem words. Default is FALSE.}

\item{dictionary}{A character vector to be tabulated against. No other
terms will be listed in the result. Defaults to NULL which means
that all terms in doc are listed.}
}
\value{
\code{\link{TermFreqFactoryDtm}} object.
}
\description{
\code{TermFreqFactoryDtm}  Strategy for building term frequency matrix objects from tm package DocumentTermMatrix objects.
}
\details{
A wrapper for tm package DocumentTermMatrix, this classes creates a sparse
document-term matrix for a Corpus object.
Source \url{https://cran.r-project.org/web/packages/tm/tm.pdf.}
}
\seealso{
Other DataStudio Classes: \code{\link{TermFreqFactory0}},
  \code{\link{TermFreqFactoryDfm}},
  \code{\link{TermFreqFactoryTdm}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
