% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermFreqFactoryTdm.R
\docType{class}
\name{TermFreqFactoryTdm}
\alias{TermFreqFactoryTdm}
\title{TermFreqFactoryTdm}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TermFreqFactoryTdm$new(x, tolower = TRUE, stem = FALSE, dictionary = NULL)$execute()
}
\arguments{
\item{x}{A Corpus object}

\item{tolower}{if TRUE, converts all letters to lower case. Default is TRUE.}

\item{stem}{if TRUE, stem words. Default is FALSE.}

\item{dictionary}{A character vector to be tabulated against. No other
terms will be listed in the result. Defaults to NULL which means
that all terms in doc are listed.}
}
\value{
\code{\link{TermFreqFactoryTdm}} object.
}
\description{
\code{TermFreqFactoryTdm}  Strategy for building term frequency matrix objects from tm package TermDocumentMatrix objects.
}
\details{
A wrapper for tm package TermDocumentMatrix, this classes creates a sparse
term-document matrix for a TokensCollection object.
Source \url{https://cran.r-project.org/web/packages/tm/tm.pdf.}
}
\section{Methods}{

 \describe{
  \item{\code{new(...)}}{Instantiates the class object.}
  \item{\code{execute()}}{Executes the process on the object.}
}
}

\section{DataStudio Family of Classes Overview}{

The DataStudio family of classes include the DataStudio class, which is
responsible for executing one or more data processing commands on a
Corpus or Document resource, the core class group that performs the
data processing functions on the objects, and the command class
group which instructs the DataStudio class to perform the
data processing function on the Corpus or Document object.
}

\seealso{
Other DataStudio Classes: \code{\link{TermFreqFactory0}},
  \code{\link{TermFreqFactoryDfm}},
  \code{\link{TermFreqFactoryDtm}}, \code{\link{Tokenize}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
