% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermFreqStrategy0.R
\docType{class}
\name{TermFreqStrategy0}
\alias{TermFreqStrategy0}
\title{TermFreqStrategy0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TermFreqStrategy0
}
\description{
\code{TermFreqStrategy0}  Abstract "strategy" class that defines interface for the TermFreq classes.
}
\seealso{
Other DataStudio Classes: \code{\link{TermFreqStrategyDfm}},
  \code{\link{TermFreqStrategyDtm}},
  \code{\link{TermFreqStrategyTdm}}, \code{\link{Tokenize}}

Other Term Frequency Classes: \code{\link{TermFreqDfm}},
  \code{\link{TermFreqDtm}}, \code{\link{TermFreqTdm}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
