% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermFreqTdm.R
\docType{class}
\name{TermFreqTdm}
\alias{TermFreqTdm}
\title{TermFreqTdm}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TermFreqTdm
}
\arguments{
\item{x}{A tm package TermDocumentMatrix object.}
}
\value{
The tdm method returns a TermDocumentMatrix object.
}
\description{
\code{TermFreqTdm} Domain class containing a term frequency matrix produced by the \code{\link[NLPStudio]{TermFreqFactoryTdm}} class.

@param x The principal object
 @param key Character string or a vector of strings containing a key or
 keys for metadata key/value pair(s).
 @param value Character or numeric, or a vector thereof, containing a value
 or values for metadata key/value pair(s).
 @param format Character string used in the metadata method. Indicates
 whether to return the metadata as  list of lists, or a data.frame.This method
 returns metadata in a list if no key/value pair is provided. The list
 may contain other lists or data frames associated with the type of
 metadata. This parameter specifies whether to return the metadata
 as a list of lists, or a list of data frames. Valid values are:
 c('list', 'data.frame')
}
\details{
TermFreqTdm objects are wrappers for the tm package TermDocumentMatrix class.
}
\section{Core Methods}{

 \itemize{
  \item{\code{new(x)}}{Initializes an object of the TermFreqTdm class.}
  \item{\code{tdm()}}{Active binding that returns the TermDocumentMatrix object.
  object. The current TermDocumentMatrix is set through assignment.}
  }
}

\seealso{
Other Term Frequency Classes: \code{\link{TermFreqDfm}},
  \code{\link{TermFreqDtm}},
  \code{\link{TermFreqFactory0}}, \code{\link{TermFreq}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
