% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Text.R
\docType{class}
\name{Text}
\alias{Text}
\title{Text}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
myText <- Text$new(name = "skiReport", x = avalanche)
}
\arguments{
\item{x}{Character vectors containing actual content of the Text object
or the path to a file containing the text content}

\item{path}{Character string designated the relative path to a file.}

\item{io}{Object of the IOStrategy class. This is an optional parameter.
If io is not specified, the default IOStrategy for the file type will
be used. See ?IOStrategy for further assistance.}

\item{key}{Character string or character vector containing the key or
keys associated a metadata field or fields.}

\item{value}{Character string or character vector, equal in length to
the key vector, containing the metadata value or values.}
}
\value{
Text object, containing the text content, the metadata and
the methods to manage both.
}
\description{
\code{Text} Class for creating, managing, reading and writing text content and metadata.
}
\details{
Class contains text content and metadata for the Text object.
The object is instantiated with a name and a character
vector containing the text. Once the object is created, users may
add, change and remove metadata via key / value pairs. IO
methods support reading and writing the texts in a variety
of formats using the \code{\link[NLPStudio]{IOFactory}} factory
class.
}
\section{Core Methods}{

 \itemize{
  \item{\code{new()}}{Initializes an object of the Text class.}
 }
}

\section{IO Methods}{

 \itemize{
  \item{\code{read(path, io = NULL)}}{Reads an object from a specified
  file using the default or designated IO Strategy. See IOStrategy?
  for details w.r.t. the io parameter. If the IO strategy is not
  indicated, the default io strategy for the file type is used.}
  \item{\code{write(path, io = NULL)}}{Writes an object to a specified
  file using the default or designated IO Strategy. See IOStrategy?
  for details w.r.t. the io parameter. If the IO strategy is not
  indicated, the default io strategy for the file type is used.}
 }
}

\section{Metadata Method}{

 \itemize{
  \item{\code{meta(key = NULL, value = NULL)}}{Provides facility for
  managing an object's metadata represented as key/value pairs in
  a list format. Metadata may be indicated via single key and
  value character strings or by pairs of character vectors. If
  no parameters are passed to the method, then it returns the
  current metadata, as well as system generated application
  and system metadata, in data frame format. If a character
  string or vector is passed via the key parameter, the value
  or values associated with the keys are returned.}
}
}

\examples{
avalanche <- c("SAN FRANCISCO  — She was snowboarding with her boyfriend when ",
          "she heard someone scream 'Avalanche!'",
          "Then John, 39, saw 'a cloud of snow coming down.'")
avalancheText <- Text$new(name = 'skiReport', content = avalanche)

avalancheText <- myText$meta(key = c("author", "editor", "year"),
                      value = c("Dorfmeister", "Huffington", "2018"))

avalancheText$meta()

}
\seealso{
Other Core Classes: \code{\link{Corpus}},
  \code{\link{Document}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
