% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextConfig.R
\docType{class}
\name{TextConfig}
\alias{TextConfig}
\title{TextConfig}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TextConfig
}
\description{
\code{TextConfig} Class containing configuration for text pre-processing
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
