% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextDocument.R
\docType{class}
\name{TextDocument}
\alias{TextDocument}
\title{TextDocument}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
skiReport <- TextDocument$new(name = "skiReport", purpose = 'Train')
}
\arguments{
\item{name}{Character string containing the name for the TextDocument object.}

\item{purpose}{Character string used to indicate how the document will be used, e.g. 'train', 'test'.}

\item{note}{Character string containing a comment associated with a call to the
content method. The contents of the note variable are written to the TextDocuments
log. This is used to track changes to the content, perhaps made during preprocessing.}
}
\value{
TextDocument object, containing the TextDocument content, the metadata and
the methods to manage both.
}
\description{
\code{TextDocument} Entity class for text documents.

@param x The principal object
 @param key Character string or a vector of strings containing a key or
 keys for metadata key/value pair(s).
 @param value Character or numeric, or a vector thereof, containing a value
 or values for metadata key/value pair(s).
}
\details{
Entity class for text documents with methods for adding content from character
vectors.
}
\section{Core Methods}{

 \itemize{
  \item{\code{new(name = NULL, purpose = NULL)}}{Initializes an object of the TextDocument class.}
  \item{\code{content(x, note = NULL)}}{Method for obtaining/adding/updating content. If no
  parameters are presented, the current content is returned.  Otherwise, the content
  is updated with the contents of the character vector 'x'. Sentence, word, token, type,
  sentence and word length statistics are also computed and the metadata is updated
  accordingly.}
  \item{\code{summary()}}{Summarizes the TextDocument object.}
 }
}

\examples{
report <- c("SAN FRANCISCO  — She was snowboarding with her boyfriend when ",
          "she heard someone scream 'Avalanche!'",
          "Then John, 39, saw 'a cloud of snow coming down.'")
avalanche <- TextDocument$new(name = 'avalanche', purpose = 'raw')
avalance$content <- report
key <- c('genre', 'author', 'year')
value <- c('weather', 'chris jones', 2018)
avalanche$meta$setDescriptive(key = key value = value)

}
\seealso{
Other Document Classes: \code{\link{Collection0}},
  \code{\link{Corpus}}, \code{\link{Document0}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
