% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextStudio.R
\docType{class}
\name{TextStudio}
\alias{TextStudio}
\title{TextStudio}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TextStudio
}
\arguments{
\item{object}{The object to be processed.}

\item{queue}{The job queue containing text processing commands.}
}
\description{
\code{TextStudio} Class for performing text cleaning and preprocessing
}
\section{TextStudio Family of Classes Overview}{

The TextStudio family of classes include the TextStudio class, which is
responsible for executing one or more text processing commands on a
text resource, the core class group that performs the processing functions
on the text, and the command class group which instructs the TextStudio
class to perform an a text processing function on a text.
}

\section{TextStudio methods}{

\strong{Core Methods:}
 \itemize{
  \item{\code{new()}}{Method for instantiating a TextStudio.}
  \item{\code{addCommand()}}{Method that adds a text processing command to the queue. }
  \item{\code{removeCommand()}}{Method that for removing a command from the queue.}
  \item{\code{execute()}}{Method that executes the job queue. }
  \item{\code{getResult()}}{Method that returns the object following execution of the job queue. }
 }

\strong{Other Methods:}
 \itemize{
  \item{\code{accept(visitor)}}{Accepts a visitor object. Not implemented for this abstract class}
  \item{\code{logIt(level = 'Info')}}{Logs events relating to the TextStudio.}
 }
}

\section{Parameters}{

}

\seealso{
Other TextStudio classes: \code{\link{TextStudio0}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
