% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextStudio0.R
\docType{class}
\name{TextStudio0}
\alias{TextStudio0}
\title{TextStudio0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TextStudio0
}
\arguments{
\item{x}{Object of to be processed. This can be an object of the Corpus, Document, character classes, or a list of character vectors.}
}
\description{
\code{TextStudio0} Abstract class  for the TextStudio family of classes.
}
\details{
This abstract class defines a common interface and methods for the TextStudio
family of classes.
}
\section{Methods}{

 \describe{
  \item{\code{new(...)}}{Instantiates the class object.}
  \item{\code{execute()}}{Executes the process on the object.}
}
}

\section{TextStudio Family of Classes Overview}{

The TextStudio family of classes include the TextStudio class, which is
responsible for executing one or more text processing commands on a
text resource, the core class group that performs the processing functions
on the text, and the command class group which instructs the TextStudio
class to perform an a text processing function on a text.
}

\section{TextStudio Design}{

This TextStudio family of classes is an implementation of the Strategy
design pattern, as described in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). The Strategy pattern defines
the family of text cleaning classes, encapsulates class and allows
text cleaning algorithms to be selected and exchagned at runtime.
}

\seealso{
Other TextStudio classes: \code{\link{TextStudio}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
