% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Token.R
\docType{class}
\name{Token}
\alias{Token}
\title{Token}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Token
}
\arguments{
\item{x}{Corpus object or a character vector containing text}

\item{n}{Numeric for the nGram method. Indicates the nGram size.}
}
\value{
Tokenizer object
}
\description{
\code{Token} Tokens class
}
\details{
Creates character, word, sentence and paragraph tokens and nGrams using the
tokenizer package.
}
\section{Methods}{

 \itemize{
  \item{\code{new()}{Initializes an object of the Tokenizer class.}}
  \item{\code{character(x)}}
  \item{\code{word(x)}}
  \item{\code{sentence(x)}}
  \item{\code{paragraph(x)}}
  \item{\code{nGrams(x, n)}}
 }
}

\seealso{
Other Corpus Tokenizer Family of Classes: \code{\link{Tokenizer}}

Other CorpusStudio Family of Classes: \code{\link{CorpusStudio}},
  \code{\link{KFold}}, \code{\link{Sample0}},
  \code{\link{Sample}}, \code{\link{Segment}},
  \code{\link{Split}}, \code{\link{TokenizerNLP}},
  \code{\link{TokenizerQ}}, \code{\link{Tokenizer}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\concept{Corpus Tokenizer Family of Classes}
\concept{CorpusStudio Family of Classes}
\keyword{datasets}
