% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tokenize.R
\docType{class}
\name{Tokenize}
\alias{Tokenize}
\title{Tokenize}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Tokenize$new(x, what = c("sentence", "word"))$execute()
}
\arguments{
\item{x}{Object of to be processed. This can be an object of the Corpus, Document, character classes, or a list of character vectors.}

\item{what}{Character string containing either c('character', 'word' ,'sentence)
indicating to which format the document should be tokenized.}
}
\value{
\code{Tokenize} A tokenized Corpus, Document, or character text
object.
}
\description{
\code{Tokenize}  Tokenizes text into character, word, or sentence tokens.
}
\details{
This class is a wrapper for the  \code{\link[quanteda]{tokens}} function for
character, and word, tokenization. Sentence tokenization functionality is
provided using the openNLP package.

Sources:
\url{https://cran.r-project.org/web/packages/quanteda/quanteda.pdf}
\url{https://cran.r-project.org/web/packages/openNLP/openNLP.pdf}
}
\section{Methods}{

 \describe{
  \item{\code{new(...)}}{Instantiates the class object.}
  \item{\code{execute()}}{Executes the process on the object.}
}
}

\section{DataStudio Family of Classes Overview}{

The DataStudio family of classes include the DataStudio class, which is
responsible for executing one or more data processing commands on a
Corpus or Document resource, the core class group that performs the
data processing functions on the objects, and the command class
group which instructs the DataStudio class to perform the
data processing function on the Corpus or Document object.
}

\section{DataStudio Design}{

This DataStudio family of classes is an implementation of the Strategy
and Command design patterns, as described in the book "Design Patterns:
Elements of Reusable Object-Oriented Software" by Erich Gamma, Richard
Helm, Ralph Johnson and John Vlissides (hence Gang of Four). The
Strategy pattern defines and encapsulates the family of data processing
classes and allows text cleaning algorithms to be selected and
exhanged at runtime. THe Command pattern defines commands that
can be executed in batch mode or at a later time.
}

\examples{

}
\seealso{
Other DataStudio Classes: \code{\link{TermFreqFactory0}},
  \code{\link{TermFreqFactoryDfm}},
  \code{\link{TermFreqFactoryDtm}},
  \code{\link{TermFreqFactoryTdm}}

Other Tokens Classes: \code{\link{TokensCollection}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
