% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize_corpus.R
\name{tokenize_corpus}
\alias{tokenize_corpus}
\title{Fast Corpus Tokenization}
\usage{
tokenize_corpus(x, ncores, ...)
}
\arguments{
\item{x}{A \strong{quanteda} \link{corpus} as built by \code{create_corpus}.}

\item{ncores}{The number of \link{multisession} workers to be allocated for the tokenization.}

\item{...}{Additional arguments passed to \link{tokens}.}
}
\value{
A \link{tokens} object.
}
\description{
Tokenize a \link{corpus} in parallel via \strong{\link{future}}.
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}
}
