% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TokenizeCorpus.R
\docType{class}
\name{TokenizeCorpus}
\alias{TokenizeCorpus}
\title{TokenizeCorpus}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TokenizeCorpus
}
\arguments{
\item{x}{Corpus Object}

\item{corpusStudio}{CorpusStudio object}

\item{name}{Character string containing the name to assign to the Token object.}

\item{tokenUnit}{Character string indicating the type of Token object to
produce. Valid values are c('character', 'word', 'sentence', 'paragraph')}
}
\value{
Token object
}
\description{
\code{TokenizeCorpus} Encapsulates the command to tokenize a Corpus object.
}
\seealso{
Other FileStudio Command Classes: \code{\link{BuildCorpus}},
  \code{\link{CleanCorpus}}, \code{\link{CopyFileSet}},
  \code{\link{MoveFileSet}}, \code{\link{ReadCorpus}},
  \code{\link{RepairFileSet}}, \code{\link{SampleCorpus}},
  \code{\link{Source}}, \code{\link{SplitCorpus}},
  \code{\link{WriteCorpus}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
