% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tokenizer.R
\docType{class}
\name{Tokenizer}
\alias{Tokenizer}
\title{Tokenizer}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Tokenizer$new(x, what = c("sentence", "word"))$execute()
}
\arguments{
\item{what}{Character string containing either c('character', 'word' ,'sentence)
indicating to which format the document should be tokenized.}
}
\value{
\code{Tokenizer} A tokenized Corpus, Document, or character text
object.
}
\description{
\code{Tokenizer}  Creates tokenized corpus objects.
}
\details{
This class is a wrapper for the  \code{\link[quanteda]{tokens}} function for
character, and word, tokenization. Sentence tokenization functionality is
provided using the openNLP package.

Sources:
\url{https://cran.r-project.org/web/packages/quanteda/quanteda.pdf}
\url{https://cran.r-project.org/web/packages/openNLP/openNLP.pdf}
}
\examples{

}
\seealso{
Other DataStudio Classes: \code{\link{TermFreqFactory0}},
  \code{\link{TermFreqFactoryDfm}},
  \code{\link{TermFreqFactoryDtm}},
  \code{\link{TermFreqFactoryTdm}}

Other Tokens Classes: \code{\link{Tokens}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
