% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TokenizerNLP.R
\docType{class}
\name{TokenizerNLP}
\alias{TokenizerNLP}
\title{TokenizerNLP}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TokenizerNLP
}
\arguments{
\item{x}{Corpus object}

\item{n}{Numeric for the nGram method. Indicates the nGram size.}
}
\value{
TokenizerNLP object
}
\description{
\code{TokenizerNLP} Wrapper class for the OpenNLP package tokens functions
}
\details{
Creates character, word, sentence and paragraph tokens and nGrams using the
OpenNLP package.
}
\section{Methods}{

 \itemize{
  \item{\code{new()}{Initializes an object of the TokenizerNLP class.}}
  \item{\code{character(x)}}
  \item{\code{word(x)}}
  \item{\code{sentence(x)}}
  \item{\code{paragraph(x)}}
  \item{\code{nGrams(x, n)}}
 }
}

\seealso{
Other CorpusStudio Family of Classes: \code{\link{CorpusStudio}},
  \code{\link{Sample0}}, \code{\link{Sample}},
  \code{\link{Segment}}, \code{\link{SplitKFold}},
  \code{\link{Split}}, \code{\link{TokenizerQ}},
  \code{\link{Tokenizer}}, \code{\link{Token}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
