% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tokens.R
\docType{class}
\name{Tokens}
\alias{Tokens}
\title{Tokens}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
tokenizedCorpus <- Tokens$new(x = corpus, to = "sentence")
}
\arguments{
\item{name}{Optional character string containing the name of the object.}

\item{event}{Character string indicating a class associated with an event to be posted to the object's log.}

\item{x}{A Corpus class object to be tokenized}

\item{to}{Character string indicating the level of tokenization. Valid
values are c("sentence", "character"), The first letter of the level
of tokenization may also be used in lieu of the word.}

\item{key}{Character string or character vector containing the key or
keys associated a metadata field or fields.}

\item{value}{Character string or character vector, equal in length to
the key vector, containing the metadata value or values.}
}
\value{
Tokens object.
}
\description{
\code{Tokens} Class for creating, managing, reading and writing Tokens objects.
}
\details{
Tokens objects are Corpus level objects containing collections of Document
objects which have been tokenized using the Tokenize class.
}
\section{Core Methods}{

 \itemize{
  \item{\code{new()}}{Initializes an object of the Tokens class.}
 }
}

\section{Metadata Method}{

 \itemize{
  \item{\code{meta(key = NULL, value = NULL)}}{Provides facility for
  managing an object's metadata represented as key/value pairs in
  a list format. Metadata may be indicated via single key and
  value character strings or by pairs of character vectors. If
  no parameters are passed to the method, then it returns the
  current metadata, as well as system generated application
  and system metadata, in data frame format. If a character
  string or vector is passed via the key parameter, the value
  or values associated with the keys are returned.}
  \item{\code{log(cls = class(self)[1], event = NULL)}}{Class for posting to and
  retrieving an object's log. If the event parameter is provided,
  the event is posted to the object's log and the log is returned
  invisibly.  If the event parameter is NULL, the existing log
  is returned.}
}
}

\seealso{
Other Data Classes: \code{\link{TokensDocument}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
