% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateDate.R
\docType{class}
\name{ValidateDate}
\alias{ValidateDate}
\title{ValidateDate}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidateDate
}
\value{
date A POSIXct date object if the date was recognized or FALSE otherwise.
}
\description{
\code{ValidateDate} Class for validating and parsing dates
}
\details{
This class provide a methods for validating and parsing dates. This class
uses the 'parsedate' package to return a POSIXct object for any
recognizable date format.
}
\section{Methods}{

\describe{
 \item{\code{new()}}{Creates an object of ValidateDate class}
 \item{\code{validate(date)}}{Method for validating and parsing dates}
}
}

\seealso{
Other Validation Classes: \code{\link{Validate0}},
  \code{\link{ValidateClass}},
  \code{\link{ValidateExists}},
  \code{\link{ValidateLogical}},
  \code{\link{ValidateName}},
  \code{\link{ValidateNoSpaces}},
  \code{\link{ValidateNotEmpty}},
  \code{\link{ValidateNumeric}},
  \code{\link{ValidatePath}}, \code{\link{ValidateString}},
  \code{\link{ValidateUrl}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
