% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validators.R
\name{validateLogical}
\alias{validateLogical}
\title{validateLogical}
\usage{
validateLogical(object)

Class accpe
}
\arguments{
\item{object}{The object to be validated}
}
\description{
\code{validateLogical} Validates a parameters that take logical values.
}
\details{
Class accepts an object with a public parameters member. The parameters
member must be a list with two elements: (1) a vector of variable
names, and (2) a vector containing the values.
}
\seealso{
Other Validation Functions: \code{\link{validateDiscrete}},
  \code{\link{validateKeyValue}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
