% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidatePath.R
\docType{class}
\name{ValidatePath}
\alias{ValidatePath}
\title{ValidatePath}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidatePath
}
\arguments{
\item{cls}{Character string containing the name of the invoking class}

\item{method}{Character string containing the name of the invoking method}

\item{fieldName}{Character string containing the name of the field to validate}

\item{value}{The value of the field being validated}

\item{level}{Character string c("Error", "Warn", "Info)}

\item{msg}{Character vector containing the message to be displayed if test fails}

\item{expect}{Logical c(TRUE, FALSE) indicating expected outcome}
}
\description{
\code{ValidatePath} Class for validating a file path

This class provide a methods for validating file paths.
}
\seealso{
Other Validation Classes: \code{\link{Validate0}},
  \code{\link{ValidateClass}}, \code{\link{ValidateDate}},
  \code{\link{ValidateExists}},
  \code{\link{ValidateLogical}},
  \code{\link{ValidateName}},
  \code{\link{ValidateNoSpaces}},
  \code{\link{ValidateNotEmpty}},
  \code{\link{ValidateNumeric}},
  \code{\link{ValidateString}}, \code{\link{ValidateUrl}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
