% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validators.R
\name{validatePatternReplace}
\alias{validatePatternReplace}
\title{validatePatternReplace}
\usage{
validatePatternReplace(object)
}
\arguments{
\item{object}{The object to be validated}
}
\description{
\code{validatePatternReplace} Validates a parameters that contain pattern and replacements
}
\details{
Function accepts an object with a public parameters member. The parameters
member has one of two formats.

Single Element Format: This must be a data frame containing both patterns
and their replacements.

Two Element Format: The first element is a vector containing the patterns
to replace.  The second element must be the replacement vector. This second
vector  may be of length one, in which case, every pattern will replaced with this
single element. Alternatively the replacement vector must be the same
length of the pattern vector.
}
\seealso{
Other Validation Functions: \code{\link{validateDiscrete}},
  \code{\link{validateLogical}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
