% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validators.R
\name{validateRange}
\alias{validateRange}
\title{validateRange}
\usage{
validateRange(object)
}
\arguments{
\item{object}{The object to be validated}
}
\description{
\code{validateRange} Validates a parameter that takes a range of values.
}
\details{
Class accepts an object with a public parameters member. The parameters
member must be a list with four elements: (1) the variable name,
(2) the variable value, (3), the low range, and (4), the high range.
All ranges are inclusive.
}
\seealso{
Other Validation Functions: \code{\link{validateClass}},
  \code{\link{validateDiscrete}},
  \code{\link{validateKeyValue}},
  \code{\link{validateLogical}}, \code{\link{validatePath}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
