% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateUrl.R
\docType{class}
\name{ValidateUrl}
\alias{ValidateUrl}
\title{ValidateUrl}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidateUrl
}
\value{
A logical TRUE if valid url, FALSE otherwise
}
\description{
\code{ValidateUrl} Class for validating website urls
}
\details{
This class provide a methods for validating website urls
}
\section{Methods}{

\describe{
 \item{\code{new()}}{Creates an object of ValidateUrl class}
 \item{\code{validate(object)}}{Method for validating urls}
}
}

\seealso{
Other Validation Classes: \code{\link{Validate0}},
  \code{\link{ValidateClass}}, \code{\link{ValidateDate}},
  \code{\link{ValidateExists}},
  \code{\link{ValidateLogical}},
  \code{\link{ValidateName}},
  \code{\link{ValidateNoSpaces}},
  \code{\link{ValidateNotEmpty}},
  \code{\link{ValidateNumeric}},
  \code{\link{ValidatePath}}, \code{\link{ValidateString}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
