% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidationVisitor.R
\docType{class}
\name{ValidationVisitor}
\alias{ValidationVisitor}
\title{ValidationVisitor}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidationVisitor
}
\arguments{
\item{object}{The object in its current state}
}
\description{
\code{ValidationVisitor} Visitor class responsible for validation
}
\details{
\strong{ValidationVisitor Methods:}
The ValidationVisitor methods are as follows:
 \itemize{
  \item{\code{nlpStudio(object)}}{Method for validating the instantiation of the NLPStudio object}
}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
