% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validator.R
\docType{class}
\name{Validator}
\alias{Validator}
\title{Validator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Validator
}
\arguments{
\item{object}{Originator object}

\item{...}{Parameters being validated}
}
\value{
A logical, TRUE if validation criteria are met, FALSE, otherwise.
}
\description{
\code{Validator} Class responsible for validation of objects and parameters.

This class contains the parameter validation methods for objects. Methods
provide validation of parameters for initialization, composite, state,
and read/write operations.


\strong{Validator Methods:}
\describe{
 \item{\code{new()}}{Creates an object of Validator class}
 \item{\code{init(object)}}{Dispatches the initialization validation visitor, via the accept method of the originator.}
 \item{\code{getObject(object)}}{Dispatches the getObject validation visitor, via the accept method of the originator.}
 \item{\code{addChild(object)}}{Dispatches the addChild validation visitor, via the accept method of the originator.}
 \item{\code{removeChild(object)}}{Dispatches the removeChild validation visitor, via the accept method of the originator.}
 \item{\code{restore(object)}}{Dispatches the restore validation visitor, via the accept method of the originator.}
 \item{\code{readState(object)}}{Dispatches the readState validation visitor, via the accept method of the originator.}
 \item{\code{writeState(object)}}{Dispatches the writeState validation visitor, via the accept method of the originator.}
 \item{\code{read(object)}}{Dispatches the read validation visitor, via the accept method of the originator.}
 \item{\code{write(object)}}{Dispatches the write validation visitor, via the accept method of the originator.}
}

\strong{Validator Parameters:}
}
\seealso{
Other Validation Classes: \code{\link{Authorizer}},
  \code{\link{Validator0}}, \code{\link{ValidatorClass}},
  \code{\link{ValidatorDate}},
  \code{\link{ValidatorExists}},
  \code{\link{ValidatorLogical}},
  \code{\link{ValidatorName}},
  \code{\link{ValidatorNoSpaces}},
  \code{\link{ValidatorNotEmpty}},
  \code{\link{ValidatorNumeric}},
  \code{\link{ValidatorPath}},
  \code{\link{ValidatorString}}, \code{\link{ValidatorUrl}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
