% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validator0.R
\docType{class}
\name{Validator0}
\alias{Validator0}
\title{Validator0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Validator0
}
\arguments{
\item{class}{Character string containing the name of the invoking class}

\item{method}{Character string containing the name of the invoking method}

\item{fieldName}{Character string containing the name of the field to validate}

\item{value}{The value of the field being validated}

\item{level}{Character string c("Error", "Warn", "Info)}

\item{msg}{Character vector containing the message to be displayed if test fails}

\item{expect}{Logical c(TRUE, FALSE) indicating expected outcome}
}
\value{
A logical, TRUE if validation criteria are met, FALSE, otherwise.
}
\description{
\code{Validator0} Abstract class for the field validation classes. Defines
the interface for field validation, and provides methods for logging
and reporting validation errors.
}
\details{
\strong{Validator0 Methods:}
\describe{
 \item{\code{validate(class, method, fieldName, value, level, msg, expect)}}{Field level validation method overwritten by subclasses.}
 \item{\code{notify(class, method, fieldName, value, level, msg, expect)}}{Renders notification of field level validation.}
}

\strong{Validator0 Parameters:}
}
\seealso{
Other Validation Classes: \code{\link{Authorizer}},
  \code{\link{ValidatorClass}},
  \code{\link{ValidatorDate}},
  \code{\link{ValidatorExists}},
  \code{\link{ValidatorLogical}},
  \code{\link{ValidatorName}},
  \code{\link{ValidatorNoSpaces}},
  \code{\link{ValidatorNotEmpty}},
  \code{\link{ValidatorNumeric}},
  \code{\link{ValidatorPath}},
  \code{\link{ValidatorString}},
  \code{\link{ValidatorUrl}}, \code{\link{Validator}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
