% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidatorClass.R
\docType{class}
\name{ValidatorClass}
\alias{ValidatorClass}
\title{ValidatorClass}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidatorClass
}
\value{
A logical TRUE if class is equal to expected value, FALSE otherwise.
}
\description{
\code{ValidatorClass} Class for validating object class
}
\details{
This class provide a methods for validating an object's class.
}
\section{Methods}{

\describe{
 \item{\code{new()}}{Creates an object of ValidatorClass class}
 \item{\code{validate(object)}}{Method for validating logical}
}
}

\seealso{
Other Validation Classes: \code{\link{VValidatorAddChild}},
  \code{\link{VValidatorGetObject}},
  \code{\link{VValidatorInit}},
  \code{\link{VValidatorParent}},
  \code{\link{VValidatorRemoveChild}},
  \code{\link{VValidatorRestore}},
  \code{\link{Validator0}}, \code{\link{ValidatorDate}},
  \code{\link{ValidatorExists}},
  \code{\link{ValidatorLogical}},
  \code{\link{ValidatorName}},
  \code{\link{ValidatorNoSpaces}},
  \code{\link{ValidatorNotEmpty}},
  \code{\link{ValidatorNumeric}},
  \code{\link{ValidatorPath}},
  \code{\link{ValidatorString}},
  \code{\link{ValidatorUrl}}, \code{\link{Validator}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
