% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidatorDate.R
\docType{class}
\name{ValidatorDate}
\alias{ValidatorDate}
\title{ValidatorDate}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidatorDate
}
\value{
date A POSIXct date object if the date was recognized or FALSE otherwise.
}
\description{
\code{ValidatorDate} Class for validating and parsing dates
}
\details{
This class provide a methods for validating and parsing dates. This class
uses the 'parsedate' package to return a POSIXct object for any
recognizable date format.
}
\section{Methods}{

\describe{
 \item{\code{new()}}{Creates an object of ValidatorDate class}
 \item{\code{validate(date)}}{Method for validating and parsing dates}
}
}

\seealso{
Other Validation Classes: \code{\link{VValidatorAddChild}},
  \code{\link{VValidatorGetObject}},
  \code{\link{VValidatorInit}},
  \code{\link{VValidatorParent}},
  \code{\link{VValidatorRemoveChild}},
  \code{\link{VValidatorRestore}},
  \code{\link{Validator0}}, \code{\link{ValidatorClass}},
  \code{\link{ValidatorExists}},
  \code{\link{ValidatorLogical}},
  \code{\link{ValidatorName}},
  \code{\link{ValidatorNoSpaces}},
  \code{\link{ValidatorNotEmpty}},
  \code{\link{ValidatorNumeric}},
  \code{\link{ValidatorPath}},
  \code{\link{ValidatorString}},
  \code{\link{ValidatorUrl}}, \code{\link{Validator}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
