% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidatorExists.R
\docType{class}
\name{ValidatorExists}
\alias{ValidatorExists}
\title{ValidatorExists}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidatorExists
}
\value{
A logical  TRUE if the object exists, FALSE otherwise
}
\description{
\code{ValidatorExists} Class for validating whether an object exists
}
\details{
This class provide a methods for validating the existence of objects in
the global environment
}
\seealso{
Other Validation Classes: \code{\link{Authorizer}},
  \code{\link{Validator0}}, \code{\link{ValidatorClass}},
  \code{\link{ValidatorDate}},
  \code{\link{ValidatorLogical}},
  \code{\link{ValidatorName}},
  \code{\link{ValidatorNoSpaces}},
  \code{\link{ValidatorNotEmpty}},
  \code{\link{ValidatorNumeric}},
  \code{\link{ValidatorPath}},
  \code{\link{ValidatorString}},
  \code{\link{ValidatorUrl}}, \code{\link{Validator}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
