% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidatorLogical.R
\docType{class}
\name{ValidatorLogical}
\alias{ValidatorLogical}
\title{ValidatorLogical}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidatorLogical
}
\value{
A logical TRUE if valid logical, FALSE otherwise
}
\description{
\code{ValidatorLogical} Class for validating logicals
}
\details{
This class provide a methods for validating logicals.
}
\section{Methods}{

\describe{
 \item{\code{new()}}{Creates an object of ValidatorLogical class}
 \item{\code{validate(object)}}{Method for validating logical}
}
}

\seealso{
Other Validation Classes: \code{\link{Authorizer}},
  \code{\link{Validator0}}, \code{\link{ValidatorClass}},
  \code{\link{ValidatorDate}},
  \code{\link{ValidatorExists}},
  \code{\link{ValidatorName}},
  \code{\link{ValidatorNoSpaces}},
  \code{\link{ValidatorNotEmpty}},
  \code{\link{ValidatorNumeric}},
  \code{\link{ValidatorPath}},
  \code{\link{ValidatorString}},
  \code{\link{ValidatorUrl}}, \code{\link{Validator}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
