% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidatorNoSpaces.R
\docType{class}
\name{ValidatorNoSpaces}
\alias{ValidatorNoSpaces}
\title{ValidatorNoSpaces}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidatorNoSpaces
}
\value{
a logical TRUE if valid string, FALSE otherwise
}
\description{
\code{ValidatorNoSpaces} Class for validating string has no spaces
}
\details{
This class provide a methods for validating character string has no spaces
}
\section{Methods}{

\describe{
 \item{\code{new()}}{Creates an object of ValidatorNoSpaces class}
 \item{\code{validate(object)}}{Method for validating string}
}
}

\seealso{
Other Validation Classes: \code{\link{Authorizer}},
  \code{\link{Validator0}}, \code{\link{ValidatorClass}},
  \code{\link{ValidatorDate}},
  \code{\link{ValidatorExists}},
  \code{\link{ValidatorLogical}},
  \code{\link{ValidatorName}},
  \code{\link{ValidatorNotEmpty}},
  \code{\link{ValidatorNumeric}},
  \code{\link{ValidatorPath}},
  \code{\link{ValidatorString}},
  \code{\link{ValidatorUrl}}, \code{\link{Validator}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
