% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidatorNotEmpty.R
\docType{class}
\name{ValidatorNotEmpty}
\alias{ValidatorNotEmpty}
\title{ValidatorNotEmpty}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidatorNotEmpty
}
\value{
a logical TRUE if not empty, FALSE otherwise
}
\description{
\code{ValidatorNotEmpty} Class for validating string is not empty
}
\details{
This class provide a methods for validating character string is not empty
}
\section{Methods}{

\describe{
 \item{\code{new()}}{Creates an object of ValidatorNotEmpty class}
 \item{\code{validate(object)}}{Method for validating string}
}
}

\seealso{
Other Validation Classes: \code{\link{VValidatorAddChild}},
  \code{\link{VValidatorGetObject}},
  \code{\link{VValidatorInit}},
  \code{\link{VValidatorParent}},
  \code{\link{VValidatorRemoveChild}},
  \code{\link{VValidatorRestore}},
  \code{\link{Validator0}}, \code{\link{ValidatorClass}},
  \code{\link{ValidatorDate}},
  \code{\link{ValidatorExists}},
  \code{\link{ValidatorLogical}},
  \code{\link{ValidatorName}},
  \code{\link{ValidatorNoSpaces}},
  \code{\link{ValidatorNumeric}},
  \code{\link{ValidatorPath}},
  \code{\link{ValidatorString}},
  \code{\link{ValidatorUrl}}, \code{\link{Validator}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
