% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidatorNumeric.R
\docType{class}
\name{ValidatorNumeric}
\alias{ValidatorNumeric}
\title{ValidatorNumeric}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidatorNumeric
}
\value{
A logical TRUE if valid numeric, FALSE otherwise
}
\description{
\code{ValidatorNumeric} Class for validating numerics
}
\details{
This class provide a methods for validating numerics.
}
\section{Methods}{

\describe{
 \item{\code{new()}}{Creates an object of ValidatorNumeric class}
 \item{\code{validate(object)}}{Method for validating numeric}
}
}

\seealso{
Other Validation Classes: \code{\link{VValidatorAddChild}},
  \code{\link{VValidatorGetObject}},
  \code{\link{VValidatorInit}},
  \code{\link{VValidatorParent}},
  \code{\link{VValidatorRemoveChild}},
  \code{\link{VValidatorRestore}},
  \code{\link{Validator0}}, \code{\link{ValidatorClass}},
  \code{\link{ValidatorDate}},
  \code{\link{ValidatorExists}},
  \code{\link{ValidatorLogical}},
  \code{\link{ValidatorName}},
  \code{\link{ValidatorNoSpaces}},
  \code{\link{ValidatorNotEmpty}},
  \code{\link{ValidatorPath}},
  \code{\link{ValidatorString}},
  \code{\link{ValidatorUrl}}, \code{\link{Validator}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
