% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidatorPath.R
\docType{class}
\name{ValidatorPath}
\alias{ValidatorPath}
\title{ValidatorPath}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidatorPath
}
\arguments{
\item{class}{Character string containing the name of the invoking class}

\item{method}{Character string containing the name of the invoking method}

\item{fieldName}{Character string containing the name of the field to validate}

\item{value}{The value of the field being validated}

\item{level}{Character string c("Error", "Warn", "Info)}

\item{msg}{Character vector containing the message to be displayed if test fails}

\item{expect}{Logical c(TRUE, FALSE) indicating expected outcome}
}
\description{
\code{ValidatorPath} Class for validating a file path

This class provide a methods for validating file paths.
}
\seealso{
Other Validation Classes: \code{\link{VValidatorAddChild}},
  \code{\link{VValidatorGetObject}},
  \code{\link{VValidatorInit}},
  \code{\link{VValidatorParent}},
  \code{\link{VValidatorRemoveChild}},
  \code{\link{VValidatorRestore}},
  \code{\link{Validator0}}, \code{\link{ValidatorClass}},
  \code{\link{ValidatorDate}},
  \code{\link{ValidatorExists}},
  \code{\link{ValidatorLogical}},
  \code{\link{ValidatorName}},
  \code{\link{ValidatorNoSpaces}},
  \code{\link{ValidatorNotEmpty}},
  \code{\link{ValidatorNumeric}},
  \code{\link{ValidatorString}},
  \code{\link{ValidatorUrl}}, \code{\link{Validator}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
