% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidatorString.R
\docType{class}
\name{ValidatorString}
\alias{ValidatorString}
\title{ValidatorString}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidatorString
}
\value{
a logical TRUE if valid string, FALSE otherwise
}
\description{
\code{ValidatorString} Class for validating strings
}
\details{
This class provide a methods for validating character strings
}
\section{Methods}{

\describe{
 \item{\code{new()}}{Creates an object of ValidatorString class}
 \item{\code{validate(object)}}{Method for validating string}
}
}

\seealso{
Other Validation Classes: \code{\link{Authorizer}},
  \code{\link{Validator0}}, \code{\link{ValidatorClass}},
  \code{\link{ValidatorDate}},
  \code{\link{ValidatorExists}},
  \code{\link{ValidatorLogical}},
  \code{\link{ValidatorName}},
  \code{\link{ValidatorNoSpaces}},
  \code{\link{ValidatorNotEmpty}},
  \code{\link{ValidatorNumeric}},
  \code{\link{ValidatorPath}}, \code{\link{ValidatorUrl}},
  \code{\link{Validator}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
