% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VUpdate.R
\docType{class}
\name{VCurator}
\alias{VCurator}
\title{VCurator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VCurator
}
\arguments{
\item{current}{The object in its current state}

\item{prior}{The object restored from a prior state.}
}
\description{
\code{VCurator} Visitor class responsible for restoring an object to a prior stete
}
\details{
\strong{VCurator Methods:}
The VCurator methods are as follows:
 \itemize{
  \item{\code{nlpStudio(current, prior)}}{Method for restoring the NLPStudio object a prior state.}
  \item{\code{lab(current, prior)}}{Method for restoring a Lab object to a prior state.}
  \item{\code{documentCollection(current, prior)}}{Method for restoring a DocumentCollection object to a prior state.}
  \item{\code{documentText(current, prior)}}{Method for updating a DocumentText object to a prior state.}
  \item{\code{documentCsv(current, prior)}}{Method for updating a DocumentCsv object to a prior state.}
  \item{\code{documentRdata(current, prior)}}{Method for updating a DocumentRdata object to a prior state.}
  \item{\code{documentXlsx(current, prior)}}{Method for updating a DocumentXlsx object to a prior state.}
}
}
\seealso{
Other State Classes: \code{\link{StateManager}},
  \code{\link{StateServer}}, \code{\link{State}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
