% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisitorGetObject.R
\docType{class}
\name{VisitorGetObject}
\alias{VisitorGetObject}
\title{VisitorGetObject}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VisitorGetObject
}
\arguments{
\item{lab}{An object of the Lab class.}

\item{collection}{An object of the DocumentCollection class.}

\item{document}{An object of the Document class.}
}
\value{
object A list containing the object meta data as well as the object itself. The format is as follows:
\describe{
 \item{metaData}{The meta data for the object in list and data frame format.}
  \describe{
   \item{metaDataList}{Meta data in list format.}
   \item{metaDataDf}{Meta data in data frame format.}
   }
 \item{object}{The actual object }
}
}
\description{
\code{VisitorGetObject} Class for obtaining an object and its meta data.
}
\details{
\strong{Visitor Family of Classes Overview:}

The Visitor family of classes is an implementation of the visitor
pattern documented in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). This pattern allows operations
to be defined without changing the clases of the elemnts upon which
it operates.

\strong{Visitor Family of Classes Participants:}
\itemize{
 \item Visitor0: This component class specifies an abstract interface
 for all concrete visitors
 \item VisitorGetObject: This visitor class responsible for retrieving objects and meta data on the objects from concrete elements.
 \item VisitorPrintObject: Visitor class responsible for printing object meta data to the console.
 }

\strong{VisitorGetObject Class Collaborators:}
The collaborators of the VisitorGetObject class are:
\itemize{
 \item Lab: Class responsible for managing collections of documents and analysis and processing workflows.
 \item DocumentCollection: Composite class containing documents.
 \item Documents: Class where each object represents a single document.
 }

\strong{VisitorGetObject Class Methods:}
The methods areas follows:
\itemize{
 \item{\code{visitLab(lab)}}{Base method for visiting a object of the Lab class.}
 \item{\code{visitCollection(collection)}}{Base method visiting a object of the DocumentCollection class.}
 \item{\code{visitDocument(document)}}{Base method visiting a object of the Document class.}
}
}
\seealso{
Other Visitor classes: \code{\link{Visitor0}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
