% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VString.R
\docType{class}
\name{VString}
\alias{VString}
\title{VString}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VString
}
\arguments{
\item{value}{Character string containing the value to be validated}
}
\value{
A logical TRUE if class is equal to expected value, FALSE otherwise.
}
\description{
\code{VString} Class for validating strings
}
\details{
This class provide a methods for validating character strings.
}
\seealso{
Other Validation Classes: \code{\link{Validator}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
