% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VValidatorGetObject.R
\docType{class}
\name{VValidatorGetObject}
\alias{VValidatorGetObject}
\title{VValidatorGetObject}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VValidatorGetObject
}
\arguments{
\item{object}{The object for which the getObject method is invoked.}

\item{requester}{The object invoking the getObject method.}
}
\description{
\code{VValidatorGetObject} Visitor class responsible for validating requests
for object information.  Only permits objects of the State class to invoke
getObject methods.
}
\details{
\strong{VValidatorGetObject Methods:}
The VValidatorGetObject methods are as follows:
 \itemize{
  \item{\code{nlpStudio(object, ...)}}{Method for validating the instantiation of the NLPStudio object}
  \item{\code{lab(object, ...)}}{Method for validating the instantiation of the Lab object}
  \item{\code{documentCollection(object, ...)}}{Method for validating the instantiation of the DocumentCollection object.}
  \item{\code{documentText(object, ...)}}{Method for validating the instantiation of the DocumentText object.}
  \item{\code{documentCsv(object, ...)}}{Method for validating the instantiation of the DocumentCsv object.}
  \item{\code{documentRdata(object, ...)}}{Method for validating the instantiation of the DocumentRdata object.}
  \item{\code{documentXlsx(object, ...)}}{Method for validating the instantiation of the DocumentXlsx object.}
}
}
\seealso{
Other Validation Visitor Classes: \code{\link{VValidatorAddChild}},
  \code{\link{VValidatorInit}},
  \code{\link{VValidatorParent}},
  \code{\link{VValidatorRestore}}, \code{\link{VValidator}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
