% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VVAttach.R
\docType{class}
\name{VVAttach}
\alias{VVAttach}
\title{VVAttach}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
visitor <- VVAttach$new()
}
\arguments{
\item{object}{The parent object}
}
\value{
status List containing:
\itemize{
 \item status Logical TRUE, if validation was passed, FALSE otherwise
 \item msg Character string describing the error, if status is FALSE
}
}
\description{
\code{VVAttach} Visitor class responsible for validating attach operations.
}
\details{
Class contains the methods for validating attach requests  originating from
'parent' classes.
}
\section{Core Methods}{

The VVAttach methods are as follows:
 \itemize{
  \item{\code{nlpStudio(object)}}{Method for validating the attach/detach methods parameters of the NLPStudio object.}
  \item{\code{corpus(object)}}{Method for validating the attach/detach methods parameters of the Corpus object}
  \item{\code{document(object)}}{Method for validating the attach/detach methods parameters of the Document object}
}
}

\examples{
visitor <- VVAttach$new()
object$accept(visitor)

}
\seealso{
Other Validation Visitor Classes: \code{\link{VVDetach}},
  \code{\link{VVGet}}, \code{\link{VVInit}},
  \code{\link{VVRead}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
