% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VVDetach.R
\docType{class}
\name{VVDetach}
\alias{VVDetach}
\title{VVDetach}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
visitor <- VVDetach$new()
}
\arguments{
\item{object}{The parent object}
}
\value{
status List containing:
\itemize{
 \item status Logical TRUE, if validation was passed, FALSE otherwise
 \item msg Character string describing the error, if status is FALSE
}
}
\description{
\code{VVDetach} Visitor class responsible for validating detach operations.
}
\details{
Class contains the methods for validating detach requests  originating from
'parent' classes.
}
\section{Core Methods}{

The VVDetach methods are as follows:
 \itemize{
  \item{\code{nlpStudio(object)}}{Method for validating the detach/detach methods parameters of the NLPStudio object.}
  \item{\code{corpus(object)}}{Method for validating the detach/detach methods parameters of the Corpus object}
  \item{\code{document(object)}}{Method for validating the detach/detach methods parameters of the Document object}
}
}

\examples{
visitor <- VVDetach$new()
object$accept(visitor)

}
\seealso{
Other Validation Visitor Classes: \code{\link{VVAttach}},
  \code{\link{VVGet}}, \code{\link{VVInit}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
