% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VVInit.R
\docType{class}
\name{VVInit}
\alias{VVInit}
\title{VVInit}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VVInit
}
\arguments{
\item{object}{The object in its current state}
}
\description{
\code{VVInit} Visitor class responsible for validating the initialization of objects.
}
\details{
\strong{VVInit Methods:}
The VVInit methods are as follows:
 \itemize{
  \item{\code{nlpStudio(object)}}{Method for validating the instantiation of the NLPStudio object}
}
}
\seealso{
Other Validation Visitor Classes: \code{\link{VVAttach}},
  \code{\link{VVDetach}}, \code{\link{VVGet}},
  \code{\link{VVRead}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
