% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VVMetadata.R
\docType{class}
\name{VVMetadata}
\alias{VVMetadata}
\title{VVMetadata}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VVMetadata
}
\arguments{
\item{object}{The object in its current state}
}
\description{
\code{VVMetadata} Visitor class responsible for validating calls to metadata methods.
}
\details{
\strong{VVMetadata Methods:}
The VVMetadata methods are as follows:
 \itemize{
  \item{\code{corpus(object)}}{Method for validating calls to metadata from Corpus objects.}
  \item{\code{document(object)}}{Method for validating calls  to metadata from Document objects.}
}
}
\seealso{
Other Validation Visitor Classes: \code{\link{VVAttach}},
  \code{\link{VVDetach}}, \code{\link{VVInit}},
  \code{\link{VVRead}}, \code{\link{VVSource}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
