% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VVRead.R
\docType{class}
\name{VVRead}
\alias{VVRead}
\title{VVRead}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
visitor <- VVRead$new()
}
\arguments{
\item{object}{The domain object}
}
\value{
status List containing:
\itemize{
 \item status Logical TRUE, if validation was passed, FALSE otherwise
 \item msg Character string describing the error, if status is FALSE
}
}
\description{
\code{VVRead} Visitor class responsible for validating read methods
}
\details{
Validates read methods. Confirms validity of file paths and file types.
}
\section{Core Methods}{

The VVRead methods are as follows:
 \itemize{
  \item{\code{text(object)}}{Method for validating the read method for Text objects.}
}
}

\examples{
visitor <- VVRead$new()
object$accept(visitor)

}
\seealso{
Other Validation Visitor Classes: \code{\link{VVAttach}},
  \code{\link{VVDetach}}, \code{\link{VVGet}},
  \code{\link{VVInit}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
