% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VVSource.R
\docType{class}
\name{VVSource}
\alias{VVSource}
\title{VVSource}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VVSource
}
\arguments{
\item{object}{The object in its current state}
}
\description{
\code{VVSource} Visitor class responsible for validating the Corpus source parameters.
}
\details{
\strong{VVSource Methods:}
The VVSource methods are as follows:
 \itemize{
  \item{\code{csourceVector(object)}}{Method for validating the parameters.}
  \item{\code{csourceDir(object)}}{Method for validating the parameters.}
  \item{\code{csourceQuanteda(object)}}{Method for validating the parameters.}
  \item{\code{csourceTM(object)}}{Method for validating the parameters.}
  \item{\code{csourceKorpus(object)}}{Method for validating the parameters.}
  \item{\code{csourceQdap(object)}}{Method for validating the parameters.}
}
}
\seealso{
Other Validation Visitor Classes: \code{\link{VVAttach}},
  \code{\link{VVDetach}}, \code{\link{VVGet}},
  \code{\link{VVInit}}, \code{\link{VVRead}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
