% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteCorpus.R
\docType{class}
\name{WriteCorpus}
\alias{WriteCorpus}
\title{WriteCorpus}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
WriteCorpus
}
\arguments{
\item{x}{Corpus or CVSet Object}

\item{corpusStudio}{CorpusStudio object}

\item{path}{Character string the directory or file path
the object.}
}
\value{
Token object
}
\description{
\code{WriteCorpus} Encapsulates the command to write a Corpus or CVSet to file.
}
\seealso{
Other FileStudio Command Classes: \code{\link{CopyFileSet}},
  \code{\link{MoveFileSet}}, \code{\link{ReadCorpus}},
  \code{\link{RepairFileSet}}, \code{\link{SampleCorpus}},
  \code{\link{Source}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\concept{FileStudio Command Classes}
\keyword{datasets}
