% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clone.R
\docType{class}
\name{Clone}
\alias{Clone}
\title{Clone}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Clone
}
\arguments{
\item{x}{Object to be Cloned.}

\item{name}{Character string containing the name to be assigned to the
replicant.}
}
\value{
A Replicant of the input object
}
\description{
\code{Clone0} Abstract class for the Clone family of classes.
}
\details{
Abstract class that defines the interface and common methods for the
Clone family of classes.
}
\section{Methods}{

\describe{
 \item{\code{new(x, name = NULL)}}{Instantiates the factory.}
 \item{\code{execute()}}{Returns the replicant object.}
}
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
