% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CloneCorpus.R
\docType{class}
\name{CloneCorpus}
\alias{CloneCorpus}
\title{CloneCorpus}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CloneCorpus
}
\arguments{
\item{x}{Object to be cloned. Currently supports Corpus and Document objects.}

\item{name}{Character string containing the name to be assigned to the
replicant.}
}
\value{
A Replicant of the input object
}
\description{
\code{CloneCorpus} Class responsible for cloning Corpus and Document objects.
}
\details{
Class responsible for cloning Corpus and Document objects.
}
\section{Methods}{

\describe{
 \item{\code{new(x, name = NULL)}}{Instantiates the factory.}
 \item{\code{execute()}}{Returns the replicant object.}
}
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
