% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection0.R
\docType{class}
\name{Collection0}
\alias{Collection0}
\title{Collection0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Collection0
}
\description{
\code{Collection0} Abstract class for all Collection classes
}
\details{
This abstract class defines members and methods common to the Corpus and
Document collection / composite objects.
}
\section{Collection0 methods}{

 \itemize{
  \item{\code{new(name = NULL, purpose = NULL)}}{Not implemented.}
  \item{\code{getDocument(key, value)}}{Returns the document or documents that
  match the key/value pair(s). }
  \item{\code{addDocument(x)}}{Adds a document to the Collection object.}
  \item{\code{removeDocument(x)}}{Removes a document from the Collection
  object.}
  \item{\code{summary(x)}}{Summarizes a Collection document.}
 }

 @param x Object or list of objects to be attached
 @param key Character string or vector of strings indicating the metadata
 variable or variables used for matching
 @param value Character string or vector of strings indicating the
 metadata value associated with the key(s) parameter.
}

\seealso{
Other Document Classes: \code{\link{Corpus}},
  \code{\link{Document0}}, \code{\link{Document}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
