% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConverterTM.R
\docType{class}
\name{ConverterTM}
\alias{ConverterTM}
\title{ConverterTM}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ConverterTM$new()$convert(x)
}
\arguments{
\item{x}{Object to be converted}
}
\description{
\code{ConverterTM} Converts NLPStudio Corpus objects to and from tm corpus objects.
}
\seealso{
Other Converter Classes: \code{\link{ConverterQuanteda}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
