% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Corpus.R
\docType{class}
\name{Corpus}
\alias{Corpus}
\title{Corpus}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
skiReport <- Corpus$new(name = "skiReport", purpose = 'Train')
}
\arguments{
\item{name}{Character string containing the name for the Corpus object.}

\item{purpose}{Character string used to indicate how the document will be used, e.g. 'train', 'test'.}

\item{note}{Character string containing a comment associated with a call to the
text method. The texts of the note variable are written to the Corpuss
log. This is used to track changes to the text, perhaps made during preprocessing.}
}
\value{
Corpus object, containing the Corpus text, the metadata and
the methods to manage both.
}
\description{
\code{Corpus} Entity class for text documents.
}
\details{
Entity class for text documents with methods for adding text from character
vectors.
}
\section{Core Methods}{

 \itemize{
  \item{\code{new(name = NULL)}}{Initializes an object of the Corpus class.}
  \item{\code{text(x, note = NULL)}}{Method for obtaining/adding/updating text. If no
  parameters are presented, the current text is returned.  Otherwise, the text
  is updated with the texts of the character vector 'x'. Sentence, word, token, type,
  sentence and word length statistics are also computed and the metadata is updated
  accordingly.}
  \item{\code{summary()}}{Summarizes the Corpus object.}
 }
}

\examples{
report <- c("SAN FRANCISCO  — She was snowboarding with her boyfriend when ",
          "she heard someone scream 'Avalanche!'",
          "Then John, 39, saw 'a cloud of snow coming down.'")
avalanche <- Corpus$new(name = 'avalanche', purpose = 'raw')
avalance$content <- report
key <- c('genre', 'author', 'year')
value <- c('weather', 'chris jones', 2018)
avalanche$meta$setDescriptive(key = key value = value)

}
\seealso{
Other Document Classes: \code{\link{Document}},
  \code{\link{Primitive0}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\concept{Document Classes}
\keyword{datasets}
