% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Corpus.R
\docType{class}
\name{Corpus}
\alias{Corpus}
\title{Corpus}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
myCorpus <- Corpus$new(name = "machineLearning", content = mlAdvances)
}
\arguments{
\item{name}{Optional character string containing the name of the object.}

\item{event}{Character string indicating a class associated with an event to be posted to the object's log.}

\item{x}{A Document class object, or a list thereof, to be attached to the Corpus object}

\item{key}{Character string or character vector containing the key or
keys associated a metadata field or fields.}

\item{value}{Character string or character vector, equal in length to
the key vector, containing the metadata value or values.}
}
\value{
Corpus object.
}
\description{
\code{Corpus} Class for creating, managing, reading and writing Corpus objects.
}
\details{
Corpus objects are collections of Document objects and the primary level
of document aggregation at which document processing, feature engineering,
selection, analysis and modeling occurs.
}
\section{Core Methods}{

 \itemize{
  \item{\code{new()}}{Initializes an object of the Corpus class.}
 }
}

\section{Metadata Method}{

 \itemize{
  \item{\code{meta(key = NULL, value = NULL)}}{Provides facility for
  managing an object's metadata represented as key/value pairs in
  a list format. Metadata may be indicated via single key and
  value character strings or by pairs of character vectors. If
  no parameters are passed to the method, then it returns the
  current metadata, as well as system generated application
  and system metadata, in data frame format. If a character
  string or vector is passed via the key parameter, the value
  or values associated with the keys are returned.}
  \item{\code{log(cls = class(self)[1], event = NULL)}}{Class for posting to and
  retrieving an object's log. If the event parameter is provided,
  the event is posted to the object's log and the log is returned
  invisibly.  If the event parameter is NULL, the existing log
  is returned.}
}
}

\examples{

# Instantiate
corpus <- Corpus$new(name = 'corpus')
corpus$summary()

# Attach / Detach Document objects

## Create Text objects
blogsTxt <- Text$new(name = 'blogs (raw)', x = "./data/en_US.blogs.txt")
newsTxt <- Text$new(name = 'news (raw)', x = "./data/en_US.news.txt")
twitterTxt <- Text$new(name = 'twitter (raw)', x = "./data/en_US.twitter.txt")

## Create Document objects and attach Texts
blogsDoc <- TextDocument$new()$attach(blogsTxt)
newsDoc <- TextDocument$new()$attach(newsTxt)
twitterDoc <- TextDocument$new()$attach(twitterTxt)

## Attach Document objects to Corpus
corpus$attach(x = list(blogsDoc, newsDoc, twitterDoc))

## Detach Document object from Corpus
corpus$detach(key = 'name', value = 'blogsDoc')

}
\seealso{
Other Core Classes: \code{\link{TextDocument}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
