% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_corpus.R
\name{create_corpus}
\alias{create_corpus}
\title{Build a SEC EDGAR Corpus}
\usage{
create_corpus(
  df_container,
  form_type = c("10-K"),
  to_disk = FALSE,
  path_out = NULL
)
}
\arguments{
\item{df_container}{A list of data.table containing the raw textual documents
as built by \code{\link{from_json_to_df}}.}

\item{form_type}{Character specifying the SEC form type that is contained in \code{df_container}.
Default to \code{10-K}.}

\item{to_disk}{Logical declaring whether to save the corpora to disk or not (see 'Details').
Default is \code{FALSE}.}

\item{path_out}{Character specifying the directory where to save the corpora by fiscal year.}
}
\value{
Either \code{NULL} or a quanteda corpus object.
}
\description{
Build a quanteda corpus from the data.table container.
}
\details{
When \code{to_disk = FALSE}, the function returns a, potentially, much larger object in
memory in the class \code{\link[quanteda]{corpus}}. In this case, be mindful of RAM consumption.
If \code{to_disk = TRUE}, the function returns \code{NULL} and saves to disk
the objects at the location specified by \code{path_out}. The execution time will eventually be
longer due to the usual I/O hiccups.
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}
}
