% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_corpus.R
\name{create_corpus}
\alias{create_corpus}
\title{Build a SEC EDGAR Corpus}
\usage{
create_corpus(df_container)
}
\arguments{
\item{df_container}{A \code{data.table} containing the raw textual documents
as built by \code{\link{from_json_to_df}}.}
}
\value{
A quanteda \code{\link[quanteda]{corpus}} object.
}
\description{
Build a quanteda corpus from the data.table container.
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}
}
