% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\docType{class}
\name{CV}
\alias{CV}
\title{CV}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CV
}
\arguments{
\item{name}{Character string containing the name for the CV object.}
}
\value{
CV object, containing the CVSets.
}
\description{
\code{CV} Class containing cross-validation or CVSet objects.

@param x The principal object
 @param key Character string or a vector of strings containing a key or
 keys for metadata key/value pair(s).
 @param value Character or numeric, or a vector thereof, containing a value
 or values for metadata key/value pair(s).
 @param format Character string used in the metadata method. Indicates
 whether to return the metadata as  list of lists, or a data.frame.This method
 returns metadata in a list if no key/value pair is provided. The list
 may contain other lists or data frames associated with the type of
 metadata. This parameter specifies whether to return the metadata
 as a list of lists, or a list of data frames. Valid values are:
 c('list', 'data.frame')
}
\details{
Class containing cross-validation or CVSet objects.
}
\section{Core Methods}{

 \itemize{
  \item{\code{new(name = NULL)}}{Initializes an object of the CV class.}
  \item{\code{summary()}}{Summarizes the CV object.}
 }
}

\seealso{
Other Cross Validation Classes: \code{\link{CVFactory0}},
  \code{\link{CVFactoryHoldOut}},
  \code{\link{CVFactoryKFold}}, \code{\link{CVSet}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
