% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVKFold.R
\docType{class}
\name{CVKFold}
\alias{CVKFold}
\title{CVKFold}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CVKFold
}
\arguments{
\item{x}{Corpus object for which the CVSetKFold object is to be created.}

\item{unit}{Character string indicating whether the unit for random sampling is
the vector or sentence. If the latter, the documents will be tokenized into sentences
prior to sampling.}

\item{stratify}{Logical. If TRUE, sampling will be conducted at the individual
document level. Otherwise, the Document objects will be combined into a single
Document object prior to sampling.}

\item{proportions}{Numeric vector of length 2 or 3, indicating the proportions to allocate
to the training (validation) and test sets.  If the length of the vector is 2, a
training and test set will be created in accordance with the proportions indicated. If
the length is 3, a validation set will be created in addition to the training and
test sets.}

\item{seed}{}
}
\description{
\code{CVKFold} Creates a hold out CVSet object.
}
\details{
Class responsible for creating hold out cross-validation sets or CVSetKFold objects.
CVSetKFold objects contain a training, test and an optional validation test set.
}
\section{Methods}{

 \itemize{
  \item{\code{new(X, name = NULL)}}{Initializes an object of the CVKFold class.}
  \item{\code{execute()}}{Executes the process of sourcing the Corpus object.}
 }
}

\seealso{
Other Cross Validation Classes: \code{\link{CVHoldOut}},
  \code{\link{CVSet}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
