% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVSet.R
\docType{class}
\name{CVSet}
\alias{CVSet}
\title{CVSet}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CVSet
}
\arguments{
\item{name}{Character string containing the name for the CVSet object.}

\item{purpose}{Character string used to indicate how the document will be used, e.g. 'train', 'test'.}

\item{note}{Character string containing a comment associated with a call to the
text method. The texts of the note variable are written to the CVSets
log. This is used to track changes to the text, perhaps made during preprocessing.}
}
\value{
CVSet object, containing the CVSet text, the metadata and
the methods to manage both.
}
\description{
\code{CVSet} Class representing a cross-validation set.

@param x The principal object
 @param key Character string or a vector of strings containing a key or
 keys for metadata key/value pair(s).
 @param value Character or numeric, or a vector thereof, containing a value
 or values for metadata key/value pair(s).
 @param format Character string used in the metadata method. Indicates
 whether to return the metadata as  list of lists, or a data.frame.This method
 returns metadata in a list if no key/value pair is provided. The list
 may contain other lists or data frames associated with the type of
 metadata. This parameter specifies whether to return the metadata
 as a list of lists, or a list of data frames. Valid values are:
 c('list', 'data.frame')
}
\details{
Class representing cross-validation sets. A cross-validation set may be
a training, validation, test, or k-fold set of Document objects.
}
\section{Methods}{

 \itemize{
  \item{\code{new(name = NULL)}}{Initializes an object of the CVSet class.}
  \item{\code{text(x, note = NULL)}}{Method for obtaining/adding/updating text. If no
  parameters are presented, the current text is returned.  Otherwise, the text
  is updated with the texts of the character vector 'x'. Sentence, word, token, type,
  sentence and word length statistics are also computed and the metadata is updated
  accordingly.}
  \item{\code{summary()}}{Summarizes the CVSet object.}
 }
}

\seealso{
Other Cross Validation Classes: \code{\link{CVHoldOut}},
  \code{\link{CVKFold}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
