% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{data_dictionary_Cannon_Ling_Wang_Watanabe}
\alias{data_dictionary_Cannon_Ling_Wang_Watanabe}
\title{Cannon, Ling, Wang, and Watanabe CSR Dictionary}
\format{
An object of class \code{dictionary2} of length 4.
}
\source{
\url{https://www.tandfonline.com/doi/epdf/10.1080/09638180.2019.1670223?src=getftr}
}
\usage{
data_dictionary_Cannon_Ling_Wang_Watanabe
}
\description{
A \code{\link[quanteda]{dictionary}} containing the
the 2020 version of the Cannon, Ling, Wang, and Watanabe Corporate Social Responsbility Word List.
}
\details{
The dictionary contains four categories of unigrams and bigrams:
\itemize{
\item \emph{Philantropy}
\item \emph{Business Practice}
\item \emph{Product}
\item \emph{General}
}
}
\references{
Cannon, J. N., Ling, Z., Wang, Q., & Watanabe, O. V. (2020).
10-K disclosure of corporate social responsibility and firms’ competitive advantages.
\emph{European Accounting Review}, 29(1), 85-113.
}
\keyword{data}
