% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{data_dictionary_LoughranMcDonald_Complexity}
\alias{data_dictionary_LoughranMcDonald_Complexity}
\title{Loughran and McDonald Firm Complexity Dictionary}
\format{
An object of class \code{dictionary2} of length 1.
}
\source{
\url{https://sraf.nd.edu/complexity/}

\url{https://www.cambridge.org/core/journals/journal-of-financial-and-quantitative-analysis/article/measuring-firm-complexity/D737FD0A697AF699C5AADD62842ACAB8}
}
\usage{
data_dictionary_LoughranMcDonald_Complexity
}
\description{
A \link[quanteda:dictionary]{quanteda::dictionary} containing the 2024 Loughran and McDonald firm complexity
word list. The object has a single top-level entry, \code{Firm_Complexity}, with
53 terms for complexity-oriented text analysis.
}
\references{
Loughran, T. & McDonald, B. (2024).
\href{https://www.cambridge.org/core/journals/journal-of-financial-and-quantitative-analysis/article/measuring-firm-complexity/D737FD0A697AF699C5AADD62842ACAB8}{Measuring Firm Complexity},
\emph{Journal of Financial and Quantitative Analysis},
2023:1-28.
}
\keyword{data}
