% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{data_dictionary_SDG}
\alias{data_dictionary_SDG}
\title{U.N. Sustainable Development Goals (SDG) Mapping Dictionary}
\format{
An object of class \code{dictionary2} of length 16.
}
\source{
\url{https://www.researchsquare.com/article/rs-2544385/v2}

\url{https://www.sdgmapping.auckland.ac.nz/}
}
\usage{
data_dictionary_SDG
}
\description{
A \link[quanteda:dictionary]{quanteda::dictionary} containing the
key terms to identify SDGs withing research papers as mapped by Wang et al. (2023).
}
\details{
The Auckland Approach, an enhanced method for mapping research publications to the Sustainable
Development Goals (SDGs), utilizes advanced text-mining techniques and n-gram analyses to
refine and expand keyword lists derived from publication metadata. This approach, built on
foundational work by Elsevier, Sustainable Development Solutions Network, and the
United Nations (UN), improves the precision and coverage of SDG
mapping by incorporating both globally standardized and locally relevant keywords.
The methodology ensures a more comprehensive, contextualized, and data-driven mapping process,
effectively tailoring global SDG themes to specific local research needs and narratives.
The dictionary contains 16 categories.
}
\references{
Wang, W., Kang, W., & Mu, J. (2023). Mapping research to the sustainable development goals
\emph{Working Paper} available \href{https://assets.researchsquare.com/files/rs-2544385/v3/de08997f5c48f22a37cc4f93.pdf?c=1678389566}{here}.
}
\keyword{data}
