% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataStudio.R
\docType{class}
\name{DataStudio}
\alias{DataStudio}
\title{DataStudio}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataStudio
}
\arguments{
\item{object}{The object to be processed.}

\item{queue}{The job queue containing text processing commands.}
}
\description{
\code{DataStudio} Class for creating data from text
}
\details{
The NLPStudio pipeline starts with the TextStudio, a collection of classes
that performs preprocessing on text. This class is the second step in the
pipeline and is responsible for converting text to data for downstream
analysis.
}
\section{DataStudio methods}{

\strong{Core Methods:}
 \itemize{
  \item{\code{new()}}{Method for instantiating a DataStudio.}
  \item{\code{token(x, what = c("char", "word", "sentence"))}}{Method for tokenizing an object.}
  \item{\code{stem()}}{Method that for removing a command from the queue.}
  \item{\code{ngram()}}{Method that executes the job queue. }
  \item{\code{dfm()}}{Method that returns the object following execution of the job queue. }
  \item{\code{pos()}}{Method that returns the object following execution of the job queue. }
 }

\strong{Other Methods:}
 \itemize{
  \item{\code{accept(visitor)}}{Accepts a visitor object. Not implemented for this abstract class}
  \item{\code{logIt(level = 'Info')}}{Logs events relating to the DataStudio.}
 }
}

\section{Parameters}{

}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
