% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Director.R
\docType{class}
\name{Director}
\alias{Director}
\title{Director}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Director
}
\arguments{
\item{cmd}{The class encapsulating a particular command.}
}
\description{
\code{Director} Class responsible for executing an NLP Pipeline
}
\section{Director methods}{

 \itemize{
  \item{\code{new()}}{Method for instantiating a Director.}
  \item{\code{addCommand()}}{Method that adds a text processing command to the queue. }
  \item{\code{removeCommand()}}{Method that removes a command from the queue.}
  \item{\code{validate()}}{Method that validates inputs/outputs at each stage of pipeline}
  \item{\code{execute()}}{Method that executes the job queue. }
  \item{\code{printTasks()}}{Prints task list. }
  \item{\code{getResult()}}{Method that returns the object following execution of the job queue. }
 }
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
