% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document.R
\docType{class}
\name{Document}
\alias{Document}
\title{Document}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
skiReport <- Document$new(name = "skiReport", purpose = 'Train')
}
\arguments{
\item{name}{Character string containing the name for the Document object.}

\item{purpose}{Character string used to indicate how the document will be used, e.g. 'train', 'test'.}

\item{note}{Character string containing a comment associated with a call to the
text method. The texts of the note variable are written to the Documents
log. This is used to track changes to the text, perhaps made during preprocessing.}
}
\value{
Document object, containing the Document text, the metadata and
the methods to manage both.
}
\description{
\code{Document} Entity class for text documents.

@param x The principal object
 @param key Character string or a vector of strings containing a key or
 keys for metadata key/value pair(s).
 @param value Character or numeric, or a vector thereof, containing a value
 or values for metadata key/value pair(s).
 @param format Character string used in the metadata method. Indicates
 whether to return the metadata as  list of lists, or a data.frame.This method
 returns metadata in a list if no key/value pair is provided. The list
 may contain other lists or data frames associated with the type of
 metadata. This parameter specifies whether to return the metadata
 as a list of lists, or a list of data frames. Valid values are:
 c('list', 'data.frame')
}
\details{
Entity class for text documents with methods for adding text from character
vectors.
}
\section{Core Methods}{

 \itemize{
  \item{\code{new(name = NULL, purpose = NULL)}}{Initializes an object of the Document class.}
  \item{\code{text(x, note = NULL)}}{Method for obtaining/adding/updating text. If no
  parameters are presented, the current text is returned.  Otherwise, the text
  is updated with the texts of the character vector 'x'. Sentence, word, token, type,
  sentence and word length statistics are also computed and the metadata is updated
  accordingly.}
  \item{\code{overview()}}{Provides a subset of the metadata in a one-row data.frame format.
  This is used by the parent class's summary method.  }
 }
}

\examples{
report <- c("SAN FRANCISCO  — She was snowboarding with her boyfriend when ",
          "she heard someone scream 'Avalanche!'",
          "Then John, 39, saw 'a cloud of snow coming down.'")
avalanche <- Document$new(name = 'avalanche', purpose = 'raw')
avalance$content <- report
key <- c('genre', 'author', 'year')
value <- c('weather', 'chris jones', 2018)
avalanche$meta$setDescriptive(key = key value = value)

}
\seealso{
Other Document Classes: \code{\link{Corpus}},
  \code{\link{Primitive0}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
