% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document.R
\docType{class}
\name{Document}
\alias{Document}
\title{Document}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
myDocument <- Document$new(name = "skiReport", x = avalanche)
}
\arguments{
\item{x}{Character vector containing Document}

\item{key}{Character string or character vector containing the key or
keys associated a metadata field or fields.}

\item{value}{Character string or character vector, equal in length to
the key vector, containing the metadata value or values.}
}
\value{
Document object, containing the Document content, the metadata and
the methods to manage both.
}
\description{
\code{Document} Class for creating, managing, reading and writing
Document content and metadata.
}
\details{
Class contains Document content and metadata for the Document object.
The object is instantiated with a name and character
vectors containing the Document text. Once the object is created, users may
add, change and remove metadata via key / value pairs. IO
methods support reading and writing the Documents in a variety
of formats using the \code{\link[NLPStudio]{IOFactory}} factory
class.
}
\section{Core Methods}{

 \itemize{
  \item{\code{new()}}{Initializes an object of the Document class.}
 }
}

\section{Metadata Method}{

 \itemize{
  \item{\code{meta(key = NULL, value = NULL)}}{Provides facility for
  managing an object's metadata represented as key/value pairs in
  a list format. Metadata may be indicated via single key and
  value character strings or by pairs of character vectors. If
  no parameters are passed to the method, then it returns the
  current metadata, as well as system generated application
  and system metadata, in data frame format. If a character
  string or vector is passed via the key parameter, the value
  or values associated with the keys are returned.}
}
}

\examples{
avalanche <- c("SAN FRANCISCO  — She was snowboarding with her boyfriend when ",
          "she heard someone scream 'Avalanche!'",
          "Then John, 39, saw 'a cloud of snow coming down.'")
avalancheDocument <- Document$new(name = 'skiReport', content = avalanche)

avalancheDocument <- myDocument$meta(key = c("author", "editor", "year"),
                      value = c("Dorfmeister", "Huffington", "2018"))

avalancheDocument$meta()

news <- "./data/en_US.news.txt"
newsDocument <- Document$new(name = 'news', x = news)

}
\seealso{
Other Core Classes: \code{\link{Corpus}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
