% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document.R
\docType{class}
\name{Document}
\alias{Document}
\title{Document}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
myDocument <- Document$new(name = "machineLearning", content = mlAdvances)
}
\arguments{
\item{name}{Optional character string containing the name of the object.}

\item{x}{A Text, Data or Analyis object to be attached to the Document object.}

\item{key}{Character string or character vector containing the key or
keys associated a metadata field or fields.}

\item{value}{Character string or character vector, equal in length to
the key vector, containing the metadata value or values.}
}
\value{
Document object.
}
\description{
\code{Document} Class for creating, managing, reading and writing Documents.
}
\details{
Document objects are abstractions of information resources comprised of
one or several versions of a single resource or text, each encapsulated in
a Text object, as well as Data and information or Analysis objects. This broad
definition allows one to "attach" any number of raw, preprocessed,
training, or cross-validation Text objects to a single Document. Data
objects are the transformations of the original or preprocessed Text
object into forms which can be analyzed and studied. A Document's Analysis
objects include analyses such as readibility, collocation, document similarity,
semantic analysis and so on. This class provides allows users to
encapsulate raw text, its conversions into data, and all associated analyses
into a single document unit for processing, experimentation, modeling,
testing, inference and prediction.
}
\section{Core Methods}{

 \itemize{
  \item{\code{new()}}{Initializes an object of the Document class.}
 }
}

\section{Metadata Method}{

 \itemize{
  \item{\code{meta(key = NULL, value = NULL)}}{Provides facility for
  managing an object's metadata represented as key/value pairs in
  a list format. Metadata may be indicated via single key and
  value character strings or by pairs of character vectors. If
  no parameters are passed to the method, then it returns the
  current metadata, as well as system generated application
  and system metadata, in data frame format. If a character
  string or vector is passed via the key parameter, the value
  or values associated with the keys are returned.}
}
}

\examples{

# Instantiate
blogsDoc <- Document$new(name = "Blogs")

# Attach Text object
## Create Text object
blogsTxt <- Text$new(name = 'blogs (raw)')$read(path = "./data/en_US.blogs.txt")

## Attach to Document object
blogsDoc$attach(blogsTxt)

## Detach
blogsDoc$detach(key ='name', value = 'blogs (raw)')

}
\seealso{
Other Core Classes: \code{\link{Corpus}},
  \code{\link{Text}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
