% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document.R
\docType{class}
\name{Document}
\alias{Document}
\title{Document}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
myDocument <- Document$new(name = "machineLearning", content = mlAdvances)
}
\arguments{
\item{x}{A Text, Data or Analyis object, or the name thereto. Used in the detach method.}

\item{cls}{Class of the object to be detached. Required for detaching objects when the
x parameter contains the name of the object.}

\item{path}{Character string designated the relative path to a file.}

\item{io}{Object of the IOStrategy class. This is an optional parameter.
If io is not specified, the default IOStrategy for the file type will
be used. See ?IOStrategy for further assistance.}

\item{key}{Character string or character vector containing the key or
keys associated a metadata field or fields.}

\item{value}{Character string or character vector, equal in length to
the key vector, containing the metadata value or values.}
}
\value{
Document object, containing related Text, Data, and Analysis objects.
}
\description{
\code{Document} Class for creating, managing, reading and writing Documents.
}
\details{
Document objects are abstractions of information resources comprised of
one or several versions of a single resource or text, each encapsulated in
a Text object, as well as Data and information or Analysis objects. This broad
definition allows one to "attach" any number of raw, preprocessed,
training, or cross-validation Text objects to a single Document. Data
objects are the transformations of the original or preprocessed Text
object into forms which can be analyzed and studied. A Document's Analysis
objects include analyses such as readibility, collocation, document similarity,
semantic analysis and so on. This class provides allows users to
encapsulate raw text, its conversions into data, and all associated analyses
into a single document unit for processing, experimentation, modeling,
testing, inference and prediction.
}
\section{Core Methods}{

 \itemize{
  \item{\code{new()}}{Initializes an object of the Document class.}
 }
}

\section{IO Methods}{

 \itemize{
  \item{\code{read(path, io = NULL)}}{Reads an object from a specified
  file using the default or designated IO Strategy. See IOStrategy?
  for details w.r.t. the io parameter. If the IO strategy is not
  indicated, the default io strategy for the file type is used.}
  \item{\code{write(path, io = NULL)}}{Writes an object to a specified
  file using the default or designated IO Strategy. See IOStrategy?
  for details w.r.t. the io parameter. If the IO strategy is not
  indicated, the default io strategy for the file type is used.}
 }
}

\section{Metadata Method}{

 \itemize{
  \item{\code{meta(key = NULL, value = NULL)}}{Provides facility for
  managing an object's metadata represented as key/value pairs in
  a list format. Metadata may be indicated via single key and
  value character strings or by pairs of character vectors. If
  no parameters are passed to the method, then it returns the
  current metadata, as well as system generated application
  and system metadata, in data frame format. If a character
  string or vector is passed via the key parameter, the value
  or values associated with the keys are returned.}
}
}

\examples{
rawText <- c("Capsule networks (CapsNets) are a hot new neural net
              architecture that may well have a profound impact on deep
              learning, in particular for computer vision. Wait, isn't
              computer vision pretty much solved already? Haven't we all
              seen fabulous examples of convolutional neural networks (CNNs)
              reaching super-human level in various computer vision tasks,
              such as classification, localization, object detection,
              semantic segmentation or instance segmentation.")

 raw <- Text$new(name = "capsNets", content = rawText)

 mlDoc <- Document$new(name = "capsNets")

 m1Doc <- m1Doc$attach(object = raw)

}
\seealso{
Other Core Classes: \code{\link{Corpus}},
  \code{\link{Text}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
