% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document.R
\docType{class}
\name{Document}
\alias{Document}
\title{Document}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Document
}
\description{
\code{Document} Leaf class for instantiating, printing, reading and writing non-composite, leaf documents
}
\details{
This leaf class contains the data and methods used to instantiate objects of
the Document class, to print the document meta data and to read  and write
the document. The methods are as follows:
\describe{
 \item{\code{new()}}{Method for instantiating an object of the Document class}
 \item{\code{getDocument()}}{Method for returning and printing the document object meta data.}
 \item{\code{readDocument(reader)}}{Method for reading the document using the designated reader.}
 \item{\code{writeDocument(writer)}}{Method for writing the document using the designated writer.}
}
}
\seealso{
Other Document classes: \code{\link{Document0}},
  \code{\link{DocumentCollection}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
