% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection0.R
\docType{class}
\name{Document0}
\alias{Document0}
\title{Document0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Document0
}
\description{
\code{Document0} Component class defining base methods for all documents.
}
\details{
This composite pattern component class defines members and methods common
across all the document related classes, including without limitation,
Corpus, TextDocument, TermFreq, POS, NGram, and Tokens objects.
}
\section{Document0 methods}{

 \itemize{
  \item{\code{new()}}{Method not implemented for this component class.}
  \item{\code{attach(x)}}{Method for attaching objects.}
  \item{\code{detach(key, value)}}{Method for detaching objects.}
  \item{\code{attachments(key = NULL, value = NULL)}}{Lists an object's attachments.}
 }

 @param x Object or list of objects to be attached
 @param key Character string or vector of strings indicating the metadata
 variable or variables used for matching
 @param value Character string or vector of strings indicating the
 metadata value associated with the key(s) parameter.
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
