% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document0.R
\docType{class}
\name{Document0}
\alias{Document0}
\title{Document0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Document0
}
\description{
\code{Document0} Component class defining base methods for all documents.
}
\details{
This composite pattern component class defines members and methods common
across all the document related classes, including without limitation,
Corpus, Document, TermFreq, POS, NGram, and Tokens objects.
}
\section{Document0 methods}{

 \itemize{
  \item{\code{new()}}{Method not implemented for this component class.}
  \item{\code{summary()}}{Prints a summary of a Document family class object.}
 }
}

\seealso{
Other Document Classes: \code{\link{Collection0}},
  \code{\link{Corpus}}, \code{\link{Document}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
