% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document0.R
\docType{class}
\name{Document0}
\alias{Document0}
\title{Document0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Document0
}
\description{
\code{Document0} Abstract class for concrete Document classes.
}
\details{
Abstract class that defines common methods for concrete Document classes,
such as the TextDocument, TokensDocument, NGramDocument, POSDocument,
and StemDocument classes.
}
\section{Metadata Method}{

 \itemize{
  \item{\code{meta(key = NULL, value = NULL)}}{Provides facility for
  managing an object's metadata represented as key/value pairs in
  a list format. Metadata may be indicated via single key and
  value character strings or by pairs of character vectors. If
  no parameters are passed to the method, then it returns the
  current metadata, as well as system generated application
  and system metadata, in data frame format. If a character
  string or vector is passed via the key parameter, the value
  or values associated with the keys are returned.}
  \item{\code{log(cls = class(self)[1], event = NULL)}}{Class for posting to and
  retrieving an object's log. If the event parameter is provided,
  the event is posted to the object's log and the log is returned
  invisibly.  If the event parameter is NULL, the existing log
  is returned.}
}
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
