% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document0.R
\docType{class}
\name{Document0}
\alias{Document0}
\title{Document0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Document0
}
\arguments{
\item{name}{Character string indicating the name of the document or file. Required for all objects.}

\item{desc}{Character string containing the description of the document.}

\item{parent}{An object of the Lab or DocumentCollection class that represents the parent object.}

\item{visitor}{An object from one of the visitor classes.}

\item{prior}{A Document Family Class object, deserialized from a prior state.}

\item{stateId}{Character string that uniquely identifies an object and its
state at a specific point in time.}
}
\description{
\code{Document0} Component class for Document composite in the NLPStudio.
}
\seealso{
Other Document classes: \code{\link{DocumentCollection}},
  \code{\link{DocumentCsv}}, \code{\link{DocumentRdata}},
  \code{\link{DocumentText}}, \code{\link{DocumentXlsx}},
  \code{\link{Document}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
