% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document0.R
\docType{class}
\name{Document0}
\alias{Document0}
\title{Document0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Document0
}
\arguments{
\item{name}{Character string indicating the name of the document (required)}

\item{desc}{Character string indicating the description for the document}

\item{path}{Character string containing the relative directory path to the document}

\item{documents}{A list of Document objects contained in the object of the DocumentCollection class}

\item{created}{A date time variable indicating the date / time the object was created}

\item{modified}{A date time variable indicating the date / time the object was modified}
}
\description{
\code{Document0} Abstract class for all documents in the NLPStudio
}
\details{
This abstract class is the component class within the Document0 composition
pattern. To enable common handling of both leaf and composite objects, this
object includes the methods common to all objects, plus those methods
used to manipulate a composite of objects. The methods, therefore, fall
into three categories: (1), core methods implemented by both primative
(documents) and composite objects (collections), (2) methods to manage child
objects (documents), and (3) behavior strategy setting methods used as part of
the strategy pattern to select specific behaviors for the core methods
at run time.
}
\section{Core Methods}{

The core methods are shared by both primative and composite objects and are
as follows:
\describe{
 \item{\code{new()}}{Base method for initializing an object of class Document0 or its descendants. Specific behaviors implemented in the Collection subclasses }
 \item{\code{getDocument()}}{Base method for retrieving the meta data for a document. Specific behaviors implemented in the Document and Collection subclasses }
 \item{\code{readDocument()}}{Base method for reading documents. Specific behaviors implemented in the Document and Collection subclasses }
 \item{\code{writeDocument()}}{Base method for writing documents. Specific behaviors implemented in the Document and Collection subclasses }
 #'  \item{\code{downloadDocument()}}{Method for downloading the document collection from the web. Specific behaviors implemented in the Document and Collection subclasses}
 \item{\code{unzipDocument()}}{Method for unzipping the document collection. Specific behaviors implemented in the Document and Collection subclasses}
}
}

\section{Composite Methods}{

The composite methods are used to manage the collection of documents within the composite objects or corpora within
other corpora. For instance,the training corpus may be a collection of training corpora of varying sizes. These methods include:
\describe{
 \item{\code{addDocument(document)}}{Base method for adding documents to a collection. Specific behaviors implemented in the Collection composite class}
 \item{\code{removeDocument(document)}}{Base method for removing documents from a collection. Specific behaviors implemented in the Collection composite class}
 \item{\code{searchDocuments(name)}}{Base method for searching for documents.  Specific behaviors implemented in the Collection composite class}
 \item{\code{listDocuments()}}{Base method for getting a list of documents in a collection. Specific behaviors implemented in the Collection composite class}
}
}

\section{Behavior Strategy Methods}{

These methods enable clients to define behaviors for Register and Korpus class object at runtime.
\describe{
 \item{getReader}{Method for getting the current Read strategy for the object.}
 \item{setReader}{Method for setting the current Read strategy for the object.}
 \item{getWriter}{Method for getting the current Write strategy for the object.}
 \item{setWriter}{Method for setting the current Write strategy for the object.}
}
}

\seealso{
Other Document classes: \code{\link{DocumentCollection}},
  \code{\link{Document}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
