% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentCollection.R
\docType{class}
\name{DocumentCollection}
\alias{DocumentCollection}
\title{DocumentCollection}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DocumentCollection
}
\description{
\code{DocumentCollection} Composite class for creating and interacting with documents
}
\details{
This compsite class provides data and methods for creating, interacting with,
and transforming documents and document collections. The methods fall
into three categories: (1), core methods for instantiating, printing,
reading and writing document collections. (2) methods to manage child
objects (documents) and (3) methods to set sourcing, reading and
writing behaviors. Note that document collections may include other
collections.
}
\section{Core Methods}{

\describe{
 \item{\code{initialize()}}{Method for instantiating a document collection.}
 \item{\code{getDocument()}}{Method for retrieving and printing collection meta data.}
 \item{\code{readDocument()}}{Method for reading the document collection, using the read behavior strategy set below.}
 \item{\code{writeDocument()}}{Method for writing the document collection, using the write behavior strategy set below.}
}
}

\section{Composite Methods}{

The composite methods control the document tree structure and include:
\describe{
 \item{\code{addDocument(document)}}{Method for adding a document to a collection.}
 \item{\code{searchDocuments(name)}}{Method for searching documents or collections within collections.}
 \item{\code{removeDocument(name)}}{Method for removing a document or collection from a collection.}
 \item{\code{listDocuments()}}{Method for obtaining and printing a list of documents in a collection.}
}
}

\section{Behavior Strategy Methods}{

These methods enable clients to define behaviors for Register and Korpus class object at runtime.
\describe{
 \item{getReader}{Method for getting the current Read strategy for the object.}
 \item{setReader}{Method for setting the current Read strategy for the object.}
 \item{getWriter}{Method for getting the current Write strategy for the object.}
 \item{setWriter}{Method for setting the current Write strategy for the object.}
}
}

\seealso{
Other Document classes: \code{\link{Document0}},
  \code{\link{Document}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
