% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entity.R
\docType{class}
\name{Entity}
\alias{Entity}
\title{Entity}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Entity
}
\description{
\code{Entity} Abstract including methods common to all Entity classes.
}
\details{
Class provides methods common to entity classes, such as Corpus Document, DFM,
POS, and NGram.
}
\section{Entity methods}{

 \itemize{
  \item{\code{new()}}{Not implemented for this abstract class.}
  \item{\code{summary()}}{Summarizes an object in terms of its meta data.}
 }
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
