% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileSource0.R
\docType{class}
\name{FileSource0}
\alias{FileSource0}
\title{FileSource0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileSource0
}
\arguments{
\item{x}{Character string containing the file path or URL for the source content}

\item{destination}{Character string indicating the file path for the resulting FileSet object.}
}
\description{
\code{FileSource0}  Abstract class for the FileSource family of classes.
}
\details{
Abstract strategy class which defines the methods common to the FileSource
family of classes.
}
\seealso{
Other FileSource Family of Classes: \code{\link{FileSourceCSV}},
  \code{\link{FileSourceDOCX}},
  \code{\link{FileSourceJSON}},
  \code{\link{FileSourcePDF}}, \code{\link{FileSourceTXT}},
  \code{\link{FileSourceURL}}, \code{\link{FileSourceXML}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
