% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileSourceXml.R
\docType{class}
\name{FileSourceXml}
\alias{FileSourceXml}
\title{FileSourceXml}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileSourceXml
}
\arguments{
\item{path}{Character string containing a relative path to a file.}
}
\description{
\code{FileSourceTxt} Class creates a File object from .xml file.
}
\seealso{
Other FileSource Family of Classes: \code{\link{FileSource0}},
  \code{\link{FileSourceCsv}},
  \code{\link{FileSourceDocx}},
  \code{\link{FileSourceHtml}},
  \code{\link{FileSourceJSON}}, \code{\link{FileSourceTxt}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
