% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileStudio.R
\docType{class}
\name{FileStudio}
\alias{FileStudio}
\title{FileStudio}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileStudio
}
\arguments{
\item{x}{FileSet object}

\item{path}{Character string containing a relative path to a file or
directory.  It may also contain a glob expression that resolves to one
or several files.}
}
\description{
\code{FileStudio}  Class responsible for performing operations on files.
}
\details{
Class registers files and filesets, and manipulates them, usually at raw binary files.
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
