% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileStudio.R, R/RemoveCtrlApp.R
\docType{class}
\name{FileStudio}
\alias{FileStudio}
\alias{FileStudio}
\title{FileStudio}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileStudio

FileStudio
}
\arguments{
\item{x}{The object to be processed.}

\item{queue}{The job queue containing file processing commands.}

\item{path}{Character string indicating the path to a file to be repaired.}

\item{code}{Numeric vector containing up to 32 numbers between 0 and 31,
in which each number relates to the decimal ASCII code for the control
characters to remove from the file.}
}
\description{
\code{FileStudio} Class for performing file cleaning and preprocessing

\code{FileStudio} Class for treating and repair files.
}
\details{
Class responsible for treating and performing repairs on text files
at binary level.
}
\section{TextStudio Family of Classes Overview}{

The TextStudio family of classes include four types of classes. The manager classes which
are responsible for executing one or more core class objects through the implementation
of a command class object, supported by support class objects.
}

\section{FileStudio methods}{

\strong{Core Methods:}
 \itemize{
  \item{\code{new()}}{Method for instantiating a FileStudio.}
  \item{\code{addCommand()}}{Method that adds a file processing command to the queue. }
  \item{\code{removeCommand()}}{Method that removes a command from the queue.}
  \item{\code{execute()}}{Method that executes the job queue. }
  \item{\code{getResult()}}{Method that returns the object following execution of the job queue. }
 }
}

\section{Parameters}{

}

\seealso{
Other File Classes: \code{\link{FileSet}},
  \code{\link{File}}
}
\author{
John James, \email{jjames@datasciencesalon.org}

John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
