% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_json_to_df.R
\name{from_json_to_df}
\alias{from_json_to_df}
\title{Convert JSON to data.frame}
\usage{
from_json_to_df(json_list, ncores = 1, bind = TRUE)
}
\arguments{
\item{json_list}{A list of JSON files as built by \code{\link{get_json_files}}.}

\item{ncores}{The number of cores to assign to \code{\link[parallel]{makeCluster}}. Default to 1.}

\item{bind}{Logical indicating whether to return a single \code{data.table} or a \code{list}.
Default to \code{TRUE}.}
}
\value{
Either a single \code{data.table} when \code{bind = TRUE} or a list of data.table
where each element represents a fiscal year. Each observation contains
several identification columns in addition to the document itself.
}
\description{
Convenient function to covert a container list of JSON files to data.frame structures.
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}
}
