% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_json_to_df.R
\name{from_json_to_df}
\alias{from_json_to_df}
\title{Convert JSON to data.frame}
\usage{
from_json_to_df(json_list)
}
\arguments{
\item{json_list}{A list of JSON files as built by \code{\link{get_json_files}}.}
}
\value{
I don't remember yet...check in later...
}
\description{
Convenient function to covert a container list of JSON files to data.frame structures.
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}
}
