% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_json_to_df.R
\name{from_json_to_df}
\alias{from_json_to_df}
\title{Convert JSON to data.frame}
\usage{
from_json_to_df(json_list, ncores = 1, drop_late_filers = FALSE)
}
\arguments{
\item{json_list}{A list of JSON files as built by \link{get_json_files}.}

\item{ncores}{The number of cores to assign to \link{makeCluster}. Default to 1.}

\item{drop_late_filers}{Logical for late filers removal. Default to \code{FALSE}.}
}
\value{
A single \code{data.table} containing
several identification columns in addition to the document itself.
}
\description{
Convenient function to convert a container list of JSON files to data.table.
}
\details{
This function accepts a list of JSON files as detected by \link{get_json_files} that have been downloaded
by the \emph{edgar-crawler} library. This is a different approach as that of \link{json_to_tabular} and should
be used directly on the split JSON files coming out of the crawler.

The parallel backend does not work over multiple SEC filings at once. Rather, it speeds up the
reshaping of the JSON structure within each filing by converting it to a data.table of size
1 x 20 if the filing is post 2005 and includes the Item 1A.
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}
}
