% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_json_files.R
\name{get_json_files}
\alias{get_json_files}
\title{JSON Files Container}
\usage{
get_json_files(root_path, pattern = NULL, fyear = NULL)
}
\arguments{
\item{root_path}{Character giving the path to the root folder location (see 'Details').}

\item{pattern}{An optional \link[base:regex]{base::regex}. Only file names which match the regular expression
will be returned as in \link{list.files}. Default to \code{NULL}.}

\item{fyear}{Optional numeric vector specifying the time window in years. When \code{NULL},
all the years detected in \code{root_path} are considered. Default to \code{NULL}.}
}
\value{
A list whose elements are the years. Each element contains the pointers to each JSON file within
each year.
}
\description{
Create a container list with the pointers to all the JSON files detected in the input folder.
}
\details{
The function works in conjunction with the EDGAR crawler structure. In general, this creates three
main folders as containers: (i) the original HTML files, (ii) the metadata of each filing, (iii)
the extracted items in JSON format. Each of these structures is usually, and conveniently, organized
by year. This is not an imposition but rather a strong suggestion. Ideally, one could put
all the filings into the root_path
}
