% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KN.R
\docType{class}
\name{KN}
\alias{KN}
\title{KN}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
KN
}
\arguments{
\item{x}{CVSet object containing a training and test set.}

\item{modelSize}{Numeric indicating the model size in terms of nGrams. Defaults to trigram model.}

\item{open}{Logical indicating whether the vocabulary is open
or closed. Open indicates that it is possible to encounter out of vocabulary
words in the test set. If TRUE, OOV processing will be performed on
the training set. Default is TRUE.}

\item{name}{Character string. Represents name to be assigned to the KN object.}
}
\description{
\code{KN} Class containing the Kneser Ney Language Model object.
}
\seealso{
Other Language Model Classes: \code{\link{KNStudio0}},
  \code{\link{KNStudio}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
