% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNEstimate.R
\docType{class}
\name{KNEstimate}
\alias{KNEstimate}
\title{KNEstimate}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
KNEstimate
}
\arguments{
\item{x}{Language model object}
}
\description{
\code{KNEstimate} Computes Modified Kneser Ney estimates
}
\details{
Computes Modified Kneser Ney estimates for all nGrams in the Model
}
\seealso{
Other KNStudio Classes: \code{\link{KNCounts}},
  \code{\link{KNDocument}}, \code{\link{KNStudio0}}

Other LMStudio Classes: \code{\link{KNCounts}},
  \code{\link{KNDocument}}, \code{\link{KNStudio0}},
  \code{\link{LMStudio0}}, \code{\link{MKNCounts}},
  \code{\link{MKNDocument}}, \code{\link{MKNEstimate}},
  \code{\link{MKNStudio0}}, \code{\link{MKN}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
