% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNEvaluate.R
\docType{class}
\name{KNEvaluate}
\alias{KNEvaluate}
\title{KNEvaluate}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
KNEvaluate
}
\arguments{
\item{x}{Language model object}
}
\description{
\code{KNEvaluate} Evaluates a Kneser-Ney Language Model on Test Set
}
\details{
Evaluates a Kneser-Ney Language Model on Test Set by computing
test set perplexity
}
\seealso{
Other KNStudio Classes: \code{\link{KNBuild}},
  \code{\link{KNEstimate}}

Other LMStudio Classes: \code{\link{KNBuild}},
  \code{\link{KNEstimate}}, \code{\link{LMCorpora}},
  \code{\link{LMStudio0}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
