% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNStudio0.R
\docType{class}
\name{KNStudio0}
\alias{KNStudio0}
\title{KNStudio0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
KNStudio0
}
\arguments{
\item{x}{Language model object}
}
\description{
\code{KNStudio0} Abstract class for the Kneser Ney language model state classes.
}
\details{
This abstract class defines a common interface and methods of the
Kneser Ney language model state classes.
}
\seealso{
Other KNStudio Classes: \code{\link{KNCounts}},
  \code{\link{KNDocument}}, \code{\link{KNEstimate}}

Other LMStudio Classes: \code{\link{KNCounts}},
  \code{\link{KNDocument}}, \code{\link{KNEstimate}},
  \code{\link{LMStudio0}}, \code{\link{MKNCounts}},
  \code{\link{MKNDocument}}, \code{\link{MKNEstimate}},
  \code{\link{MKNStudio0}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
