% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab.R
\docType{class}
\name{Lab}
\alias{Lab}
\title{Lab}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Lab
}
\arguments{
\item{name}{A character string containing the name of the Lab}

\item{desc}{A chararacter string containing the description of the Lab}

\item{collections}{A list containing collection objects}
}
\description{
\code{Lab} Class for managing collections of documents.
}
\details{
The primary function of this class is to maintain collections of documents.
Methods have been developed to add, and remove document collections.
Additionally, functionality is included to get/print and archive the lab.
}
\section{Methods}{

\describe{
 \item{\code{new()}}{Creates an object of Lab Class}
 \item{\code{getLab(verbose = FALSE)}}{Retrieves the meta data for the Lab, as well as a list of its document collections. If verbose is false, a data frame is returned.  If verbose is true, data frame is returned and the lab and its colllections are printed to console.}
  \item{\code{archiveLab()}}{Compresses and stores the lab in the archive directory. This method is also called prior to removing a lab from the Studio. }
  \item{\code{addDocument(name, desc)}}{Adds a document collection to the list of collections in the lab}
  \item{\code{removeDocument(name, purge = FALSE)}}{Removes a document collection from the list and removes the collection from memory of purge = TRUE}
}
}

\section{Private Fields}{

}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
