% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_tokens.R
\name{lookup_tokens}
\alias{lookup_tokens}
\title{Fast Tokens Lookup}
\usage{
lookup_tokens(x, ncores, ...)
}
\arguments{
\item{x}{A \strong{quanteda} \link{tokens} object.}

\item{ncores}{The number of \link{multisession} workers to be allocated for
the lookup.}

\item{...}{Additional arguments passed to \link{tokens_lookup}.}
}
\value{
A \link{tokens} object.
}
\description{
Lookup a \link{tokens} object in parallel via \strong{\link{future}}.
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}
}
