% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_tokens.R
\name{lookup_tokens}
\alias{lookup_tokens}
\title{Fast Tokens Lookup}
\usage{
lookup_tokens(x, ncores, ...)
}
\arguments{
\item{x}{A \code{\link[quanteda]{tokens}} object.}

\item{ncores}{The number of \code{\link[future]{multisession}} workers to be allocated for
the lookup.}

\item{...}{Additional arguments passed to \link{tokens_lookup}.}
}
\value{
A \code{\link[quanteda]{tokens}} object.
}
\description{
Lookup a \code{\link[quanteda]{tokens}} object in parallel via \strong{\link{future}}.
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}
}
