% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meta.R
\docType{class}
\name{Meta}
\alias{Meta}
\title{Meta}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Meta
}
\description{
\code{Meta} Base class for metadata objects.
}
\details{
Provides the methods for creating, managing, and reporting metadata. This
base functionality is shared by all the Metadata family classes.
}
\section{Meta methods}{

}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\concept{Metadata Classes}
\keyword{datasets}
