% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MKNCounts.R
\docType{class}
\name{MKNCounts}
\alias{MKNCounts}
\title{MKNCounts}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MKNCounts
}
\arguments{
\item{x}{Language model object}
}
\description{
\code{MKNCounts} Computes counts used to estimate Modified Kneser Ney probabilities
}
\details{
Computes counts used in the probability calculations for the Modified Kneser Ney
probabilities.
}
\seealso{
Other MKNStudio Classes: \code{\link{MKNDocument}},
  \code{\link{MKNEstimate}}, \code{\link{MKNStudio0}}

Other LMStudio Classes: \code{\link{KNCounts}},
  \code{\link{KNDocument}}, \code{\link{KNEstimate}},
  \code{\link{KNStudio0}}, \code{\link{LMStudio0}},
  \code{\link{MKNDocument}}, \code{\link{MKNEstimate}},
  \code{\link{MKNStudio0}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
