% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MKNStudio0.R
\docType{class}
\name{MKNStudio0}
\alias{MKNStudio0}
\title{MKNStudio0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MKNStudio0
}
\arguments{
\item{x}{Language model object}
}
\description{
\code{MKNStudio0} Abstract class for the Modified Kneser Ney language model state classes.
}
\details{
This abstract class defines a common interface and methods of the
Modified Kneser Ney language model state classes.
}
\seealso{
Other MKNStudio Classes: \code{\link{MKNData}}

Other LMStudio Classes: \code{\link{LMStudio0}},
  \code{\link{MKNData}}, \code{\link{MKN}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
