% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLPStudio.R
\docType{class}
\name{NLPStudio}
\alias{NLPStudio}
\title{NLPStudio}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NLPStudio
}
\arguments{
\item{name}{A character string containing the studio name}

\item{desc}{A character string containing the studio description}

\item{current}{A logical indicating whether to set the lab current.}

\item{lab}{An object of class 'Lab'}
}
\description{
\code{NLPStudio} Class the creates and manages labs.
}
\details{
This class creates and manages data labs  A data lab is essentially a
directory in which project data reside.  Multiple data labs can be created
to house separate versions of the data for analysis. Note: This class is a
singleton pattern. An NLPStudio object called nlpStudio is instantiated at
load time.  Any subsequent initializations will return the single nlpStudio
instance.
}
\section{Fields}{

\describe{
\item{\code{nlpStudio:}}{The name of the single NLPStudio object created at load time.}
}}

\section{Methods}{

\describe{
 \item{\code{new()}}{Initializes the NLPStudio. This is a singleton class in which its only object is created when the package is loaded.}
 \item{\code{getStudio()}}{Retrieves current studio object, returns in list format and prints to console}
 \item{\code{addLab(lab, current)}}{Adds an existing lab to the NLPStudio object list of labs.  If current is set to TRUE, it is also assigned to the 'currentLab' variable.}
 \item{\code{removeLab()}}{Method which archives the lab and removes it from the global environment and from cache.}
 \item{\code{listlabs()}}{Returns a data frame of labs in the NLPStudio and prints them to console}
}
}

\section{Parameters}{

}

\section{Active Bindings}{

\describe{
 \item{\code{desc()}}{An active binding that gets and sets the NLPStudio description}
 \item{\code{current()}}{An active binding that gets and sets the current lab. With the exception of the cross-lab classes, all objects act on the current lab. }
}
}

\examples{
\dontrun{
nlpStudio$getStudio()
nlpStudio$desc <- "Homer's NLPStudio"  # Changes the description of the studio
nlpStudio$addLab("Lisa", current = FALSE)
nlpStudio$addLab("Bart", current = TRUE)
nlpStudio$removeLab("Bart") # Fails: Cannot remove a current lab
nlpStudio$currentLab <- "Lisa"
nlpStudio$removeLab("Bart") # Success!
nlpStudio$listlabs()  # Renders a list of labs to console and to data frame
}

}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
