% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLPStudio.R
\docType{class}
\name{NLPStudio}
\alias{NLPStudio}
\title{NLPStudio}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NLPStudio
}
\arguments{
\item{name}{A character string containing the studio name}

\item{desc}{A character string containing the studio description}
}
\description{
\code{NLPStudio} Class the creates and manages labs.
}
\details{
This class creates and manages data labs  A data lab is
essentially a directory in which project data reside.  Multiple data
labs can be created to house separate versions of the data
for analysis and management.

Note: This class is a singleton pattern. An NLPStudio object called
      nlpStudio is  instantiated at load time.  Any subsequent
      initializations will return the nlpStudio instance.
}
\section{Methods}{

\describe{
 \item{\code{new()}}{Initializes the NLPStudio. Note, an instance called nlpStudio is created at load time}
 \item{\code{getLab()}}{Retrieves current studio object, returns in list format and prints to console}
 \item{\code{searchLab(name)}}{Searches the studio for a named lab}
 \item{\code{archiveLab(name)}}{Archives the labs in the studio}
 \item{\code{addLab(name, desc, current)}}{Creates an lab and sets its current status to TRUE or FALSE}
 \item{\code{listlabs()}}{Returns a data frame of labs in the NLPStudio and prints them to console}
 \item{\code{removeLab(name)}}{Removes an lab (after archiving it)}
}
}

\section{Parameters}{

}

\section{Active Bindings}{

\describe{
 \item{\code{currentLab}}{Gets and sets the current lab}
}
}

\examples{
\dontrun{
nlpStudio$getStudio()
nlpStudio$archiveLab("Main")
nlpStudio$removeLab("Main")
nlpStudio$searchStudio("Main")
nlpStudio$listlabs()
nlpStudio$addLab("Alpha", "Alpha Test Lab", FALSE)
nlpStudio$currentLab <- "Alpha"
nlpStudio$listlabs()
}

}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
