% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_corpus.R
\name{parse_corpus}
\alias{parse_corpus}
\title{Fast Corpus Parsing via spaCy}
\usage{
parse_corpus(x, ncores = 1, ...)
}
\arguments{
\item{x}{A \strong{quanteda} \link[quanteda:corpus]{quanteda::corpus}.}

\item{ncores}{The number of \link[future:multisession]{future::multisession} workers to be allocated for the
parsing. Default to 1.}

\item{...}{Additional arguments passed to \code{\link[spacyr:spacy_parse]{spacyr::spacy_parse()}}.}
}
\value{
A \link[data.table:data.table]{data.table::data.table} of tokenized, parsed, and annotated tokens.
}
\description{
Parse a \link[quanteda:corpus]{quanteda::corpus} in parallel under the \strong{\link[future:future]{future::future}} paradigm .
This function is a wrapper of \code{\link[spacyr:spacy_parse]{spacyr::spacy_parse()}}. Thus, it is critical to have
a working installation of the \strong{\link[spacyr:spacyr-package]{spacyr::spacyr}} package. Please refer to the
\href{https://github.com/quanteda/spacyr?tab=readme-ov-file#installing-the-package}{installation guide}
to troubleshoot issues.
}
\details{
The workhorse of this function is \code{\link[spacyr:spacy_parse]{spacyr::spacy_parse()}} such that all the usual parameters
can be passed to \code{parse_corpus()} as well. It is critical to have a proper installation of the
\href{https://spacy.io/}{\strong{spaCy}} library and all of its components. \code{parse_corpus()} does not
initizalize any instance of spaCy so call \code{\link[spacyr:spacy_initialize]{spacyr::spacy_initialize()}} beforehand.

In particular, one can pass and use any language model as currently supported by version 3.7 via
the argument \code{model} in \code{\link[spacyr:spacy_initialize]{spacyr::spacy_initialize()}}.
By default, \code{\link[spacyr:spacy_install]{spacyr::spacy_install()}} downloads and uses the smallest English model
\code{en_core_web_sm}. It is recommended to use \code{\link[spacyr:spacy_download_langmodel]{spacyr::spacy_download_langmodel()}}
to properly download and activate the desired model.

To avoid any issue, \code{parse_corpus()} always calls \code{\link[spacyr:spacy_finalize]{spacyr::spacy_finalize()}} while closing via \code{on.exit()}.
}
\note{
Although parsing can be parallelized across multiple CPU cores, memory usage
grows quickly with both the number of cores and the size of the corpus.
On large corpora, allocating too many workers may exhaust available RAM and
significantly slow down or even terminate the process. It is recommended to
increase \code{ncores} gradually and monitor memory consumption.
}
\seealso{
\code{\link[spacyr:spacy_install]{spacyr::spacy_install()}} \code{\link[spacyr:spacy_initialize]{spacyr::spacy_initialize()}} \code{\link[spacyr:spacy_parse]{spacyr::spacy_parse()}}
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}
}
