% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{printHeading}
\alias{printHeading}
\title{printHeading}
\usage{
printHeading(text, symbol = "=", newlines = 1)
}
\description{
\code{printHeading} Prints a three line heading with text centered on 2nd line
}
\seealso{
Other Internal Functions: \code{\link{listFiles}},
  \code{\link{proper}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
