% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{proper}
\alias{proper}
\title{proper}
\usage{
proper(x)
}
\description{
\code{proper} Converts text string to proper case
}
\seealso{
Other Internal Functions: \code{\link{listFiles}},
  \code{\link{printHeading}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
