% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_corpus.R
\name{reshape_corpus}
\alias{reshape_corpus}
\title{Fast Corpus Reshape}
\usage{
reshape_corpus(x, ncores, to = "sentences", ...)
}
\arguments{
\item{x}{A \strong{quanteda} \link[quanteda:corpus]{quanteda::corpus}.}

\item{ncores}{The number of \link[future:multisession]{future::multisession} workers to be allocated for the reshaping.}

\item{to}{Character; type of reshaping. Default is \code{"sentences"}. Passed to \code{\link[quanteda:corpus_reshape]{quanteda::corpus_reshape()}}.}

\item{...}{Additional arguments passed to \code{\link[quanteda:corpus_reshape]{quanteda::corpus_reshape()}} (see 'Details').}
}
\value{
A reshaped \link[quanteda:corpus]{quanteda::corpus} object.
}
\description{
Reshape a \link[quanteda:corpus]{quanteda::corpus} in parallel under the \strong{\link[future:future]{future::future}} paradigm.
}
\details{
This function wraps \code{\link[quanteda:corpus_reshape]{quanteda::corpus_reshape()}} but executes in parallel using the
\link[future:future]{future::future} framework. The most common use is reshaping to sentences (\code{to = "sentences"}),
which enables sentence-level text analysis and alignment with document-level metadata.

Document order is guaranteed to match the input \code{corpus}. After reshaping,
documents are reordered to ensure consistency with the original corpus ordering.

Use \code{...} to further control the reshaping process. For instance, one can respahe the corpus to
paragraphs by passing \code{to = "paragraphs"}.
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}
}
