% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample.R
\docType{class}
\name{Sample}
\alias{Sample}
\title{Sample}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Sample
}
\arguments{
\item{x}{Corpus object.}

\item{n}{Numeric of length one, or numeric vector of length equal to
the number of documents  in the Corpus.}

\item{stratify}{Logical. If TRUE (default), samples will be taken from
each document in accordance with the proportions or counts indicated
in the 'n' parameter.}

\item{seed}{Numeric used to initialize a pseudorandom number generator.}
}
\description{
\code{Sample} Creates Corpus from a Corpus sample.
}
\details{
Creates a Corpus object from a sample of a Corpus object.
}
\section{Methods}{

 \itemize{
  \item{\code{new()}}{Not implemented.}
 }
}

\seealso{
Other Corpus Sample Family of Classes: \code{\link{KFold}},
  \code{\link{Sample0}}, \code{\link{Segment}}

Other CorpusStudio Family of Classes: \code{\link{CorpusStudio}},
  \code{\link{KFold}}, \code{\link{Sample0}},
  \code{\link{Segment}}, \code{\link{Split}},
  \code{\link{TokenizerNLP}}, \code{\link{TokenizerQ}},
  \code{\link{Tokenizer}}, \code{\link{Token}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
