% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segment.R
\docType{class}
\name{Segment}
\alias{Segment}
\title{Segment}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Segment
}
\arguments{
\item{x}{Corpus object.}

\item{n}{Numeric of indicating the number of units to include in each segment.}

\item{size}{Numeric indicating the size of each units to include in each segment}

\item{stratify}{Logical. If TRUE (default), Segments will be taken from
each document in accordance with the proportions or counts indicated
in the 'n' parameter.}

\item{seed}{Numeric used to initialize a pseudorandom number generator.}
}
\description{
\code{Segment} Segments a Corpus into n equal units of a designated size.
}
\section{Methods}{

 \itemize{
  \item{\code{new()}}{Not implemented.}
 }
}

\seealso{
Other Corpus Sample Family of Classes: \code{\link{Sample0}},
  \code{\link{Sample}}, \code{\link{SplitKFold}}

Other CorpusStudio Family of Classes: \code{\link{CorpusStudio}},
  \code{\link{Sample0}}, \code{\link{Sample}},
  \code{\link{SplitKFold}}, \code{\link{Split}},
  \code{\link{TokenizerNLP}}, \code{\link{TokenizerQ}},
  \code{\link{Tokenizer}}, \code{\link{Token}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
