% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_ff_industries.R
\name{set_ff_industries}
\alias{set_ff_industries}
\title{Pull Fama-French Industries and Map Them to a Corpus}
\usage{
set_ff_industries(x, ind, fill_category = FALSE, ...)
}
\arguments{
\item{x}{A \code{data.table} as extracted by a quanteda \link[quanteda:corpus]{quanteda::corpus} object.}

\item{ind}{Fama-French industry grouping (See 'Details').}

\item{fill_category}{Should unmatched industries be labelled as \code{Unclassified}? Default to \code{FALSE}.}

\item{...}{Not used at the moment.}
}
\value{
A \link[quanteda:corpus]{quanteda::corpus} with the Fama-French industries as \code{docvars}.
}
\description{
This function maps Fama-French industry classifications to a quanteda corpus based on each
document’s SIC code.
}
\details{
The Fama-French industry classifications are available in several versions, with the most
commonly used being the 12-industry (FF12), 17-industry (FF17), and 48-industry (FF48) groupings.
These are widely adopted in empirical asset pricing research due to their balance between
interpretability and granularity. The 49-industry classification (FF49), which originates
from Fama and French’s 1997 Journal of Financial Economics paper, remains a frequently used
option, especially for studies requiring finer industry resolution. An older 30-industry
scheme (FF30) also exists but is largely deprecated and not recommended for current research
applications. A 38-industry classification (FF38) has been introduced more recently and is
occasionally used by Kenneth French in the construction of anomaly portfolios and other datasets.
Each scheme groups SIC codes into economically meaningful categories with varying levels of
detail, allowing users to tailor industry aggregation to their research needs.
}
