% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLM0.R
\docType{class}
\name{SLM0}
\alias{SLM0}
\title{SLM0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SLM0
}
\description{
\code{SLM0} Base class for the statistical learning family of classes.
}
\seealso{
Other Statistical Language Model Classes: \code{\link{KN}},
  \code{\link{MKN}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
