% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitCorpus.R
\docType{class}
\name{SplitCorpus}
\alias{SplitCorpus}
\title{SplitCorpus}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SplitCorpus
}
\arguments{
\item{x}{Corpus or Token Object}

\item{corpusStudio}{CorpusStudio object}

\item{name}{Character string containing the name to assign to the cross
validation (CVSet) object.}

\item{train}{Numeric indicating the proportion of the Corpus to allocate
to the training set. Acceptable values are between 0 and 1. The total of the
values for the train, validation and test parameters must equal 1.}

\item{validation}{Numeric indicating the proportion of the Corpus to allocate
to the validation set. Acceptable values are between 0 and 1. The total of the
values for the train, validation and test parameters must equal 1.}

\item{test}{Numeric indicating the proportion of the Corpus to allocate
to the test set. Acceptable values are between 0 and 1. The total of the
values for the train, validation and test parameters must equal 1.}

\item{stratify}{Logical. If TRUE (default), splits and sampling will
be stratefied.}

\item{seed}{Numeric used to initialize a pseudorandom number generator.}
}
\value{
Token object
}
\description{
\code{SplitCorpus} Encapsulates the command to split a Corpus object.
}
\details{
Encapsulates the command to split a Corpus object into a cross-validation
object (CV), containing a training, test, and optional validation sets.
}
\seealso{
Other FileStudio Command Classes: \code{\link{BuildCorpus}},
  \code{\link{CleanCorpus}}, \code{\link{CopyFileSet}},
  \code{\link{MoveFileSet}}, \code{\link{ReadCorpus}},
  \code{\link{RepairFileSet}}, \code{\link{SampleCorpus}},
  \code{\link{Source}}, \code{\link{TokenizeCorpus}},
  \code{\link{WriteCorpus}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
