% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitKFold.R
\docType{class}
\name{SplitKFold}
\alias{SplitKFold}
\title{SplitKFold}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SplitKFold
}
\arguments{
\item{corpus}{Corpus object.}

\item{k}{Numeric indicating the number of folds.}

\item{name}{Character string indicating the name for the cross-validation set.}

\item{train}{Numeric indicating the proportion of the Corpus to allocate
to the training set. Acceptable values are between 0 and 1. The total of the
values for the train, validation and test parameters must equal 1.}

\item{validation}{Numeric indicating the proportion of the Corpus to allocate
to the validation set. Acceptable values are between 0 and 1. The total of the
values for the train, validation and test parameters must equal 1.}

\item{test}{Numeric indicating the proportion of the Corpus to allocate
to the test set. Acceptable values are between 0 and 1. The total of the
values for the train, validation and test parameters must equal 1.}

\item{stratify}{Logical. If TRUE (default), splits and sampling will
be stratefied.}

\item{seed}{Numeric used to initialize a pseudorandom number generator.}
}
\value{
CVKFold object
}
\description{
\code{SplitKFold} Splits a Corpus object into a k fold cross-validation sets.
}
\details{
Splits a Corpus object into k fold cross-validation sets, each containing
a training, test and an optional validation set.
}
\section{Methods}{

 \itemize{
  \item{\code{new()}}{Initializes an object of the SplitKFold class.}
  \item{\code{execute(x, train = 0.75, validation = 0, test = 0.25,
  stratify = TRUE, seed = NULL)}}{Executes the corpus splits.}
 }
}

\seealso{
Other Corpus Split Family of Classes: \code{\link{Split}}

Other CorpusStudio Family of Classes: \code{\link{CorpusStudio}},
  \code{\link{Split}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
