% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateManager.R
\docType{class}
\name{StateManager}
\alias{StateManager}
\title{StateManager}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
StateManager
}
\arguments{
\item{key}{Character string containing the name of the object to retrieved or stored to state}

\item{value}{The object to be stored to state}
}
\description{
\code{StateManager} Class for storing the current state of all objects in
NLPStudio.
}
\details{
NLPStudio objects may be inadvertaetly removed from the global environment.
This class allows clients to restore the objects to its current state in
the global environment. The current state of all objects is stored to
file for persistence from session to session.
}
\section{Methods}{

\describe{
 \item{\code{initialize()}}{Creates the state environment.}
 \item{\code{getState(key)}}{Retrieves object from state.}
 \item{\code{setState(key, value)}}{Save a key value pair to the state environment.}
 \item{\code{loadState()}}{Loads the state from file.}
 \item{\code{saveState()}}{Restores items in state to the global environment.}
 \item{\code{restoreState()}}{Purges the entire state.}
 \item{\code{printState}}{Lists the names of the items in the state environment.}
 \item{\code{purgeState}}{Lists the names of the items in the state environment.}
 \item{\code{printState()}}{Prints the contents of the state environment.}
 \item{\code{getInstance()}}{Returns the instance of the singleton StateManager class.}
}
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
