% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StripText.R
\docType{class}
\name{StripText}
\alias{StripText}
\title{StripText}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
StripText$new(keepChars = NULL, removeDigit = TRUE, lowerCase = TRUE)
}
\arguments{
\item{x}{Object of to be processed. This can be an object of the Corpus, Document, character classes, or a list of character vectors.}

\item{keepChars}{A character vector of symbols (i.e., punctuation) that
strip should keep. The default is to strip every symbol except apostrophes
and a double tilde "~~". The double tilde "~~" is included for a
convenient means of keeping word groups together in functions that
split text apart based on space To remove double tildes "~~" set
char.keep to NULL.}

\item{removeDigit}{Logical. If TRUE, digits are removed from the text.}

\item{removeApostrophe}{Logical. If TRUE, digits are removed from the text.}

\item{lowerCase}{Logical.  If TRUE, forces all alphabetic characters to lower case.}
}
\description{
\code{StripText} Command for the StripText class.
}
\details{
Class that encapsulates the command to execute an object of the StripText
class
}
\section{Methods}{

 \describe{
  \item{\code{new(...)}}{Instantiates the class object.}
  \item{\code{execute()}}{Executes the process on the object.}
}
}

\section{TextStudio Family of Classes Overview}{

The TextStudio family of classes include four types of classes. The manager classes which
are responsible for executing one or more core class objects through the implementation
of a command class object, supported by support class objects.
}

\section{DataStudio Design}{

This DataStudio family of classes is an implementation of the Strategy
design pattern, as described in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). The Strategy pattern defines
the family of text cleaning classes, encapsulates class and allows
text cleaning algorithms to be selected and exchagned at runtime.
}

\seealso{
Other TextStudio Classes: \code{\link{AddCommaSpaceApp}},
  \code{\link{AddCommaSpace}}, \code{\link{GSubApp}},
  \code{\link{GSub}}, \code{\link{RemoveEmailApp}},
  \code{\link{RemoveEmail}},
  \code{\link{RemoveTwitterApp}},
  \code{\link{RemoveTwitter}}, \code{\link{RemoveURLApp}},
  \code{\link{RemoveURL}},
  \code{\link{RemoveWhiteSpaceApp}},
  \code{\link{RemoveWhiteSpace}},
  \code{\link{ReplaceAbbreviationsApp}},
  \code{\link{ReplaceAbbreviations}},
  \code{\link{ReplaceBacktickApp}},
  \code{\link{ReplaceBacktick}},
  \code{\link{ReplaceContractionsApp}},
  \code{\link{ReplaceContractions}},
  \code{\link{ReplaceCurlyQuotesApp}},
  \code{\link{ReplaceCurlyQuotes}},
  \code{\link{ReplaceEmojiApp}},
  \code{\link{ReplaceEmoji}},
  \code{\link{ReplaceEmoticonApp}},
  \code{\link{ReplaceEmoticon}},
  \code{\link{ReplaceHTMLApp}}, \code{\link{ReplaceHTML}},
  \code{\link{ReplaceInternetSlangApp}},
  \code{\link{ReplaceInternetSlang}},
  \code{\link{ReplaceKernApp}}, \code{\link{ReplaceKern}},
  \code{\link{ReplaceNonAsciiApp}},
  \code{\link{ReplaceNonAscii}},
  \code{\link{ReplaceNumbersApp}},
  \code{\link{ReplaceNumbers}},
  \code{\link{ReplaceOrdinalApp}},
  \code{\link{ReplaceOrdinal}},
  \code{\link{ReplaceSymbolApp}},
  \code{\link{ReplaceSymbol}},
  \code{\link{ReplaceTokensApp}},
  \code{\link{ReplaceTokens}},
  \code{\link{ReplaceWordElongationApp}},
  \code{\link{ReplaceWordElongation}},
  \code{\link{StripTextApp}}
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
