% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermFreq.R
\docType{class}
\name{TermFreq}
\alias{TermFreq}
\title{TermFreq}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TermFreq
}
\arguments{
\item{x}{A Corpus object}

\item{type}{Character string indicating the type of term frequency matrix
to produce. Valid values are 'dfm' for the quanteda dfm object, 'dtm' for
the tm package DocumentTermMatrix, and 'tdm' for the tm package
TermDocumentMatrix.}

\item{tolower}{Logical, if TRUE, converts text to lower case prior to creating
the term frequency matrix.  The Default is TRUE.}

\item{stem}{Logical, if TRUE, conducts stemming prior to creating
the term frequency matrix.}

\item{dictionary}{For the dfm objects, this is a list of named character
vectors of words to be included in the term frequency matrix. For
the tdm and dtm objects, the dictionary is a vector of terms. The default
is NULL.}
}
\value{
A dfm, TermDocumentMatrix, or a DocumentTermMatrix as requested.
}
\description{
\code{TermFreq} Class responsible for producing term frequency matrices.

@param x The principal object
 @param key Character string or a vector of strings containing a key or
 keys for metadata key/value pair(s).
 @param value Character or numeric, or a vector thereof, containing a value
 or values for metadata key/value pair(s).
 @param format Character string used in the metadata method. Indicates
 whether to return the metadata as  list of lists, or a data.frame.This method
 returns metadata in a list if no key/value pair is provided. The list
 may contain other lists or data frames associated with the type of
 metadata. This parameter specifies whether to return the metadata
 as a list of lists, or a list of data frames. Valid values are:
 c('list', 'data.frame')
}
\details{
Class receives a request for a quanteda dfm, tm TermDocumentMatrix, or a
tm DocumenTermMatrix, then dispatches the request to the appropriate
concrete factory for production.
}
\section{Core Methods}{

 \itemize{
  \item{\code{new(x, type = c("dfm", "dtm", "tdm"), tolower = TRUE,
  stem = FALSE, dictionary = NULL)}}{Initializes an object of the TermFreq class.}
  \item{\code{execute()}}{Executes the build via the appropriate factory.}
  }
}

\seealso{
Other Term Frequency Classes: \code{\link{TermFreqDfm}},
  \code{\link{TermFreqDtm}},
  \code{\link{TermFreqFactory0}}, \code{\link{TermFreqTdm}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
