% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermFreqQ.R
\docType{class}
\name{TermFreqQ}
\alias{TermFreqQ}
\title{TermFreqQ}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TermFreqQ
}
\arguments{
\item{x}{A quanteda dfm object.}

\item{name}{Optional character string containing the name of the object.}

\item{event}{Character string indicating a class associated with an event to be posted to the object's log.}

\item{key}{Character string or character vector containing the key or
keys associated a metadata field or fields.}

\item{value}{Character string or character vector, equal in length to
the key vector, containing the metadata value or values.}
}
\value{
The get method returns a quanteda dfm object.
}
\description{
\code{TermFreqQ} Domain class containing a term frequency matrix produced by the \code{\link[NLPStudio]{TermFreqStrategyQ}} class.
}
\details{
TermFreqQ objects are wrappers for the \code{\link[quanteda]{dfm}} class.
}
\section{Core Methods}{

 \itemize{
  \item{\code{new(x)}}{Initializes an object of the TermFreqQ class.}
  \item{\code{get()}}{Returns the quanteda dfm object.}
  }
}

\section{Metadata Method}{

 \itemize{
  \item{\code{meta(key = NULL, value = NULL)}}{Provides facility for
  managing an object's metadata represented as key/value pairs in
  a list format. Metadata may be indicated via single key and
  value character strings or by pairs of character vectors. If
  no parameters are passed to the method, then it returns the
  current metadata, as well as system generated application
  and system metadata, in data frame format. If a character
  string or vector is passed via the key parameter, the value
  or values associated with the keys are returned.}
  \item{\code{log(cls = class(self)[1], event = NULL)}}{Class for posting to and
  retrieving an object's log. If the event parameter is provided,
  the event is posted to the object's log and the log is returned
  invisibly.  If the event parameter is NULL, the existing log
  is returned.}
}
}

\seealso{
Other Term Frequency Classes: \code{\link{TermFreqStrategy0}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
