% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextStudio.R
\docType{class}
\name{TextStudio}
\alias{TextStudio}
\title{TextStudio}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TextStudio
}
\description{
\code{TextStudio}  Class responsible for manipulating Corpus texts.
}
\details{
Class contains methods for normalizing and cleaning Corpus object texts.
}
\author{
John James, \email{jjames@dataScienceSalon.org}
}
\keyword{datasets}
