% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Token.R
\docType{class}
\name{Token}
\alias{Token}
\title{Token}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Token
}
\description{
\code{Token} Token class. Contains character, word, sentence, or paragraph tokens.
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
