% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tokenize}
\alias{tokenize}
\title{tokenize}
\usage{
tokenize(x, tokenType = "char")
}
\description{
\code{tokenize} Creates sentence, word and character tokens
}
\seealso{
Other Internal Functions: \code{\link{listFiles}},
  \code{\link{printHeading}}, \code{\link{proper}},
  \code{\link{slice}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
