% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize_corpus.R
\name{tokenize_corpus}
\alias{tokenize_corpus}
\title{Fast Corpus Tokenization}
\usage{
tokenize_corpus(x, ncores, ...)
}
\arguments{
\item{x}{A \strong{quanteda} \link[quanteda:corpus]{quanteda::corpus} as built by \code{define_corpus}.}

\item{ncores}{The number of \link[future:multisession]{future::multisession} workers to be allocated for the tokenization.}

\item{...}{Additional arguments passed to \link[quanteda:tokens]{quanteda::tokens}.}
}
\value{
A \link[quanteda:tokens]{quanteda::tokens} object.
}
\description{
Tokenize a \link[quanteda:corpus]{quanteda::corpus} in parallel via \strong{\link[future:future]{future::future}}.
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}
}
