% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize_corpus.R
\name{tokenize_corpus}
\alias{tokenize_corpus}
\title{Fast Corpus Tokenization}
\usage{
tokenize_corpus(x, ncores, ...)
}
\arguments{
\item{x}{A \code{\link[quanteda]{corpus}} as built by \code{create_corpus}.}

\item{ncores}{The number of \code{\link[future]{multisession}} workers to be allocated for the tokenization.}

\item{...}{Additional arguments passed to \code{\link[quanteda]{tokens}}.}
}
\value{
A \code{\link[quanteda]{tokens}} object.
}
\description{
Tokenize a \code{\link[quanteda]{corpus}} in parallel via \strong{\link{future}}.
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}
}
