% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tokenizer.R
\docType{class}
\name{Tokenizer}
\alias{Tokenizer}
\title{Tokenizer}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Tokenizer
}
\arguments{
\item{x}{Corpus object.}

\item{tokenUnit}{Character string indicating the token unit. Valid values
are 'character', 'word', 'sentence', and 'paragraph'. Abbreviations
are permitted.}
}
\value{
Tokenizer object
}
\description{
\code{Tokenizer} Creates a Tokenizer object from a Corpus.
}
\details{
Class creates a character, word, sentence or paragraph Tokenizer object
from a Corpus object.
}
\section{Methods}{

 \itemize{
  \item{\code{new()}}{Initializes an object of the Tokenizer class.}
  \item{\code{execute(x, train = 0.75, validation = 0, test = 0.25,
  stratify = TRUE, seed = NULL)}}{Executes the corpus Tokenizers.}
 }
}

\seealso{
Other CorpusStudio Family of Classes: \code{\link{CorpusStudio}},
  \code{\link{SplitKFold}}, \code{\link{Split}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
