% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tokens.R
\docType{class}
\name{Tokens}
\alias{Tokens}
\title{Tokens}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
tokenizedCorpus <- Tokens$new(name = 'Sentence Tokens', what = "sentence")
}
\arguments{
\item{x}{The Corpus object to be tokenized.}

\item{name}{Character string containing the name to assign to the Tokenized Corpus}

\item{what}{Character string indicating the level of tokenization. Valid
values are c("sentence", "word", "character").}
}
\value{
Tokens object.
}
\description{
\code{Tokens} Class representing a Tokenized Corpus.

@param x The principal object
 @param key Character string or a vector of strings containing a key or
 keys for metadata key/value pair(s).
 @param value Character or numeric, or a vector thereof, containing a value
 or values for metadata key/value pair(s).
 @param format Character string used in the metadata method. Indicates
 whether to return the metadata as  list of lists, or a data.frame.This method
 returns metadata in a list if no key/value pair is provided. The list
 may contain other lists or data frames associated with the type of
 metadata. This parameter specifies whether to return the metadata
 as a list of lists, or a list of data frames. Valid values are:
 c('list', 'data.frame')
}
\details{
Tokens objects are Corpus level objects containing collections of
Tokenized Document objects.
}
\section{Core Methods}{

 \itemize{
  \item{\code{new()}}{Initializes an object of the Tokens class.}
 }
}

\seealso{
Other Tokens Classes: \code{\link{Tokenizer}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
