% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validators.R
\name{validateClass}
\alias{validateClass}
\title{validateClass}
\usage{
validateClass(object)
}
\arguments{
\item{object}{The object to be validated}
}
\value{
a list containing two elements: a code and a message. If the
validation passed, the code will be TRUE and the msg element will be NULL.
Otherwise, the code will be FALSE and the msg element will describe
the error.
}
\description{
\code{validateClass} Validates the class of one or more parameters.
}
\details{
Validates the class of one or more objects. The object must
have a public member called params which is a list containing three
elements: (1) a vector of object names, (2) a vector containing
the objects, and (3) a list of vectors, where each vector
contains the valid classes for each of the objects
}
\seealso{
Other Validation Functions: \code{\link{validateDiscrete}},
  \code{\link{validateKeyValue}},
  \code{\link{validateLogical}},
  \code{\link{validatePath}}, \code{\link{validateRange}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\concept{Validation Functions}
