% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validators.R
\name{validateKeyValue}
\alias{validateKeyValue}
\title{validateKeyValue}
\usage{
validateKeyValue(object)
}
\arguments{
\item{object}{The object to be validated}
}
\description{
\code{validateKeyValue} Validates a parameters that contain key/value pairs
}
\details{
Function accepts an object with a public parameters member. The parameters
member has one of two formats.
}
\seealso{
Other Validation Functions: \code{\link{validateClass}},
  \code{\link{validateDiscrete}},
  \code{\link{validateLogical}},
  \code{\link{validatePath}}, \code{\link{validateRange}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\concept{Validation Functions}
