% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validators.R
\name{validatePath}
\alias{validatePath}
\title{validatePath}
\usage{
validateRange(object)
}
\arguments{
\item{object}{The object to be validated}
}
\description{
\code{validatePath} Validates the existence or non-existence of a path.
}
\details{
Class returns an error if the existence of a path is not expected.
}
\seealso{
Other Validation Functions: \code{\link{validateClass}},
  \code{\link{validateDiscrete}},
  \code{\link{validateKeyValue}},
  \code{\link{validateLogical}},
  \code{\link{validateRange}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
