% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validator.R
\docType{class}
\name{Validator}
\alias{Validator}
\title{Validator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Validator
}
\arguments{
\item{object}{Object to be validated}
}
\value{
A list containing:
\itemize{
 \item status: Logical. If validation passed, then, TRUE, otherwise FALSE
 \item msg: Character string describing the error if status is FALSE
}
}
\description{
Class accepts validation requests from client objects and dispatches a
ValidationVisitor class object to perform validation.
}
\details{
Class responsible for accepting validation requests and delegating
the request to a ValidationVisitor object.
}
\section{Validator methods}{

This section summarizes the methods in the Validator class.
\describe{
 \item{\code{validate(object)}}{Dispatches the ValidationVisitor to the client object.}
 }
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\concept{Validation Classes}
\keyword{datasets}
