% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warp_lda.R
\name{warp_lda}
\alias{warp_lda}
\alias{warpLDA}
\title{Fit a WarpLDA Class Topic Model}
\usage{
warp_lda(x, k, return_theta = TRUE, return_phi = TRUE, ...)

warpLDA(...)
}
\arguments{
\item{x}{A sparse document-term matrix of class \link[Matrix:dgCMatrix-class]{Matrix::dgCMatrix}.}

\item{k}{The number of latent topics to estimate.}

\item{return_theta}{Should the Document-Topic-Weights matrix ( \eqn{\theta} ) be returned? Default to \code{TRUE}.}

\item{return_phi}{Should the Topic-Word-Weights matrix ( \eqn{\phi} ) be returned? Default to \code{TRUE}.}

\item{...}{Additional arguments passed to \code{\link[text2vec:LatentDirichletAllocation]{LDA$new}} and \link[text2vec:LatentDirichletAllocation]{LDA$fit_transform}.}
}
\value{
A named list with the following elements:
\itemize{
\item \code{lda_object}. The trained \code{\link[text2vec:LatentDirichletAllocation]{LDA}} model object, containing methods for inference,
scoring, and topic extraction.
\item \code{theta}. A \code{data.table} of document-topic weights (\eqn{\theta}), where rows are documents
and columns are topic proportions. Included if \code{return_theta = TRUE}.
\item \code{phi}. A \code{data.table} of topic-word probabilities (\eqn{\phi}), where rows are topics and columns
are terms. Included if \code{return_phi = TRUE}.
}
}
\description{
Fit a Latent Dirichlet Allocation model using \link[text2vec:LatentDirichletAllocation]{WarpLDA} class from the package \strong{text2vec}.
Conveniently returns utility objects such as \eqn{\theta} and \eqn{\phi} matrices for sub-sequent
analysis.
}
\details{
This function wraps the \link[text2vec:LatentDirichletAllocation]{text2vec::LDA} class to fit a probabilistic topic model using the
WarpLDA algorithm—a highly efficient method for estimating Latent Dirichlet Allocation (LDA)
models on large and sparse document-term matrices. Internally, the function initializes the
LDA model via \code{LDA$new()} and fits it using \code{LDA$fit_transform()}, passing user-supplied
parameters through \code{...}.

The output includes the trained model object and, optionally, the matrices \eqn{\theta}
(document-topic proportions) and \eqn{\phi} (topic-word distributions), both in \code{data.table} format
for convenient downstream analysis. Note that \eqn{\phi} is expressed as a probability
distribution over words for each topic. That is, it already represents likelihood estimates.
}
\references{
Chen, J., Li, K., Zhu, J., & Chen, W. (2015). Warplda: a cache efficient O(1) algorithm for
Latent Dirichlet Allocation. \href{https://arxiv.org/abs/1510.08628}{\emph{arXiv preprint arXiv:1510.08628}}.

Blei, D. M., Ng, A. Y., & Jordan, M. I. (2003). \href{https://www.jmlr.org/papers/volume3/blei03a/blei03a.pdf}{Latent Dirichlet Allocation}.
\emph{Journal of Machine Learning Research}, 3(Jan), 993-1022.
}
\seealso{
\code{\link[text2vec:LatentDirichletAllocation]{LDA}}
}
